### 2023-01-14 #5081 @editorialbot
[PRE REVIEW]: BridgeStan: Efficient in-memory access to the methods of a Stan model
**Submitting author:** @roualdes (<a href="http://orcid.org/0000-0002-8757-3463">Edward Roualdes</a>)
**Repository:** https://github.com/roualdes/bridgestan
**Branch with paper.md** (empty if default branch): joss
**Version:** v.1.0.0
