### 2023-01-14 fe01d57
clear up compilation and JSON string literal wording
