### 2023-01-14 0136328
change JSON object to JSON string literal for all BridgeStan owned files
