### 2023-01-15 #5081 @bob-carpenter
Hi, @arfon.  Do you have any guidance on how long it might take?  We didn't submit to *JStatSoft* because our paper there was over 3 years in process!  Even our last *JMLR* paper took nearly 1.5 years.  There's a lot of burnout in this field among reviewers due to the huge volume of conferences.

P.S.  I'd be happy to review for JOSS and I'm guessing my co-authors would be too.  I could review submissions like ours that involve Bayesian stats and computing in multiple analysis languages and C++ (though I'm less familiar with Julia than with R and Python).
