### 2023-01-20 2679e77
Use artifact if user doesn't specify BRIDGESTAN path
