### 2023-01-20 bc6baa9
Update tests to use get_bridgestan()
