### 2023-01-20 0fa9b4e
Don't set BRIDGESTAN for Julia tests in CI
