+++ .github/workflows/main.yaml
@@ -142,7 +142,6 @@ jobs:
 
       - name: Run tests
         run: |
-          export BRIDGESTAN=$(pwd)
           julia --project=./julia -t 2 -e "using Pkg; Pkg.test()"
 
   test_R_client:
@@ -233,7 +232,6 @@ jobs:
 
       - name: Run Julia tests
         run: |
-          $env:BRIDGESTAN = $(pwd)
           julia --project=./julia -t 2 -e "using Pkg; Pkg.test()"
 
       - name: Run R tests
