+++ julia/Project.toml
@@ -4,4 +4,4 @@ authors = ["Brian Ward <bward@flatironinstitute.org>", "Bob Carpenter <bcarpente
 version = "1.0.0"
 
 [deps]
-Artifacts = "56f22d72-fd6d-98f1-02f0-08ddc0907c33"
+LazyArtifacts = "4af54fe1-eca0-43a8-85a7-787d91b784e3"
