### 2023-01-21 97310aa
Revert "Don't set BRIDGESTAN for Julia tests in CI"
This reverts commit 0fa9b4ed4527fadd9e7bd8330822eafff17b4839.
