### 2023-01-27 7266a20
Document get_bridgestan
