### 2023-01-27 d1d449b
Document and export get_bridgestan_path
