+++ julia/src/compile.jl
@@ -4,7 +4,7 @@ function get_make()
 end
 
 """
-    get_bridgestan() -> String
+    get_bridgestan_path() -> String
 
 Return the path the the BridgeStan directory.
 
@@ -12,7 +12,7 @@ If the environment variable `BRIDGESTAN` is set, this will be returned. Otherwis
 function downloads an artifact containing the BridgeStan repository and returns the path to
 the extracted directory.
 """
-function get_bridgestan()
+function get_bridgestan_path()
     path = get(ENV, "BRIDGESTAN", "")
     if path == ""
         artifact_path = artifact"bridgestan"
@@ -66,7 +66,7 @@ function compile_model(
     stanc_args::AbstractVector{String} = String[],
     make_args::AbstractVector{String} = String[],
 )
-    bridgestan = get_bridgestan()
+    bridgestan = get_bridgestan_path()
     validate_stan_dir(bridgestan)
 
     if !isfile(stan_file)
