+++ julia/src/BridgeStan.jl
@@ -20,6 +20,7 @@ export StanModel,
     log_density,
     log_density_gradient,
     log_density_hessian,
+    get_bridgestan_path,
     set_bridgestan_path!,
     compile_model
 
