+++ julia/docs/src/julia.md
@@ -77,5 +77,6 @@ param_unconstrain_json!
 ### Compilation utilities
 ```@docs
 compile_model
+get_bridgestan_path
 set_bridgestan_path!
 ```
