### 2023-01-27 e81c927
Add test for artifact being downloaded
