+++ julia/test/compile_tests.jl
@@ -36,3 +36,9 @@ end
     @test_throws ErrorException BridgeStan.set_bridgestan_path!("dummy")
     @test_throws ErrorException BridgeStan.set_bridgestan_path!(models)
 end
+
+@testset "download artifact" begin
+    withenv("BRIDGESTAN" => nothing) do
+        BridgeStan.validate_stan_dir(BridgeStan.get_bridgestan_path())
+    end
+end
