+++ .github/workflows/RegisterJuliaPackage.yaml
@@ -1,15 +1,0 @@
-name: RegisterJuliaPackage
-on:
-  workflow_dispatch:
-    inputs:
-      version:
-        description: Version to register or component to bump
-        required: true
-jobs:
-  register:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: julia-actions/RegisterAction@latest
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-          subdir: julia
