+++ docs/languages/julia.md
@@ -22,8 +22,35 @@
 ## Installation
 
 
-This assumes you have followed the [Getting Started guide](../getting-started.rst) to install BridgeStan's pre-requisites and downloaded a copy of the BridgeStan source code.
-Note that the Julia interface has the ability to install the BridgeStan source if the ``BRIDGESTAN`` environment variable is not set.
+<a id='From-JuliaRegistries'></a>
+
+<a id='From-JuliaRegistries-1'></a>
+
+### From JuliaRegistries
+
+
+BridgeStan is registered on JuliaRegistries each release.
+
+
+```julia
+] add BridgeStan
+```
+
+
+The first time you need it, the BridgeStan source code will be downloaded as an [Artifact](https://pkgdocs.julialang.org/v1/artifacts/). If you prefer to use a source distribution of BridgeStan, consult the following section.
+
+
+Note that the system pre-requisites from the [Getting Started Guide](../getting-started.rst) are still required and will not be automatically installed by this method.
+
+
+<a id='From-Source'></a>
+
+<a id='From-Source-1'></a>
+
+### From Source
+
+
+This section assumes you have followed the [Getting Started guide](../getting-started.rst) to install BridgeStan's pre-requisites and downloaded a copy of the BridgeStan source code.
 
 
 To install the Julia interface, you can either install it directly from Github by running the following inside a Julia REPL
@@ -94,18 +121,22 @@ StanModel(lib, datafile="", seed=204, chain_id=0)
 
 A StanModel instance encapsulates a Stan model instantiated with data.
 
-The constructor a Stan model from the supplied library file path and data. Data should either be a string containing a JSON object or a path to a data file ending in `.json`. If seed or chain_id are supplied, these are used to initialize the RNG used by the model.
+Construct a Stan model from the supplied library file path and data. Data should either be a string containing a JSON string literal or a path to a data file ending in `.json`. If seed or chain_id are supplied, these are used to initialize the RNG used by the model.
 
 ```
 StanModel(;stan_file, data="", seed=204, chain_id=0)
 ```
 
-Construct a StanModel instance from a `.stan` file, compiling if necessary.
+Construct a `StanModel` instance from a `.stan` file, compiling if necessary.
 
-This is equivalent to calling `compile_model` and then the original constructor of StanModel.
+```
+StanModel(;stan_file, stanc_args=[], make_args=[], data="", seed=204, chain_id=0)
+```
+
+Construct a `StanModel` instance from a `.stan` file.  Compilation occurs if no shared object file exists for the supplied Stan file or if a shared object file exists and the Stan file has changed since last compilation.  This is equivalent to calling `compile_model` and then the original constructor of `StanModel`.
 
 
-<a target='_blank' href='https://github.com/roualdes/bridgestan/blob/main/julia/src/model.jl#L4-L18' class='documenter-source'>source</a><br>
+<a target='_blank' href='https://github.com/roualdes/bridgestan/blob/main/julia/src/model.jl#L13-L33' class='documenter-source'>source</a><br>
 
 <a id='BridgeStan.log_density' href='#BridgeStan.log_density'>#</a>
 **`BridgeStan.log_density`** &mdash; *Function*.
@@ -121,7 +152,7 @@ Return the log density of the specified unconstrained parameters.
[…]
