+++ python/setup.cfg
@@ -1,5 +1,10 @@
 [metadata]
 name = bridgestan
+author = Brian Ward, Edward Roualdes, Bob Carpenter
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = BSD 3-Clause License
+url=https://github.com/roualdes/bridgestan
 version = 1.0.0
 
 [options]
