+++ julia/Project.toml
@@ -5,4 +5,6 @@ version = "1.0.0"
 
 [deps]
 LazyArtifacts = "4af54fe1-eca0-43a8-85a7-787d91b784e3"
+
+[compat]
 julia = "1.8"
