### 2023-01-27 4fa2b79
Download bridgestan repo as artifact for BridgeStan.jl (#70)
* Remove Manifest.toml

* Add bridgestan artifact

* Use artifact if user doesn't specify BRIDGESTAN path

* Update tests to use get_bridgestan()

* Don't set BRIDGESTAN for Julia tests in CI

* Create RegisterJuliaPackage.yaml

* Download BridgeStan lazily

* Revert "Don't set BRIDGESTAN for Julia tests in CI"

This reverts commit 0fa9b4ed4527fadd9e7bd8330822eafff17b4839.

* Document get_bridgestan

* Add note that BridgeStan can be automatically installed.

* Document and export get_bridgestan_path

* Add test for artifact being do
[…]
