### 2023-01-27 89ed5dd
Add test for `@const`
