### 2023-01-30 4dad4cf
Add compatibility with Julia LTS (#72)
* Remove and ignore all manifests

* Add utility macro to optionally declare a field as const

* Add Julia LTS to CI matrix

* Specify lower Julia compat

* Revert "Remove and ignore all manifests"

This reverts commit 9ea41c3d2ecce85e45cbf5f9e6f8a3f6180f6b58.

* Add test for `@const`
