+++ julia/Project.toml
@@ -7,4 +7,4 @@ version = "1.0.0"
 LazyArtifacts = "4af54fe1-eca0-43a8-85a7-787d91b784e3"
 
 [compat]
-julia = "1.8"
+julia = "1.6"
