### 2023-02-01 c09d2cf
Add unified release action (#73)
* Add release action

* Add subdir to JuliaRegistrator

* Mark new releases as drafts by default

Julia tests may fail initially since link won't exist
