+++ julia/test/runtests.jl
@@ -1,2 +1,3 @@
 include("compile_tests.jl")
 include("model_tests.jl")
+include("util_tests.jl")
