+++ src/version.hpp
@@ -13,7 +13,7 @@
 
 #define BRIDGESTAN_MAJOR 1
 #define BRIDGESTAN_MINOR 0
-#define BRIDGESTAN_PATCH 0
+#define BRIDGESTAN_PATCH 1
 
 namespace bridgestan {
 
