+++ julia/Artifacts.toml
@@ -1,7 +1,7 @@
 [bridgestan]
-git-tree-sha1 = "730b3ee2b8c9fd643dcb2bee6c0ea5995db142d7"
+git-tree-sha1 = "3c724e7bfa60bdeb77f132469e27de72a5c98e2f"
 lazy = true
 
-    [[bridgestan.download]]
-    sha256 = "b66a01322b650022bbc7944fa5229cbc90bab63fab28b063c0eae714cd927872"
-    url = "https://github.com/roualdes/bridgestan/releases/download/v1.0.0/bridgestan-1.0.0.tar.gz"
+    [bridgestan.download]
+    sha256 = "6cdb8f330c9814f6c451e3a66db73ea5284b69833d9039e906c580d29085bc65"
+    url = "https://github.com/roualdes/bridgestan/releases/download/v1.0.1/bridgestan-1.0.1.tar.gz"
