### 2023-02-01 7399404
Add LICENSE to Julia folder
Redundant, but required for General registry
