### 2023-02-01 a1c3acd
Bump r-lib/actions from 2.3.1 to 2.4.0
Bumps [r-lib/actions](https://github.com/r-lib/actions) from 2.3.1 to 2.4.0.
- [Release notes](https://github.com/r-lib/actions/releases)
- [Commits](https://github.com/r-lib/actions/compare/v2.3.1...v2.4.0)

---
updated-dependencies:
- dependency-name: r-lib/actions
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
