+++ julia/README.md
@@ -5,16 +5,13 @@
 
 ## Installation
 
-**From Github**:
+**From the General Registry**:
 ```julia
-] add https://github.com/roualdes/bridgestan.git:julia
-```
-Alternatively,
-```julia
-> using Pkg; Pkg.add(url="https://github.com/roualdes/bridgestan.git", subdir="julia")
+] add BridgeStan
 ```
 
 **From the downloaded repository**:
+Assuming you have already cloned https://github.com/roualdes/bridgestan
 ```julia
 # from folder julia/
 ] dev .
