### 2023-02-07 0a946ab
Do not create ChainableStacks except when threading
