### 2023-02-14 2733ca6
Add a few more tests to the R interface
