### 2023-02-14 bf37c6f
Add a few more tests to the R interface (#76)
