+++ R/tests/testthat/test-bridgestan.R
@@ -1,7 +1,15 @@
 base = "../../.."
 
+load_model <- function(name) {
+    model <- StanModel$new(file.path(base,paste0("/test_models/", name, "/",name,"_model.so")), file.path(base, paste0("/test_models/", name, "/",name,".data.json")), 1234, 0)
+    return(model)
+}
 
-simple <- StanModel$new(file.path(base, "/test_models/simple/simple_model.so"), file.path(base, "/test_models/simple/simple.data.json"), 1234, 0)
+test_that("missing data throws error", {
+    expect_error(StanModel$new(file.path(base,paste0("/test_models/simple/simple_model.so")), "", 1234, 0))
+})
+
+simple <- load_model("simple")
 test_that("simple_model name is correct", {
     expect_identical(simple$name(), "simple_model")
 })
@@ -30,7 +38,7 @@ test_that("simple_model Hessian is -I",{
 })
 
 
-bernoulli <- StanModel$new(file.path(base, "/test_models/bernoulli/bernoulli_model.so"), file.path(base, "/test_models/bernoulli/bernoulli.data.json"), 1234, 0)
+bernoulli <- load_model("bernoulli")
 
 test_that("loading another library didn't break prior ones", {
     if (.Platform$OS.type == "windows"){
@@ -56,4 +64,39 @@ test_that("bernoulli unconstrain works", {
     q <- log(x / (1 - x))
     expect_equal(q, bernoulli$param_unconstrain(x))
     expect_equal(q, bernoulli$param_unconstrain_json(paste("{\"theta\":", as.character(x), "}")))
+})
+
+
+fr_gaussian <- load_model("fr_gaussian")
+
+cov_constrain <- function(v,D){
+    L <- matrix(c(0), D, D)
+    L[upper.tri(L, diag=TRUE)] <- v
+    diag(L) <- exp(diag(L))
+    return(t(L) %*% L)
+}
+
+test_that("param_constrain works for a nontrivial case", {
+    D <- 4
+    unc_size <- 10
+
+    a <- rnorm(unc_size)
+    B_expected <- cov_constrain(a, D)
+
+    b <- fr_gaussian$param_constrain(a)
+    B <- array(b, dim=c(D,D))
+
+    expect_equal(B, B_expected)
+})
+
+test_that("param_unconstrain works for a nontrivial case", {
+    D <- 4
+    unc_size <- 10
+
+    a <- rnorm(unc_size)
+    B <- cov_constrain(a, D)
+    
+    c <- fr_gaussian$param_unconstrain(B)
+
+    expect_equal(c, a)
 })
