### 2023-02-23 f2e1bd9
example.R: add missing instructions
