### 2023-02-23 5af8519
Re-build R documentation files
