### 2023-02-23 91f0109
bridgestan.R: fix a few doc typos and use consistent punctuation (#77)
* bridgestan.R: fix a few doc typos and use consistent punctuation

* Re-build R documentation files

---------

Co-authored-by: Brian Ward <bward@flatironinstitute.org>
