### 2023-02-23 3981b0a
Change color of code in docs
