+++ docs/conf.py
@@ -37,6 +37,7 @@ html_theme = "pydata_sphinx_theme"
 html_static_path = ["_static"]
 html_css_files = [
     "css/Documenter.css",
+    "css/custom.css",
 ]
 
 html_show_sphinx = False
