### 2023-02-28 #5081 @Nikoleta-v3
:wave: @salleuska @andrewjradcliffe @adity-om @dhvalden would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "BridgeStan: Efficient in-memory access to the methods of a Stan model". You can find more information at the top of this Github issue 🆙
