### 2023-03-08 #5081 @salleuska
> 👋 @salleuska @andrewjradcliffe @adity-om @dhvalden would any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The submission I'd like you to review is titled: "BridgeStan: Efficient in-memory access to the methods of a Stan model". You can find more information at the top of this Github issue 🆙
>
> Please let me know if you're available 😄 Thank you!

Hi @Nikoleta-v3, happy to review the submission. Apologies for the delayed reply, I am not sure why I am not receiving emails anymore for github mentions!
