### 2023-03-08 3cf651a
Make version of package part of Python source
