+++ python/setup.cfg
@@ -5,7 +5,6 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 license = BSD 3-Clause License
 url=https://github.com/roualdes/bridgestan
-version = 1.0.1
 
 [options]
 python_requires = >=3.9
