+++ python/test/test_download.py
@@ -1,0 +1,9 @@
+import os
+
+import bridgestan as bs
+from unittest import mock
+
+
+@mock.patch.dict(os.environ, {"BRIDGESTAN": ""})
+def test_download_bridgestan():
+    bs.compile.verify_bridgestan_path(bs.compile.get_bridgestan_path())
