### 2023-03-08 0536fcf
Fix version path in setup script
