### 2023-03-09 #5081 @Nikoleta-v3
Thank you @salleuska! 🙏🏻

> I am not sure why I am not receiving emails anymore for github mentions!

🤔 not sure if that's a general issue. I personally have turned off emails from GitHub and I just check the notifications everyday 😄
