### 2023-03-10 15ddb27
Make sure -fPIC is set globally so Sundials can link
