### 2023-03-10 2e90397
Make sure -fPIC is set globally so Sundials can link (#81)
* Make sure -fPIC is set globally so Sundials can link

* Also update c-example/makefile
