### 2023-03-10 eac2788
Update list of test models in Makefile
