### 2023-03-10 3a8292c
Update list of test models in Makefile (#82)
* Update list of test models in Makefile

* Update GHA cache version
