### 2023-03-16 144356b
Python: Download bridgestan source if needed (#80)
* Make version of package part of Python source

* Download to ~/.bridgestan if needed

* Update documentation

* Fix version path in setup script

* Fix test
