+++ python/bridgestan/__init__.py
@@ -1,4 +1,5 @@
 from .compile import compile_model, set_bridgestan_path
 from .model import StanModel
+from .__version import __version__
 
 __all__ = ["StanModel", "set_bridgestan_path", "compile_model"]
