+++ CITATION.cff
@@ -3,6 +3,7 @@ message: "If you use this software, please cite it as below."
 authors:
 - family-names: "Roualdes"
   given-names: "Edward"
+  orcid: "https://orcid.org/0000-0002-8757-3463"
 - family-names: "Ward"
   given-names: "Brian"
 - family-names: "Axen"
