+++ CITATION.cff
@@ -6,6 +6,7 @@ authors:
   orcid: "https://orcid.org/0000-0002-8757-3463"
 - family-names: "Ward"
   given-names: "Brian"
+  orcid: "https://orcid.org/0000-0002-9841-3342"  
 - family-names: "Axen"
   given-names: "Seth"
 - family-names: "Carpenter"
