### 2023-03-22 ed85dd9
Add CITATION file (#86)
* adding CITATION file

* add Edward ORCID

* Add B.Ward ORCID

* Add Zenodo DOI

---------

Co-authored-by: Edward A. Roualdes <eroualdes@csuchico.edu>
Co-authored-by: Brian Ward <bward@flatironinstitute.org>
