+++ CITATION.cff
@@ -1,0 +1,19 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: "Roualdes"
+  given-names: "Edward"
+  orcid: "https://orcid.org/0000-0002-8757-3463"
+- family-names: "Ward"
+  given-names: "Brian"
+  orcid: "https://orcid.org/0000-0002-9841-3342"
+- family-names: "Axen"
+  given-names: "Seth"
+- family-names: "Carpenter"
+  given-names: "Bob"
+  orcid: "https://orcid.org/0000-0002-2433-9688"
+title: "BridgeStan: Efficient in-memory access to Stan programs through Python, Julia, and R"
+version: 1.0.2
+doi: 10.5281/zenodo.7760173
+date-released: 2023-03-22
+url: "https://github.com/roualdes/bridgestan"
