+++ rust/src/binding.rs
@@ -1,0 +1,239 @@
+/* automatically generated by rust-bindgen 0.64.0 */
+
+#[repr(C)]
+#[derive(Debug, Copy, Clone)]
+pub struct bs_model_rng {
+    _unused: [u8; 0],
+}
+pub type bool_ = ::std::os::raw::c_int;
+extern crate libloading;
+pub struct Bridgestan {
+    __library: ::libloading::Library,
+    pub bs_construct: unsafe extern "C" fn(
+        data_file: *mut ::std::os::raw::c_char,
+        seed: ::std::os::raw::c_uint,
+        chain_id: ::std::os::raw::c_uint,
+    ) -> *mut bs_model_rng,
+    pub bs_destruct: unsafe extern "C" fn(mr: *mut bs_model_rng) -> ::std::os::raw::c_int,
+    pub bs_name: unsafe extern "C" fn(mr: *mut bs_model_rng) -> *const ::std::os::raw::c_char,
+    pub bs_model_info: unsafe extern "C" fn(mr: *mut bs_model_rng) -> *const ::std::os::raw::c_char,
+    pub bs_param_names: unsafe extern "C" fn(
+        mr: *mut bs_model_rng,
+        include_tp: bool_,
+        include_gq: bool_,
+    ) -> *const ::std::os::raw::c_char,
+    pub bs_param_unc_names:
+        unsafe extern "C" fn(mr: *mut bs_model_rng) -> *const ::std::os::raw::c_char,
+    pub bs_param_num: unsafe extern "C" fn(
+        mr: *mut bs_model_rng,
+        include_tp: bool_,
+        include_gq: bool_,
+    ) -> ::std::os::raw::c_int,
+    pub bs_param_unc_num: unsafe extern "C" fn(mr: *mut bs_model_rng) -> ::std::os::raw::c_int,
+    pub bs_param_constrain: unsafe extern "C" fn(
+        mr: *mut bs_model_rng,
+        include_tp: bool_,
+        include_gq: bool_,
+        theta_unc: *const f64,
+        theta: *mut f64,
+    ) -> ::std::os::raw::c_int,
+    pub bs_param_unconstrain: unsafe extern "C" fn(
+        mr: *mut bs_model_rng,
+        theta: *const f64,
+        theta_unc: *mut f64,
+    ) -> ::std::os::raw::c_int,
+    pub bs_param_unconstrain_json: unsafe extern "C" fn(
+        mr: *mut bs_model_rng,
+        json: *const ::std::os::raw::c_char,
+        theta_unc: *mut f64,
+    ) -> ::std::os::raw::c_int,
+    pub bs_log_density: unsafe extern "C" fn(
+        mr: *mut bs_model_rng,
+        propto: bool_,
+        jacobian: bool_,
+        theta: *const f64,
+        lp: *mut f64,
+    ) -> ::std::os::raw::c_int,
+    pub bs_log_density_gradient: unsafe extern "C" fn(
+        mr: *mut bs_model_rng,
+        propto: bool_,
+        jacobian: bool_,
+        theta: *const f64,
+        val: *mut f64,
+        grad: *mut f64,
+    ) -> ::std::os::raw::c_int,
+    pub bs_log_density_hessian: unsafe extern "C" fn(
+        mr: *mut bs_model_rng,
+        propto: bool_,
+        jacobian: bool_,
+        theta: *const f64,
+        val: *mut f64,
+        grad: *mut f64,
+        hessian: *mut f64,
+    ) -> ::std::os::raw::c_int,
+}
+impl Bridgestan {
+    pub unsafe fn new<P>(path: P) -> Result<Self, ::libloading::Error>
+    where
+        P: AsRef<::std::ffi::OsStr>,
+    {
[…]
