+++ rust/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "bridgestan"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+libloading = "0.7.4"
+thiserror = "1.0.40"
