+++ rust/Makefile
@@ -1,0 +1,2 @@
+refresh_bindings:
+	bindgen ../src/bridgestan.h --opaque-type bs_model_rng --dynamic-loading Bridgestan --dynamic-link-require-all > src/binding.rs
