### 2023-03-24 4afd96f
Fix: Const correctness in C api
