+++ rust-example/Cargo.toml
@@ -4,6 +4,7 @@ version = "1.0.0"
 edition = "2021"
 
 [dependencies]
+libloading = "0.7.4"
 nuts-rs = { version = "0.4.1", optional = true }
 thiserror = "1.0.38"
 
