### 2023-03-24 f07232a
Fix: Const correctness in C api (#89)
