+++ rust-example/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bridgestan"
-version = "1.0.0"
+version = "1.0.2"
 edition = "2021"
 
 [dependencies]
