### 2023-03-24 e820f38
Start sketch of error messaging
