+++ src/bridgestan.h
@@ -19,6 +19,8 @@ typedef int bool;
  * @param[in] seed seed for PRNG
  * @param[in] chain_id identifier for concurrent sequence of PRNG
  * draws
+ * @param[out] error_msg a pointer to a string that will be allocated if there
+ * is an error. This must later be freed by calling `bs_free_error_msg`.
  * @return pointer to constructed model or `nullptr` if construction
  * fails
  */
@@ -30,10 +32,12 @@ bs_model_rng* bs_construct(const char* data_file, unsigned int seed,
  * exception while freeing memory.
  *
  * @param[in] mr pointer to model and RNG structure
+ * @param[out] error_msg a pointer to a string that will be allocated if there
+ * is an error. This must later be freed by calling `bs_free_error_msg`.
  * @return 0 for success and -1 if there is an exception freeing one
  * of the model components.
  */
-int bs_destruct(bs_model_rng* mr);
+int bs_destruct(bs_model_rng* mr, char** error_msg);
 
 /**
  * Free the error messages created by other methods.
@@ -142,11 +146,14 @@ int bs_param_unc_num(const bs_model_rng* mr);
  * @param[in] include_gq `true` to include generated quantities
  * @param[in] theta_unc sequence of unconstrained parameters
  * @param[out] theta sequence of constrained parameters
+ * @param[out] error_msg a pointer to a string that will be allocated if there
+ * is an error. This must later be freed by calling `bs_free_error_msg`.
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
 int bs_param_constrain(bs_model_rng* mr, bool include_tp, bool include_gq,
-                       const double* theta_unc, double* theta);
+                       const double* theta_unc, double* theta,
+                       char** error_msg);
 
 /**
  * Set the sequence of unconstrained parameters based on the
@@ -158,11 +165,13 @@ int bs_param_constrain(bs_model_rng* mr, bool include_tp, bool include_gq,
  * @param[in] mr pointer to model and RNG structure
  * @param[in] theta sequence of constrained parameters
  * @param[out] theta_unc sequence of unconstrained parameters
+ * @param[out] error_msg a pointer to a string that will be allocated if there
+ * is an error. This must later be freed by calling `bs_free_error_msg`.
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
 int bs_param_unconstrain(const bs_model_rng* mr, const double* theta,
-                         double* theta_unc);
+                         double* theta_unc, char** error_msg);
 
 /**
  * Set the sequence of unconstrained parameters based on the JSON
@@ -175,11 +184,13 @@ int bs_param_unconstrain(const bs_model_rng* mr, const double* theta,
  * @param[in] mr pointer to model and RNG structure
  * @param[in] json JSON-encoded constrained parameters
  * @param[out] theta_unc sequence of unconstrained parameters
+ * @param[out] error_msg a pointer to a string that will be allocated if there
+ * is an error. This must later be freed by calling `bs_free_error_msg`.
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
 int bs_param_unconstrain_json(const bs_model_rng* mr, const char* json,
-                              double* theta_unc);
+                              double* theta_unc, char** error_msg);
 
 /**
  * Set the log density of the specified parameters, dropping
@@ -193,11 +204,13 @@ int bs_param_unconstrain_json(const bs_model_rng* mr, const char* json,
  * @param[in] jacobian `true` to include change-of-variables terms
  * @param[in] theta_unc unconstrained parameters
  * @param[out] lp log density to be set
+ * @param[out] error_msg a pointer to a string that will be allocated if there
[…]
