+++ python/bridgestan/model.py
@@ -122,11 +122,17 @@ class StanModel:
             ctypes.c_int,
             double_array,
             double_array,
+            star_star_char,
         ]
 
         self._param_unconstrain = self.stanlib.bs_param_unconstrain
         self._param_unconstrain.restype = ctypes.c_int
-        self._param_unconstrain.argtypes = [ctypes.c_void_p, double_array, double_array]
+        self._param_unconstrain.argtypes = [
+            ctypes.c_void_p,
+            double_array,
+            double_array,
+            star_star_char,
+        ]
 
         self._param_unconstrain_json = self.stanlib.bs_param_unconstrain_json
         self._param_unconstrain_json.restype = ctypes.c_int
@@ -134,6 +140,7 @@ class StanModel:
             ctypes.c_void_p,
             ctypes.c_char_p,
             double_array,
+            star_star_char,
         ]
 
         self._log_density = self.stanlib.bs_log_density
@@ -144,6 +151,7 @@ class StanModel:
             ctypes.c_int,
             double_array,
             ctypes.POINTER(ctypes.c_double),
+            star_star_char,
         ]
 
         self._log_density_gradient = self.stanlib.bs_log_density_gradient
@@ -155,6 +163,7 @@ class StanModel:
             double_array,
             ctypes.POINTER(ctypes.c_double),
             double_array,
+            star_star_char,
         ]
 
         self._log_density_hessian = self.stanlib.bs_log_density_hessian
@@ -167,11 +176,12 @@ class StanModel:
             ctypes.POINTER(ctypes.c_double),
             double_array,
             double_array,
+            star_star_char,
         ]
 
         self._destruct = self.stanlib.bs_destruct
         self._destruct.restype = ctypes.c_int
-        self._destruct.argtypes = [ctypes.c_void_p]
+        self._destruct.argtypes = [ctypes.c_void_p, star_star_char]
 
     @classmethod
     def from_stan_file(
@@ -214,7 +224,7 @@ class StanModel:
         Destroy the Stan model and free memory.
         """
         if hasattr(self, "model_rng") and hasattr(self, "_destruct"):
-            self._destruct(self.model_rng)
+            self._destruct(self.model_rng, None)
 
     def __repr__(self) -> str:
         data = f"{self.data_path!r}, " if self.data_path else ""
@@ -324,13 +334,12 @@ class StanModel:
             raise ValueError(
                 "Error: out must be same size as number of constrained parameters"
             )
+        err = ctypes.pointer(ctypes.c_char_p())
         rc = self._param_constrain(
-            self.model_rng, int(include_tp), int(include_gq), theta_unc, out
+            self.model_rng, int(include_tp), int(include_gq), theta_unc, out, err
         )
         if rc:
-            raise RuntimeError(
-                "param_constrain failed on C++ side; see stderr for messages"
-            )
+            raise self._handle_error(err.contents, "param_constrain")
         return out
 
     def param_unconstrain(
@@ -357,11 +366,10 @@ class StanModel:
             raise ValueError(
                 f"out size = {out.size} != unconstrained params size = {dims}"
             )
-        rc = self._param_unconstrain(self.model_rng, theta, out)
+        err = ctypes.pointer(ctypes.c_char_p())
+        rc = self._param_unconstrain(self.model_rng, theta, out, err)
         if rc:
-            raise RuntimeError(
-                "param_unconstrain failed on C++ side; see stderr for messages"
-            )
+            raise self._handle_error(err.contents, "param_unconstrain")
         return out
 
     def param_unconstrain_json(
@@ -390,11 +398,10 @@ class StanModel:
                 f"out size = {out.size} != unconstrained params size = {dims}"
             )
         chars = theta_json.encode("UTF-8")
[…]
