+++ c-example/example.c
@@ -8,11 +8,21 @@ int main(int argc, char** argv) {
   } else {
     data = "";
   }
-  bs_model_rng* model = bs_construct(data, 123, 0);
+
+  // this could potentially error, and we may get information back about why.
+  char* err;
+  bs_model_rng* model = bs_construct(data, 123, 0, &err);
   if (!model) {
+    if (err) {
+      printf("Error: %s", err);
+      bs_free_error_msg(err);
+    }
     return 1;
   }
+
   printf("This model's name is %s.\n", bs_name(model));
   printf("It has %d parameters.\n", bs_param_num(model, 0, 0));
-  return bs_destruct(model);
+
+  // we don't care about the error message here, so we pass nullptr
+  return bs_destruct(model, 0);
 }

[…]
