### 2023-03-27 fbba400
Error handling in R
