+++ src/bridgestanR.h
@@ -11,9 +11,16 @@ typedef int bool;
 
 // Shim to convert to R interface requirement of void with pointer args
 // All calls directly delegated to versions without _R suffix
-void bs_construct_R(char** data, int* rng, int* chain, bs_model_rng** ptr_out);
+void bs_construct_R(char** data, int* rng, int* chain, bs_model_rng** ptr_out,
+                    char** err_msg, void** err_ptr);
 
 void bs_destruct_R(bs_model_rng** model, int* return_code);
+
+/**
+ * Free error message allocated in C++. Because R performs copies
+ * at the boundary on char**s, this uses void** pointing to the same memory
+ */
+void bs_free_error_msg_R(void** err_msg);
 
 void bs_name_R(bs_model_rng** model, char const** name_out);
 
@@ -31,24 +38,30 @@ void bs_param_unc_num_R(bs_model_rng** model, int* num_out);
 
 void bs_param_constrain_R(bs_model_rng** model, int* include_tp,
                           int* include_gq, const double* theta_unc,
-                          double* theta, int* return_code);
+                          double* theta, int* return_code, char** err_msg,
+                          void** err_ptr);
 
 void bs_param_unconstrain_R(bs_model_rng** model, const double* theta,
-                            double* theta_unc, int* return_code);
+                            double* theta_unc, int* return_code, char** err_msg,
+                            void** err_ptr);
 
 void bs_param_unconstrain_json_R(bs_model_rng** model, char const** json,
-                                 double* theta_unc, int* return_code);
+                                 double* theta_unc, int* return_code,
+                                 char** err_msg, void** err_ptr);
 
 void bs_log_density_R(bs_model_rng** model, int* propto, int* jacobian,
-                      const double* theta, double* val, int* return_code);
+                      const double* theta_unc, double* val, int* return_code,
+                      char** err_msg, void** err_ptr);
 
 void bs_log_density_gradient_R(bs_model_rng** model, int* propto, int* jacobian,
-                               const double* theta, double* val, double* grad,
-                               int* return_code);
+                               const double* theta_unc, double* val,
+                               double* grad, int* return_code, char** err_msg,
+                               void** err_ptr);
 
 void bs_log_density_hessian_R(bs_model_rng** model, int* propto, int* jacobian,
-                              const double* theta, double* val, double* grad,
-                              double* hess, int* return_code);
+                              const double* theta_unc, double* val,
+                              double* grad, double* hess, int* return_code,
+                              char** err_msg, void** err_ptr);
 
 #ifdef __cplusplus
 }
