### 2023-03-27 7825e53
Test that error messages are recieved by interfaces
