### 2023-03-27 370c47a
Make test Julia 1.6 safe
