### 2023-03-27 2a332c7
Start C++ changes for moving RNGs outside model object
