+++ src/model_rng.hpp
@@ -5,6 +5,7 @@
 #include <boost/random/additive_combine.hpp>
 #include <string>
 #include <vector>
+#include <random>
 
 /**
  * This structure holds a pointer to a model, holds a pseudorandom
@@ -12,7 +13,7 @@
  * CSV format.  Instances can be constructed with the C function
  * `construct()` and destroyed with the C function `destruct()`.
  */
-class bs_model_rng {
+class bs_model {
  public:
   /**
    * Construct a model and random number generator with cached
@@ -23,12 +24,12 @@ class bs_model_rng {
    * @param[in] chain_id number of gaps to skip in the pseudorandom
    * number generator for concurrent computations
    */
-  bs_model_rng(const char* data_file, unsigned int seed, unsigned int chain_id);
+  bs_model(const char* data_file, unsigned int seed);
 
   /**
    * Destroy this object and free all of the memory allocated for it.
    */
-  ~bs_model_rng() noexcept;
+  ~bs_model() noexcept;
 
   /**
    * Return the name of the model.  This class manages the memory,
@@ -114,7 +115,8 @@ class bs_model_rng {
    * @param[in,out] theta constrained parameters generated
    */
   void param_constrain(bool include_tp, bool include_gq,
-                       const double* theta_unc, double* theta);
+                       const double* theta_unc, double* theta,
+                       boost::ecuyer1988& rng);
 
   /**
    * Calculate the log density for the specified unconstrain
@@ -181,9 +183,6 @@ class bs_model_rng {
   /** Stan model */
   stan::model::model_base* model_;
 
-  /** pseudorandom number generator */
-  boost::ecuyer1988 rng_;
-
   /** name of the Stan model */
   char* name_ = nullptr;
 
@@ -222,6 +221,13 @@ class bs_model_rng {
 
   /** number of unconstrained parameters */
   int param_unc_num_ = -1;
+};
+
+class bs_rng {
+ public:
+  bs_rng(unsigned int seed) : rng_(seed) { rng_.discard(1); }
+
+  boost::ecuyer1988 rng_;
 };
 
 #endif
