### 2023-03-28 9552408
Consistent naming of helper function
