+++ R/man/handle_error.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/bridgestan.R
+\name{handle_error}
+\alias{handle_error}
+\title{Get and free the error message stored at the C++ pointer}
+\usage{
+handle_error(function_name, err_msg, err_ptr, lib_name)
+}
+\description{
+Get and free the error message stored at the C++ pointer
+}
+\keyword{internal}
