+++ julia/test/runtests.jl
@@ -1,3 +1,3 @@
-# include("compile_tests.jl")
+include("compile_tests.jl")
 include("model_tests.jl")
-# include("util_tests.jl")
+include("util_tests.jl")
