### 2023-03-28 88e7ca1
Send Stan outputs to stdout rather than stderr (#94)
