+++ src/bridgestan.cpp
@@ -2,12 +2,13 @@
 #include "model_rng.cpp"
 #include "bridgestanR.cpp"
 #include "version.hpp"
+#include "callback_stream.hpp"
+#include <sstream>
 
 int bs_major_version = BRIDGESTAN_MAJOR;
 int bs_minor_version = BRIDGESTAN_MINOR;
 int bs_patch_version = BRIDGESTAN_PATCH;
 
-#include <sstream>
 
 bs_model* bs_model_construct(const char* data, unsigned int seed,
                              char** error_msg) {
@@ -227,3 +228,16 @@ bs_rng* bs_rng_construct(unsigned int seed, char** error_msg) {
 }
 
 void bs_rng_destruct(bs_rng* rng) { delete (rng); }
+
+
+int bs_set_print_callback(STREAM_CALLBACK callback) {
+  try {
+    // TODO(bmw): this can't possibly be right
+    outstream = new std::ostream(new callback_ostreambuf(callback));
+    return 0;
+  } catch (...) {
+    std::cerr << "unknown exception in C++ set_print_callback()" << std::endl;
+  }
+  return -1;
+}
+
