+++ src/model_rng.hpp
@@ -10,7 +10,7 @@
 /**
  * This structure holds a pointer to a model and holds pointers to the parameter
  * names in CSV format.  Instances can be constructed with the C function
- * `bs_construct()` and destroyed with the C function `destruct()`.
+ * `bs_construct()` and destroyed with the C function `bs_destruct()`.
  */
 class bs_model {
  public:
@@ -222,6 +222,12 @@ class bs_model {
   int param_unc_num_ = -1;
 };
 
+/**
+ * A wrapper around the Boost random number generator required
+ * by the Stan model's write_array methods. Instances can be
+ * constructed with the C function `bs_construct_rng()` and destroyed
+ * with the C function `bs_destruct_rng()`.
+ */
 class bs_rng {
  public:
   bs_rng(unsigned int seed) : rng_(seed) { rng_.discard(1); }
