+++ src/bridgestan.h
@@ -179,9 +179,9 @@ int bs_param_constrain(const bs_model* m, bool include_tp, bool include_gq,
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
-int bs_param_constrain_seed(const bs_model* mr, bool include_tp, bool include_gq,
-                            const double* theta_unc, double* theta,
-                            unsigned int seed, char** error_msg);
+int bs_param_constrain_seed(const bs_model* mr, bool include_tp,
+                            bool include_gq, const double* theta_unc,
+                            double* theta, unsigned int seed, char** error_msg);
 
 /**
  * Set the sequence of unconstrained parameters based on the
@@ -294,7 +294,9 @@ int bs_log_density_hessian(const bs_model* m, bool propto, bool jacobian,
                            double* hessian, char** error_msg);
 
 /**
- * Construct an RNG object to be used for param_constrain
+ * Construct an RNG object to be used in `bs_param_constrain`.
+ * This object is not thread safe and should be constructed and
+ * destructed for each thread.
  *
  * @param[in] seed seed for the RNG
  * @param[out] error_msg a pointer to a string that will be allocated if there
@@ -303,7 +305,7 @@ int bs_log_density_hessian(const bs_model* m, bool propto, bool jacobian,
 bs_rng* bs_construct_rng(unsigned int seed, char** error_msg);
 
 /**
- * Destruct an RNG object
+ * Destruct an RNG object.
  *
  * @param[in] rng pointer to RNG object
  * @param[out] error_msg a pointer to a string that will be allocated if there
