+++ julia/src/BridgeStan.jl
@@ -28,7 +28,7 @@ include("model.jl")
 include("compile.jl")
 
 """
-    StanModel(;stan_file, stanc_args=[], make_args=[], data="", seed=204, chain_id=0)
+    StanModel(;stan_file, stanc_args=[], make_args=[], data="", seed=204)
 
 Construct a StanModel instance from a `.stan` file, compiling if necessary.
 
@@ -40,7 +40,6 @@ StanModel(;
     make_args::AbstractVector{String} = String[],
     data::String = "",
     seed = 204,
-    chain_id = 0,
-) = StanModel(compile_model(stan_file; stanc_args, make_args), data, seed, chain_id)
+) = StanModel(compile_model(stan_file; stanc_args, make_args), data, seed)
 
 end
