### 2023-03-29 e030ec6
Pass null in destruct in Julia/R
