### 2023-03-29 6e50958
Use nullptr in destruct in bridgestanR.cpp
