### 2023-03-30 a7e4d8f
Lock for callback, general cleanup
