+++ src/callback_stream.hpp
@@ -2,6 +2,7 @@
 #define CALLBACK_STREAM_H
 
 #include <streambuf>
+#include <mutex>
 
 typedef void (*STREAM_CALLBACK)(const char* data, size_t size);
 
@@ -10,11 +11,13 @@ struct callback_ostreambuf : public std::streambuf {
 
  protected:
   std::streamsize xsputn(const char_type* s, std::streamsize n) override {
+    std::lock_guard<std::mutex> lock(callback_mutex);
     callback(s, n);
     return n;
   };
 
   int_type overflow(int_type ch) {
+    std::lock_guard<std::mutex> lock(callback_mutex);
     if (ch != traits_type::eof()) {
       char c = ch;
       callback(&c, 1);
@@ -24,6 +27,7 @@ struct callback_ostreambuf : public std::streambuf {
 
  private:
   STREAM_CALLBACK callback;
+  std::mutex callback_mutex;
 };
 
 #endif
