### 2023-04-03 614b454
Expose version numbers in C API
