### 2023-04-03 7304b04
Rework to use model seed (by default) and chain offset
