### 2023-04-04 7dcd31b
Expose version numbers in C API (#101)
* Expose version numbers in C API

* Also add warning to R
