+++ python/bridgestan/__version.py
@@ -1,1 +1,2 @@
 __version__ = "1.0.2"
+__version_info__ = tuple(map(int, __version__.split(".")))
