### 2023-04-04 d8a11d0
Consistency of handle_error functions
