+++ julia/src/model.jl
@@ -71,7 +71,7 @@ mutable struct StanModel
             err,
         )
         if stanmodel == C_NULL
-            error(_handle_error(lib, err, "bs_construct"))
+            error(handle_error(lib, err, "bs_construct"))
         end
 
         sm = new(lib, stanmodel, data, seed, chain_id)
@@ -255,7 +255,7 @@ function param_constrain!(
         err,
     )
     if rc != 0
-        error(_handle_error(sm.lib, err, "param_constrain"))
+        error(handle_error(sm.lib, err, "param_constrain"))
     end
     out
 end
@@ -316,7 +316,7 @@ function param_unconstrain!(sm::StanModel, theta::Vector{Float64}, out::Vector{F
         err,
     )
     if rc != 0
-        error(_handle_error(sm.lib, err, "param_unconstrain"))
+        error(handle_error(sm.lib, err, "param_unconstrain"))
     end
     out
 end
@@ -371,7 +371,7 @@ function param_unconstrain_json!(sm::StanModel, theta::String, out::Vector{Float
         err,
     )
     if rc != 0
-        error(_handle_error(sm.lib, err, "param_unconstrain_json"))
+        error(handle_error(sm.lib, err, "param_unconstrain_json"))
     end
     out
 end
@@ -415,7 +415,7 @@ function log_density(sm::StanModel, q::Vector{Float64}; propto = true, jacobian 
         err,
     )
     if rc != 0
-        error(_handle_error(sm.lib, err, "log_density"))
+        error(handle_error(sm.lib, err, "log_density"))
     end
     lp[]
 end
@@ -470,7 +470,7 @@ function log_density_gradient!(
         err,
     )
     if rc != 0
-        error(_handle_error(sm.lib, err, "log_density_gradient"))
+        error(handle_error(sm.lib, err, "log_density_gradient"))
     end
     (lp[], out)
 end
@@ -557,7 +557,7 @@ function log_density_hessian!(
         err,
     )
     if rc != 0
-        error(_handle_error(sm.lib, err, "log_density_hessian"))
+        error(handle_error(sm.lib, err, "log_density_hessian"))
     end
     (lp[], out_grad, reshape(out_hess, (dims, dims)))
 end
@@ -587,11 +587,11 @@ function log_density_hessian(
 end
 
 """
-    _handle_error(sm::StanModel, err::Ref{Cstring}, method::String)
+    handle_error(sm::StanModel, err::Ref{Cstring}, method::String)
 
 Retrieves the error message allocated in C++ and frees it before returning a copy.
 """
-function _handle_error(lib::Ptr{Nothing}, err::Ref{Cstring}, method::String)
+function handle_error(lib::Ptr{Nothing}, err::Ref{Cstring}, method::String)
     if err[] == C_NULL
         return "Unknown error in $method."
     else
