### 2023-04-04 5b2c8b9
Cleanup: bs_destruct cannot error
