+++ src/bridgestanR.cpp
@@ -9,14 +9,12 @@ void bs_construct_R(char** data, int* rng, int* chain, bs_model_rng** ptr_out,
 void bs_free_error_msg_R(void** err_msg) {
   bs_free_error_msg(static_cast<char*>(*err_msg));
 }
-void bs_version_R(int* major, int* minor, int* patch){
+void bs_version_R(int* major, int* minor, int* patch) {
   *major = bs_major_version;
   *minor = bs_minor_version;
   *patch = bs_patch_version;
 }
-void bs_destruct_R(bs_model_rng** model, int* return_code) {
-  *return_code = bs_destruct(*model, nullptr);
-}
+void bs_destruct_R(bs_model_rng** model) { bs_destruct(*model); }
 void bs_name_R(bs_model_rng** model, char const** name_out) {
   *name_out = bs_name(*model);
 }
