### 2023-04-04 c3a67eb
Update julia/src/model.jl
Co-authored-by: Edward A. Roualdes <eroualdes@csuchico.edu>
