### 2023-04-04 983db93
Breaking change: Propagate error messages through C API (#91)
* Start sketch of error messaging

* Return error messages in Python, Julia

* Error handling in R

* Test that error messages are recieved by interfaces

* Clean up

* Make test Julia 1.6 safe

* Consistent naming of helper function

* Syntax fix

* Uncomment slower Julia tests

* Pass null in destruct in Julia/R

* Use nullptr in destruct in bridgestanR.cpp

* Consistency of handle_error functions

* Cleanup: bs_destruct cannot error

* Update julia/src/model.jl

Co-authored-by: Edward A. Roualdes <eroualdes@csuchico.edu>

-------
[…]
