+++ julia/src/model.jl
@@ -59,16 +59,19 @@ mutable struct StanModel
 
         lib = Libc.Libdl.dlopen(lib)
 
+        err = Ref{Cstring}()
+
         stanmodel = ccall(
             Libc.Libdl.dlsym(lib, "bs_construct"),
             Ptr{StanModelStruct},
-            (Cstring, UInt32, UInt32),
+            (Cstring, UInt32, UInt32, Ref{Cstring}),
             data,
             seed,
             chain_id,
+            err,
         )
         if stanmodel == C_NULL
-            error("could not construct model RNG")
+            error(handle_error(lib, err, "bs_construct"))
         end
 
         sm = new(lib, stanmodel, data, seed, chain_id)
@@ -76,7 +79,7 @@ mutable struct StanModel
         function f(sm)
             ccall(
                 Libc.Libdl.dlsym(sm.lib, "bs_destruct"),
-                UInt32,
+                Cvoid,
                 (Ptr{StanModelStruct},),
                 sm.stanmodel,
             )
@@ -238,18 +241,20 @@ function param_constrain!(
             DimensionMismatch("out must be same size as number of constrained parameters"),
         )
     end
+    err = Ref{Cstring}()
     rc = ccall(
         Libc.Libdl.dlsym(sm.lib, "bs_param_constrain"),
         Cint,
-        (Ptr{StanModelStruct}, Cint, Cint, Ref{Cdouble}, Ref{Cdouble}),
+        (Ptr{StanModelStruct}, Cint, Cint, Ref{Cdouble}, Ref{Cdouble}, Ref{Cstring}),
         sm.stanmodel,
         include_tp,
         include_gq,
         theta_unc,
         out,
+        err,
     )
     if rc != 0
-        error("param_constrain failed on C++ side; see stderr for messages")
+        error(handle_error(sm.lib, err, "param_constrain"))
     end
     out
 end
@@ -299,17 +304,18 @@ function param_unconstrain!(sm::StanModel, theta::Vector{Float64}, out::Vector{F
             ),
         )
     end
-
+    err = Ref{Cstring}()
     rc = ccall(
         Libc.Libdl.dlsym(sm.lib, "bs_param_unconstrain"),
         Cint,
-        (Ptr{StanModelStruct}, Ref{Cdouble}, Ref{Cdouble}),
+        (Ptr{StanModelStruct}, Ref{Cdouble}, Ref{Cdouble}, Ref{Cstring}),
         sm.stanmodel,
         theta,
         out,
+        err,
     )
     if rc != 0
-        error("param_unconstrain failed on C++ side; see stderr for messages")
+        error(handle_error(sm.lib, err, "param_unconstrain"))
     end
     out
 end
@@ -353,16 +359,18 @@ function param_unconstrain_json!(sm::StanModel, theta::String, out::Vector{Float
         )
     end
 
+    err = Ref{Cstring}()
     rc = ccall(
         Libc.Libdl.dlsym(sm.lib, "bs_param_unconstrain_json"),
         Cint,
-        (Ptr{StanModelStruct}, Cstring, Ref{Cdouble}),
+        (Ptr{StanModelStruct}, Cstring, Ref{Cdouble}, Ref{Cstring}),
         sm.stanmodel,
         theta,
         out,
+        err,
     )
     if rc != 0
-        error("param_unconstrain_json failed on C++ side; see stderr for messages")
+        error(handle_error(sm.lib, err, "param_unconstrain_json"))
     end
     out
 end
@@ -393,18 +401,20 @@ and includes change of variables terms for constrained parameters if `jacobian` 
 """
[…]
