### 2023-04-04 65bb3d7
Expose seed AND chain in C API, keep higher-level APIs as prior commit
