+++ src/bridgestanR.cpp
@@ -42,12 +42,12 @@ void bs_param_constrain_R(bs_model** model, int* include_tp, int* include_gq,
                                     theta, *rng, err_msg);
   *err_ptr = static_cast<void*>(*err_msg);
 }
-void bs_param_constrain_id_R(bs_model** model, int* include_tp,
+void bs_param_constrain_seeded_R(bs_model** model, int* include_tp,
                                int* include_gq, const double* theta_unc,
-                               double* theta, int* chain_id, int* return_code,
+                               double* theta, int* seed, int* chain_id, int* return_code,
                                char** err_msg, void** err_ptr) {
-  *return_code = bs_param_constrain_id(*model, *include_tp, *include_gq,
-                                         theta_unc, theta, *chain_id, err_msg);
+  *return_code = bs_param_constrain_seeded(*model, *include_tp, *include_gq,
+                                         theta_unc, theta, *seed, *chain_id, err_msg);
   *err_ptr = static_cast<void*>(*err_msg);
 }
 void bs_param_unconstrain_R(bs_model** model, const double* theta,
