+++ src/bridgestan.h
@@ -87,8 +87,7 @@ const char* bs_model_info(const bs_model* m);
  * @param[in] include_gq `true` to include generated quantities
  * @return CSV-separated, indexed, parameter names
  */
-const char* bs_param_names(const bs_model* m, bool include_tp,
-                           bool include_gq);
+const char* bs_param_names(const bs_model* m, bool include_tp, bool include_gq);
 
 /**
  * Return a comma-separated sequence of unconstrained parameters.
@@ -181,9 +180,10 @@ int bs_param_constrain(const bs_model* m, bool include_tp, bool include_gq,
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
-int bs_param_constrain_seeded(const bs_model* mr, bool include_tp, bool include_gq,
-                          const double* theta_unc, double* theta, unsigned int seed,
-                          unsigned int chain_id, char** error_msg);
+int bs_param_constrain_seeded(const bs_model* mr, bool include_tp,
+                              bool include_gq, const double* theta_unc,
+                              double* theta, unsigned int seed,
+                              unsigned int chain_id, char** error_msg);
 
 /**
  * Set the sequence of unconstrained parameters based on the
