+++ src/bridgestan.cpp
@@ -79,9 +79,9 @@ int bs_param_constrain(const bs_model* m, bool include_tp, bool include_gq,
 }
 
 int bs_param_constrain_seeded(const bs_model* m, bool include_tp,
-                            bool include_gq, const double* theta_unc,
-                            double* theta, unsigned int seed, unsigned int chain_id,
-                            char** error_msg) {
+                              bool include_gq, const double* theta_unc,
+                              double* theta, unsigned int seed,
+                              unsigned int chain_id, char** error_msg) {
   bs_rng rng(seed, chain_id);
   return bs_param_constrain(m, include_tp, include_gq, theta_unc, theta, &rng,
                             error_msg);
@@ -201,7 +201,8 @@ int bs_log_density_hessian(const bs_model* m, bool propto, bool jacobian,
   return -1;
 }
 
-bs_rng* bs_construct_rng(unsigned int seed, unsigned int chain_id, char** error_msg) {
+bs_rng* bs_construct_rng(unsigned int seed, unsigned int chain_id,
+                         char** error_msg) {
   try {
     return new bs_rng(seed, chain_id);
   } catch (const std::exception& e) {
