+++ src/bridgestanR.h
@@ -44,9 +44,10 @@ void bs_param_constrain_R(bs_model** model, int* include_tp, int* include_gq,
                           int* return_code, char** err_msg, void** err_ptr);
 
 void bs_param_constrain_seeded_R(bs_model** model, int* include_tp,
-                               int* include_gq, const double* theta_unc,
-                               double* theta, int* seed, int* chain_id, int* return_code,
-                               char** err_msg, void** err_ptr);
+                                 int* include_gq, const double* theta_unc,
+                                 double* theta, int* seed, int* chain_id,
+                                 int* return_code, char** err_msg,
+                                 void** err_ptr);
 
 void bs_param_unconstrain_R(bs_model** model, const double* theta,
                             double* theta_unc, int* return_code, char** err_msg,
@@ -70,8 +71,8 @@ void bs_log_density_hessian_R(bs_model** model, int* propto, int* jacobian,
                               double* grad, double* hess, int* return_code,
                               char** err_msg, void** err_ptr);
 
-void bs_construct_rng_R(int* seed, int* chain_id, bs_rng** ptr_out, char** err_msg,
-                        void** err_ptr);
+void bs_construct_rng_R(int* seed, int* chain_id, bs_rng** ptr_out,
+                        char** err_msg, void** err_ptr);
 
 void bs_destruct_rng_R(bs_rng** rng, int* return_code);
 

[…]
