### 2023-04-04 5f9f7fa
Rebase rust interface to use rng and error message
