+++ rust-example/src/bs_safe.rs
@@ -1,85 +1,226 @@
 use crate::ffi;
+use std::ffi::c_char;
+use std::ffi::c_int;
 use std::ffi::CStr;
 use std::ffi::CString;
 use std::ffi::NulError;
+use std::ptr::null_mut;
+use std::ptr::NonNull;
 use std::str::Utf8Error;
 
 // This is more or less equivalent to manually defining Display and From<other error types>
 use thiserror::Error;
 
-#[derive(Error, Debug, PartialEq, Eq)]
+#[derive(Error, Debug)]
 pub enum BridgeStanError {
+    #[error("Could not load target library: {0}")]
+    LoadLibraryError(#[from] libloading::Error),
+    #[error("Bad Stan library version: Got {0} but expected {1}")]
+    BadLibraryVersion(String, String),
+    #[error("The Stan library was compiled without threading support")]
+    StanThreadsError(),
     #[error("failed to encode string to null-terminated C string")]
     StringEncodeError(#[from] NulError),
     #[error("failed to decode string to UTF8")]
     StringDecodeError(#[from] Utf8Error),
-    #[error("failed to allocate model_rng")]
-    AllocateFailedError,
-    #[error("failed evaluate function, see C stderr")]
-    EvaluationFailed,
+    #[error("failed to construct model: {0}")]
+    ConstructFailedError(String),
+    #[error("failed during evaluation: {0}")]
+    EvaluationFailed(String),
 }
 
-pub fn open_library(path: &std::path::Path) -> Result<ffi::Bridgestan, libloading::Error> {
-    unsafe { ffi::Bridgestan::new(path) }
+type Result<T> = std::result::Result<T, BridgeStanError>;
+
+pub fn open_library(path: &std::path::Path) -> Result<ffi::Bridgestan> {
+    let library = unsafe { libloading::Library::new(path) }?;
+    let major: libloading::Symbol<*const c_int> = unsafe { library.get(b"bs_major_version") }?;
+    let major = unsafe { **major };
+    let minor: libloading::Symbol<*const c_int> = unsafe { library.get(b"bs_minor_version") }?;
+    let minor = unsafe { **minor };
+    let patch: libloading::Symbol<*const c_int> = unsafe { library.get(b"bs_patch_version") }?;
+    let patch = unsafe { **patch };
+    let self_major: c_int = env!("CARGO_PKG_VERSION_MAJOR").parse().unwrap();
+    let self_minor: c_int = env!("CARGO_PKG_VERSION_MINOR").parse().unwrap();
+    let self_patch: c_int = env!("CARGO_PKG_VERSION_PATCH").parse().unwrap();
+
+    if (self_major != major) | (self_minor != minor) {
+        return Err(BridgeStanError::BadLibraryVersion(
+            format!("{}.{}.{}", major, minor, patch),
+            format!("{}.{}.{}", self_major, self_minor, self_patch),
+        ));
+    }
+    Ok(unsafe { ffi::Bridgestan::from_library(library) }?)
 }
 
-/// Safe wrapper for BridgeStan C functions
-#[non_exhaustive]
-pub struct StanModel<'lib> {
-    model: *mut ffi::bs_model_rng,
+// TODO for C Api
+//
+// * Access generated data?
+// * integer generated values? (expand and generated data)
+// * StanThreads?
+// * Should bs_destruct_rng return an int? Should we panic if that happens?
+// * Return &[u8] or assume utf8 and convert to str?
+// * param_unconstrain: What lengths are allowed (ie include_tp, inclued_gq)
+
+/// A Stan model instance with data
+pub struct Model<'lib> {
+    model: NonNull<ffi::bs_model>,
[…]
