### 2023-04-04 4b43133
Rebase rust interface to use rng and error message
