+++ rust-example/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod bs_safe;
 pub(crate) mod ffi;
 
-pub use bs_safe::{StanModel, BridgeStanError, open_library};
+pub use bs_safe::{open_library, BridgeStanError, Model};
