### 2023-04-05 030aeeb
Clean up python callback registration
