+++ rust/src/ffi.rs
@@ -1,4 +1,5 @@
 #![allow(non_camel_case_types)]
 #![allow(dead_code)]
+#![allow(clippy::all)]
 // Include generated bindings file
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
