### 2023-04-05 5c7b151
Use Borrow to store reference to stan library
