+++ rust/src/main.rs
@@ -1,8 +1,9 @@
-use bridgestan::Model;
+use bridgestan::{BridgeStanError, Model};
 use std::env;
 use std::error::Error;
 use std::path::Path;
-use std::thread;
+use std::sync::Arc;
+use std::thread::{self, JoinHandle};
 
 fn main() -> Result<(), Box<dyn Error>> {
     let lib_path = env::args()
@@ -10,37 +11,45 @@ fn main() -> Result<(), Box<dyn Error>> {
         .expect("Required to pass a path to the library!");
     let lib = bridgestan::open_library(Path::new(&lib_path))?;
 
-    let data_path = env::args().nth(2).unwrap_or_default();
+    let lib = Arc::new(lib);
 
-    let model = Model::new(&lib, &data_path, 123)?;
+    let out: JoinHandle<Result<_, BridgeStanError>> = thread::spawn(move || {
+        let data_path = env::args().nth(2).unwrap_or_default();
 
-    println!(
-        "The model has {} parameters.",
-        model.param_num(false, false)
-    );
+        let model = Model::new(lib, &data_path, 123)?;
 
-    let name = model.name().unwrap();
-    println!("The model's name is {}.", name);
+        println!(
+            "The model has {} parameters.",
+            model.param_num(false, false)
+        );
 
-    // parallel evaluation of log_density_gradient
+        let name = model.name().unwrap();
+        println!("The model's name is {}.", name);
 
-    let n = model.param_unc_num();
+        // parallel evaluation of log_density_gradient
 
-    // location for the results
-    let mut grad = vec![vec![0.0; n]; 25];
-    let theta = (1..n + 1).map(|s| s as f64).collect::<Vec<f64>>();
+        let n = model.param_unc_num();
 
-    thread::scope(|s| {
-        for g in grad.iter_mut() {
-            s.spawn(|| {
-                model
-                    .log_density_gradient(theta.as_slice(), true, true, g.as_mut_slice())
-                    .unwrap();
-            });
-        }
+        // location for the results
+        let mut grad = vec![vec![0.0; n]; 25];
+        let theta = (1..n + 1).map(|s| s as f64).collect::<Vec<f64>>();
+
+        thread::scope(|s| {
+            for g in grad.iter_mut() {
+                s.spawn(|| {
+                    model
+                        .log_density_gradient(theta.as_slice(), true, true, g.as_mut_slice())
+                        .unwrap();
+                });
+            }
+        });
+
+        println!("{:?}", &grad);
+
+        Ok(())
     });
 
-    println!("{:?}", &grad);
+    out.join().unwrap().unwrap();
 
     Ok(())
 }
