+++ rust/src/bs_safe.rs
@@ -1,4 +1,5 @@
 use crate::ffi;
+use std::borrow::Borrow;
 use std::ffi::c_char;
 use std::ffi::c_int;
 use std::ffi::CStr;
@@ -10,6 +11,8 @@ use std::str::Utf8Error;
 
 // This is more or less equivalent to manually defining Display and From<other error types>
 use thiserror::Error;
+
+pub use ffi::Bridgestan as StanLibrary;
 
 #[derive(Error, Debug)]
 pub enum BridgeStanError {
@@ -31,7 +34,7 @@ pub enum BridgeStanError {
 
 type Result<T> = std::result::Result<T, BridgeStanError>;
 
-pub fn open_library(path: &std::path::Path) -> Result<ffi::Bridgestan> {
+pub fn open_library(path: &std::path::Path) -> Result<StanLibrary> {
     let library = unsafe { libloading::Library::new(path) }?;
     let major: libloading::Symbol<*const c_int> = unsafe { library.get(b"bs_major_version") }?;
     let major = unsafe { **major };
@@ -49,7 +52,7 @@ pub fn open_library(path: &std::path::Path) -> Result<ffi::Bridgestan> {
             format!("{}.{}.{}", self_major, self_minor, self_patch),
         ));
     }
-    Ok(unsafe { ffi::Bridgestan::from_library(library) }?)
+    Ok(unsafe { StanLibrary::from_library(library) }?)
 }
 
 // TODO for C Api
@@ -60,21 +63,22 @@ pub fn open_library(path: &std::path::Path) -> Result<ffi::Bridgestan> {
 // * Should bs_destruct_rng return an int? Should we panic if that happens?
 // * Return &[u8] or assume utf8 and convert to str?
 // * param_unconstrain: What lengths are allowed (ie include_tp, inclued_gq)
+// * Put lib in Arc?
 
 /// A Stan model instance with data
-pub struct Model<'lib> {
+pub struct Model<T: Borrow<StanLibrary>> {
     model: NonNull<ffi::bs_model>,
-    lib: &'lib ffi::Bridgestan,
+    lib: T,
 }
 
 // Stan model is thread safe
-unsafe impl<'lib> Sync for Model<'lib> {}
-unsafe impl<'lib> Send for Model<'lib> {}
+unsafe impl<T: Sync + Borrow<StanLibrary>> Sync for Model<T> {}
+unsafe impl<T: Send + Borrow<StanLibrary>> Send for Model<T> {}
 
 /// A random number generator for Stan
 pub struct Rng<'lib> {
     rng: NonNull<ffi::bs_rng>,
-    lib: &'lib ffi::Bridgestan,
+    lib: &'lib StanLibrary,
 }
 
 impl<'lib> Drop for Rng<'lib> {
@@ -88,7 +92,7 @@ impl<'lib> Drop for Rng<'lib> {
 
 struct ErrorMsg<'lib> {
     msg: *mut c_char,
-    lib: &'lib ffi::Bridgestan,
+    lib: &'lib StanLibrary,
 }
 
 impl<'lib> Drop for ErrorMsg<'lib> {
@@ -100,7 +104,7 @@ impl<'lib> Drop for ErrorMsg<'lib> {
 }
 
 impl<'lib> ErrorMsg<'lib> {
-    fn new(lib: &'lib ffi::Bridgestan) -> Self {
+    fn new(lib: &'lib StanLibrary) -> Self {
         Self {
             msg: std::ptr::null_mut(),
             lib,
@@ -125,41 +129,44 @@ impl<'lib> ErrorMsg<'lib> {
     }
 }
 
-impl<'lib> Model<'lib> {
+impl<T: Borrow<StanLibrary>> Model<T> {
     /// Create a new instance of the compiled Stan model.
[…]
