### 2023-04-05 823109f
Change Rng to use Borrow
