### 2023-04-05 c52efe9
Change Rng to use Borrow
