### 2023-04-06 c9b0b08
Update to reflect that destructor cannot fail
