+++ src/bridgestanR.h
@@ -74,7 +74,7 @@ void bs_log_density_hessian_R(bs_model** model, int* propto, int* jacobian,
 void bs_construct_rng_R(int* seed, int* chain_id, bs_rng** ptr_out,
                         char** err_msg, void** err_ptr);
 
-void bs_destruct_rng_R(bs_rng** rng, int* return_code);
+void bs_destruct_rng_R(bs_rng** rng);
 
 #ifdef __cplusplus
 }
