### 2023-04-06 519a644
Allow nullptr for data string
