### 2023-04-06 950b674
Explicitly handle NULL in error messages
