### 2023-04-06 e834601
Accept optional data string in rust interface
