### 2023-04-07 8c2df9a
Add some convenience function in rust intarface
