+++ rust/src/bs_safe.rs
@@ -11,6 +11,7 @@ use std::ptr::null_mut;
 use std::ptr::NonNull;
 use std::str::Utf8Error;
 
+use itertools::Itertools;
 // This is more or less equivalent to manually defining Display and From<other error types>
 use thiserror::Error;
 
@@ -22,16 +23,18 @@ pub enum BridgeStanError {
     LoadLibraryError(#[from] libloading::Error),
     #[error("Bad Stan library version: Got {0} but expected {1}")]
     BadLibraryVersion(String, String),
-    #[error("The Stan library was compiled without threading support")]
-    StanThreadsError(),
-    #[error("failed to encode string to null-terminated C string")]
+    #[error("The Stan library was compiled without threading support. Config was {0}")]
+    StanThreadsError(String),
+    #[error("Failed to encode string to null-terminated C string")]
     StringEncodeError(#[from] NulError),
-    #[error("failed to decode string to UTF8")]
+    #[error("Failed to decode string to UTF8")]
     StringDecodeError(#[from] Utf8Error),
-    #[error("failed to construct model: {0}")]
+    #[error("Failed to construct model: {0}")]
     ConstructFailedError(String),
-    #[error("failed during evaluation: {0}")]
+    #[error("Failed during evaluation: {0}")]
     EvaluationFailed(String),
+    #[error("Failed to interpret variable names")]
+    InvalidVariableNames(),
 }
 
 type Result<T> = std::result::Result<T, BridgeStanError>;
@@ -75,6 +78,7 @@ pub fn open_library<P: AsRef<OsStr>>(path: P) -> Result<StanLibrary> {
 pub struct Model<T: Borrow<StanLibrary>> {
     model: NonNull<ffi::bs_model>,
     lib: T,
+    seed: u32,
 }
 
 // Stan model is thread safe
@@ -86,6 +90,9 @@ pub struct Rng<T: Borrow<StanLibrary>> {
     rng: NonNull<ffi::bs_rng>,
     lib: T,
 }
+
+unsafe impl<T: Sync + Borrow<StanLibrary>> Sync for Rng<T> {}
+unsafe impl<T: Send + Borrow<StanLibrary>> Send for Rng<T> {}
 
 impl<T: Borrow<StanLibrary>> Drop for Rng<T> {
     fn drop(&mut self) {
@@ -154,6 +161,23 @@ impl<'lib> ErrorMsg<'lib> {
     }
 }
 
+#[non_exhaustive]
+#[derive(Clone, Debug)]
+pub struct Parameter {
+    pub name: String,
+    pub shape: Vec<usize>,
+    pub size: usize,
+    pub start_idx: usize,
+    pub end_idx: usize,
+}
+
+impl<T: Borrow<StanLibrary> + Clone> Model<T> {
+    /// Return a clone of the underlying stan library
+    pub fn clone_library(&self) -> T {
+        self.lib.clone()
+    }
+}
+
 impl<T: Borrow<StanLibrary>> Model<T> {
     /// Create a new instance of the compiled Stan model.
     /// Data is specified as a JSON file at the given path, or empty for no data
@@ -171,18 +195,27 @@ impl<T: Borrow<StanLibrary>> Model<T> {
 
         if let Some(model) = NonNull::new(model) {
             drop(err);
-            let model = Self { model, lib };
+            let model = Self { model, lib, seed };
             // If STAN_THREADS is not true, the safty guaranties we are
             // making would be incorrect
             let info = model.info()?;
             if !info.contains("STAN_THREADS=true") {
-                Err(BridgeStanError::StanThreadsError())
+                Err(BridgeStanError::StanThreadsError(info.to_string()))
             } else {
                 Ok(model)
             }
[…]
