+++ rust/src/lib.rs
@@ -1,4 +1,4 @@
 mod bs_safe;
 pub(crate) mod ffi;
 
-pub use bs_safe::{open_library, BridgeStanError, Model, Rng};
+pub use bs_safe::{open_library, BridgeStanError, Model, Parameter, Rng, StanLibrary};
