+++ rust/Cargo.toml
@@ -4,6 +4,7 @@ version = "1.0.2"
 edition = "2021"
 
 [dependencies]
+itertools = "0.10.5"
 libloading = "0.7.4"
 nuts-rs = { version = "0.4.1", optional = true }
 thiserror = "1.0.40"
