### 2023-04-11 9b2fc59
Allow nullptr as data argument to constructor (#104)
* Allow nullptr for data string

* Explicitly handle NULL in error messages

* Update doc comments
