+++ python/bridgestan/model.py
@@ -20,17 +20,6 @@ class StanModel:
     A StanModel instance encapsulates a Stan model instantiated with data
     and provides methods to access parameter names, transforms, log
     densities, gradients, and Hessians.
-
-    The constructor method instantiates a Stan model and sets constant
-    return values.  The constructor arguments are
-
-    :param model_lib: A path to a compiled shared object.
-    :param model_data: Either a JSON string literal or a
-         path to a data file in JSON format ending in ``.json``.
-    :param seed: A pseudo random number generator seed.
-    :raises FileNotFoundError or PermissionError: If ``model_lib`` is not readable or
-        ``model_data`` is specified and not a path to a readable file.
-    :raises RuntimeError: If there is an error instantiating the Stan model.
     """
 
     def __init__(
@@ -41,13 +30,15 @@ class StanModel:
         seed: int = 1234,
     ) -> None:
         """
-        Construct a StanModel object for a Stan model and data given
+        Construct a StanModel object for a compiled Stan model and data given
         constructor arguments.
 
         :param model_lib: A system path to compiled shared object.
-        :param model_data: Either a JSON string literal or a
-            system path to a data file in JSON format ending in ``.json``.
-        :param seed: A pseudo random number generator seed.
+        :param model_data: Either a JSON string literal, a
+            system path to a data file in JSON format ending in ``.json``,
+            or the empty string.
+        :param seed: A pseudo random number generator seed, used for RNG functions
+            in the ``transformed data`` block.
         :raises FileNotFoundError or PermissionError: If ``model_lib`` is not readable or
             ``model_data`` is specified and not a path to a readable file.
         :raises RuntimeError: If there is an error instantiating the
@@ -74,9 +65,9 @@ class StanModel:
         self._free_error.argtypes = [ctypes.c_char_p]
 
         err = ctypes.pointer(ctypes.c_char_p())
-        self.model_rng = self._construct(str.encode(self.data_path), self.seed, err)
+        self.model = self._construct(str.encode(self.data_path), self.seed, err)
 
-        if not self.model_rng:
+        if not self.model:
             raise self._handle_error(err.contents, "bs_construct")
 
         if self.model_version() != __version_info__:
@@ -221,8 +212,8 @@ class StanModel:
         """
         Destroy the Stan model and free memory.
         """
-        if hasattr(self, "model_rng") and hasattr(self, "_destruct"):
-            self._destruct(self.model_rng)
+        if hasattr(self, "model") and hasattr(self, "_destruct"):
+            self._destruct(self.model)
 
     def __repr__(self) -> str:
         data = f"{self.data_path!r}, " if self.data_path else ""
@@ -234,7 +225,7 @@ class StanModel:
 
         :return: The name of Stan model.
         """
-        return self._name(self.model_rng).decode("utf-8")
+        return self._name(self.model).decode("utf-8")
 
     def model_info(self) -> str:
         """
@@ -244,7 +235,7 @@ class StanModel:
 
         :return: Information about the compiled Stan model.
         """
-        return self._model_info(self.model_rng).decode("utf-8")
+        return self._model_info(self.model).decode("utf-8")
 
     def model_version(self) -> Tuple[int, int, int]:
         """
@@ -265,7 +256,7 @@ class StanModel:
         :param include_gq: ``True`` to include the generated quantities.
         :return: The number of parameters.
         """
-        return self._param_num(self.model_rng, int(include_tp), int(include_gq))
+        return self._param_num(self.model, int(include_tp), int(include_gq))
 
     def param_unc_num(self) -> int:
[…]
