### 2023-04-13 d22191f
Consistently name construct/destruct functions
