### 2023-04-14 38f9ac0
Breaking change: Implement param_constrain in a thread-safe way (#98)
* Start C++ changes for moving RNGs outside model object

* Checkpoint

* Checkpointing; rebased on error msg PR to avoid stepping on myself

* Update doc comments

* Update Python tests

* Start julia

* Update Julia tests, doc

* R interface work

* Documentation

* Update c-example

* Rework to use model seed (by default) and chain offset

* Expose seed AND chain in C API, keep higher-level APIs as prior commit

* Formatting

* Update to reflect that destructor cannot fail

* Typo fix

* Simplify interface; remove concept
[…]
