+++ julia/src/model.jl
@@ -1,5 +1,7 @@
 
-mutable struct StanModelStruct end
+struct StanModelStruct end
+
+mutable struct StanRNGStruct end
 
 # utility macro to annotate a field as const only if supported
 @eval macro $(Symbol("const"))(x)
@@ -11,19 +13,20 @@ mutable struct StanModelStruct end
 end
 
 """
-    StanModel(lib, datafile="", seed=204, chain_id=0)
+    StanModel(lib, datafile="", seed=204)
 
 A StanModel instance encapsulates a Stan model instantiated with data.
 
 Construct a Stan model from the supplied library file path and data. Data
-should either be a string containing a JSON string literal or a path to a data file ending in `.json`.
-If seed or chain_id are supplied, these are used to initialize the RNG used by the model.
+should either be a string containing a JSON string literal, a path to a data file ending in `.json`,
+or the empty string.
+If seed is supplied, it is used to initialize the RNG used by the model's constructor.
 
-    StanModel(;stan_file, data="", seed=204, chain_id=0)
+    StanModel(;stan_file, data="", seed=204)
 
 Construct a `StanModel` instance from a `.stan` file, compiling if necessary.
 
-    StanModel(;stan_file, stanc_args=[], make_args=[], data="", seed=204, chain_id=0)
+    StanModel(;stan_file, stanc_args=[], make_args=[], data="", seed=204)
 
 Construct a `StanModel` instance from a `.stan` file.  Compilation
 occurs if no shared object file exists for the supplied Stan file or
@@ -36,11 +39,9 @@ mutable struct StanModel
     stanmodel::Ptr{StanModelStruct}
     @const data::String
     @const seed::UInt32
-    @const chain_id::UInt32
 
-    function StanModel(lib::String, data::String = "", seed = 204, chain_id = 0)
+    function StanModel(lib::String, data::String = "", seed = 204)
         seed = convert(UInt32, seed)
-        chain_id = convert(UInt32, chain_id)
 
         if !isfile(lib)
             throw(SystemError("Dynamic library file not found"))
@@ -62,23 +63,22 @@ mutable struct StanModel
         err = Ref{Cstring}()
 
         stanmodel = ccall(
-            Libc.Libdl.dlsym(lib, "bs_construct"),
+            Libc.Libdl.dlsym(lib, "bs_model_construct"),
             Ptr{StanModelStruct},
-            (Cstring, UInt32, UInt32, Ref{Cstring}),
+            (Cstring, UInt32, Ref{Cstring}),
             data,
             seed,
-            chain_id,
             err,
         )
         if stanmodel == C_NULL
-            error(handle_error(lib, err, "bs_construct"))
+            error(handle_error(lib, err, "bs_model_construct"))
         end
 
-        sm = new(lib, stanmodel, data, seed, chain_id)
+        sm = new(lib, stanmodel, data, seed)
 
         function f(sm)
             ccall(
-                Libc.Libdl.dlsym(sm.lib, "bs_destruct"),
+                Libc.Libdl.dlsym(sm.lib, "bs_model_destruct"),
                 Cvoid,
                 (Ptr{StanModelStruct},),
                 sm.stanmodel,
@@ -86,6 +86,51 @@ mutable struct StanModel
         end
 
         finalizer(f, sm)
+    end
+end
+
+"""
+    StanRNG(sm::StanModel, seed)
+
+Construct a StanRNG instance from a `StanModel` instance and a seed.
+
+This can be used in the `param_constrain` and `param_constrain!` methods
+when using the generated quantities block.
+
+This object is not thread-safe, one should be created per thread.
+"""
+mutable struct StanRNG
[…]
