+++ python/test/test_stanmodel.py
@@ -43,7 +43,6 @@ def test_constructor():
         b4 = bs.StanModel(throw_data_so)
 
 
-
 def test_name():
     std_so = str(STAN_FOLDER / "stdnormal" / "stdnormal_model.so")
     b = bs.StanModel(std_so)
@@ -55,7 +54,6 @@ def test_model_info():
     b = bs.StanModel(std_so)
     assert "STAN_OPENCL" in b.model_info()
     assert "BridgeStan version: 1." in b.model_info()
-
 
 
 def test_param_num():
@@ -210,13 +208,26 @@ def test_param_constrain():
 
     full_so = str(STAN_FOLDER / "full" / "full_model.so")
     bridge2 = bs.StanModel(full_so)
+    rng = bridge2.new_rng(seed=1234)
 
     np.testing.assert_equal(1, bridge2.param_constrain(a).size)
     np.testing.assert_equal(2, bridge2.param_constrain(a, include_tp=True).size)
-    np.testing.assert_equal(3, bridge2.param_constrain(a, include_gq=True).size)
     np.testing.assert_equal(
-        4, bridge2.param_constrain(a, include_tp=True, include_gq=True).size
+        3, bridge2.param_constrain(a, include_gq=True, rng=rng).size
     )
+    np.testing.assert_equal(
+        4, bridge2.param_constrain(a, include_tp=True, include_gq=True, rng=rng).size
+    )
+
+    # reproducibility test
+    np.testing.assert_equal(
+        bridge2.param_constrain(a, include_gq=True, rng=bridge2.new_rng(seed=4567)),
+        bridge2.param_constrain(a, include_gq=True, rng=bridge2.new_rng(seed=4567)),
+    )
+
+    # test error if neither seed or rng is provided
+    with pytest.raises(ValueError):
+        bridge2.param_constrain(a, include_gq=True)
 
     # out tests, matched and mismatched
     scratch = np.zeros(16)
@@ -240,7 +251,7 @@ def test_param_constrain():
     bridge3 = bs.StanModel(throw_gq_so)
     bridge3.param_constrain(y, include_gq=False)
     with pytest.raises(RuntimeError, match="find this text: gqfails"):
-        bridge3.param_constrain(y, include_gq=True)
+        bridge3.param_constrain(y, include_gq=True, rng=bridge3.new_rng(seed=1))
 
 
 def test_param_unconstrain():
