+++ v1.0.1/_sources/languages/julia.md.txt
@@ -1,0 +1,491 @@
+
+<a id='Julia-Interface'></a>
+
+<a id='Julia-Interface-1'></a>
+
+# Julia Interface
+
+
+% NB: If you are reading this file in python/docs/languages, you are reading a generated output!
+% This should be apparent due to the html tags everywhere.
+% If you are reading this in julia/docs/src, you are reading the true source!
+% Please only make edits in the later file, since the first is DELETED each re-build.
+
+
+---
+
+
+<a id='Installation'></a>
+
+<a id='Installation-1'></a>
+
+## Installation
+
+
+<a id='From-JuliaRegistries'></a>
+
+<a id='From-JuliaRegistries-1'></a>
+
+### From JuliaRegistries
+
+
+BridgeStan is registered on JuliaRegistries each release.
+
+
+```julia
+] add BridgeStan
+```
+
+
+The first time you need it, the BridgeStan source code will be downloaded as an [Artifact](https://pkgdocs.julialang.org/v1/artifacts/). If you prefer to use a source distribution of BridgeStan, consult the following section.
+
+
+Note that the system pre-requisites from the [Getting Started Guide](../getting-started.rst) are still required and will not be automatically installed by this method.
+
+
+<a id='From-Source'></a>
+
+<a id='From-Source-1'></a>
+
+### From Source
+
+
+This section assumes you have followed the [Getting Started guide](../getting-started.rst) to install BridgeStan's pre-requisites and downloaded a copy of the BridgeStan source code.
+
+
+To install the Julia interface, you can either install it directly from Github by running the following inside a Julia REPL
+
+
+```julia
+] add https://github.com/roualdes/bridgestan.git:julia
+```
+
+
+Or, since you have already downloaded the repository, you can run
+
+
+```julia
+] dev julia/
+```
+
+
+from the BridgeStan folder.
+
+
+Note that the Julia package depends on Julia 1.8+.
+
+
+<a id='Example-Program'></a>
+
+<a id='Example-Program-1'></a>
+
+## Example Program
+
+
+An example program is provided alongside the Julia interface code in `example.jl`:
+
+
+<details>
+<summary><a>Show example.jl</a></summary>
+
+
+```{literalinclude} ../../julia/example.jl
+:language: julia
+```
+
+
+</details>
+
+
+<a id='API-Reference'></a>
+
+<a id='API-Reference-1'></a>
+
+## API Reference
+
+
+<a id='StanModel-interface'></a>
+
+<a id='StanModel-interface-1'></a>
+
+### StanModel interface
+
+<a id='BridgeStan.StanModel' href='#BridgeStan.StanModel'>#</a>
+**`BridgeStan.StanModel`** &mdash; *Type*.
+
+
+
+```julia
+StanModel(lib, datafile="", seed=204, chain_id=0)
+```
+
+A StanModel instance encapsulates a Stan model instantiated with data.
+
+Construct a Stan model from the supplied library file path and data. Data should either be a string containing a JSON string literal or a path to a data file ending in `.json`. If seed or chain_id are supplied, these are used to initialize the RNG used by the model.
+
+```
+StanModel(;stan_file, data="", seed=204, chain_id=0)
+```
+
+Construct a `StanModel` instance from a `.stan` file, compiling if necessary.
+
+```
+StanModel(;stan_file, stanc_args=[], make_args=[], data="", seed=204, chain_id=0)
+```
+
[…]
