+++ v2.1.0/_sources/languages/r.md.txt
@@ -1,0 +1,403 @@
+# R Interface
+
+% The skeleton of this page was made using Roxygen2 and tools::Rd2txt, but it is not automated.
+% Maybe it will be one day
+% Used:
+% roxygen2::roxygenize()
+% tools::Rd2txt_options(underline_titles=FALSE, itemBullet="* ", showURLs=TRUE)
+% tools::Rd2txt('./man/StanModel.Rd', 'r.txt')
+
+
+---
+
+## Installation
+
+This assumes you have followed the [Getting Started guide](../getting-started.rst)
+to install BridgeStan's pre-requisites and downloaded a copy of the BridgeStan source code.
+
+
+```R
+devtools::install_github("https://github.com/roualdes/bridgestan", subdir="R")
+```
+
+Or, since you have already downloaded the repository, you can run
+
+```R
+install.packages(file.path(getwd(),"R"), repos=NULL, type="source")
+```
+from the BridgeStan folder.
+
+Note that the R package depends on R 3+ and R6, and will install R6 if it is not
+already installed.
+
+```{note}
+Mac users have reported issues when using a copy of R installed from [conda-forge](https://conda-forge.org/). If you encounter an issue, you may need to use R from the official [R project website](https://www.r-project.org/) or a system package manager like `brew`.
+```
+
+## Example Program
+
+An example program is provided alongside the R interface code in `example.R`:
+
+<details>
+<summary><a>Show example.R</a></summary>
+
+```{literalinclude} ../../R/example.R
+:language: R
+```
+
+</details>
+
+## API Reference
+
+### *R6::R6Class* `StanModel`
+
+`R6` Class representing a compiled BridgeStan model.
+
+This model exposes log density, gradient, and Hessian information
+as well as constraining and unconstraining transforms.
+
+#### Methods
+
+**Method** `new()`:
+
+Create a Stan Model instance.
+
+_Usage_
+
+```R
+StanModel$new(lib, data, rng_seed)
+```
+
+
+_Arguments_
+
+  - `lib` A path to a compiled BridgeStan Shared Object file.
+
+  - `data` Either a JSON string literal, a path to a data file in JSON format ending in ".json", or the empty string.
+
+  - `rng_seed` Seed for the RNG used in constructing the model.
+
+_Returns_
+
+  A new StanModel.
+
+
+**Method** `name()`:
+
+Get the name of this StanModel.
+
+_Usage_
+
+```R
+StanModel$name()
+```
+
+
+_Returns_
+
+  A character vector of the name.
+
+
+**Method** `model_info()`:
+
+Get compile information about this Stan model.
+
+_Usage_
+
+```R
+StanModel$model_info()
+```
+
+
+_Returns_
+
+  A character vector of the Stan version and important flags.
+
+**Method** `param_names()`:
+
+Return the indexed names of the (constrained) parameters. For
+containers, indexes are separated by periods (.).
+
+For example, the scalar a has indexed name a, the vector entry
+a[1] has indexed name a.1 and the matrix entry a[2, 3] has
+indexed name a.2.3. Parameter order of the output is column
+major and more generally last-index major for containers.
+
+_Usage_
+
+```R
+StanModel$param_names(include_tp = FALSE, include_gq = FALSE)
+```
+
+
+_Arguments_
+
+  - `include_tp` Whether to include variables from transformed
+      parameters.
+
+  - `include_gq` Whether to include variables from generated
+      quantities.
+
+
+_Returns_
+
+  A list of character vectors of the names.
+
+
+**Method** `param_unc_names()`:
+
+Return the indexed names of the unconstrained parameters. For
[…]
