+++ v2.6.1/_sources/languages/_r/set_bridgestan_path.md.txt
@@ -1,0 +1,11 @@
+### Function `set_bridgestan_path()`
+
+```r
+set_bridgestan_path(path)
+```
+
+Set the path to BridgeStan.
+
+#### Details
+
+This should point to the top-level folder of the repository.
