+++ latest/_sources/languages/julia.md.txt
@@ -1,0 +1,564 @@
+
+<a id='Julia-Interface'></a>
+
+<a id='Julia-Interface-1'></a>
+
+# Julia Interface
+
+
+% NB: If you are reading this file in docs/languages, you are reading a generated output!
+% This should be apparent due to the html tags everywhere.
+% If you are reading this in julia/docs/src, you are reading the true source!
+% Please only make edits in the later file, since the first is DELETED each re-build.
+
+
+---
+
+
+<a id='Installation'></a>
+
+<a id='Installation-1'></a>
+
+## Installation
+
+
+<a id='From-JuliaRegistries'></a>
+
+<a id='From-JuliaRegistries-1'></a>
+
+### From JuliaRegistries
+
+
+BridgeStan is registered on JuliaRegistries each release.
+
+
+```julia
+] add BridgeStan
+```
+
+
+The first time you compile a model, the BridgeStan source code for your current version will be downloaded to a hidden directory in the users `HOME` directory. If you prefer to use a source distribution of BridgeStan, consult the following section.
+
+
+Note that the system pre-requisites from the [Getting Started guide](../getting-started.rst) are still required and will not be automatically installed by this method.
+
+
+<a id='From-Source'></a>
+
+<a id='From-Source-1'></a>
+
+### From Source
+
+
+This section assumes you have followed the [Getting Started guide](../getting-started.rst) to install BridgeStan's pre-requisites and downloaded a copy of the BridgeStan source code.
+
+
+To install the Julia interface, you can either install it directly from Github by running the following inside a Julia REPL
+
+
+```julia
+] add https://github.com/roualdes/bridgestan.git:julia
+```
+
+
+Or, since you have already downloaded the repository, you can run
+
+
+```julia
+] dev julia/
+```
+
+
+from the BridgeStan folder.
+
+
+To use the BridgeStan source you've manually downloaded instead of one the package will download for you, you must use [`set_bridgestan_path()`](BridgeStan.set_bridgestan_path!) or the `$BRIDGESTAN` environment variable.
+
+
+Note that the Julia package depends on Julia 1.10 (LTS) and the `Inflate` package.
+
+
+<a id='Example-Program'></a>
+
+<a id='Example-Program-1'></a>
+
+## Example Program
+
+
+An example program is provided alongside the Julia interface code in `example.jl`:
+
+
+<details>
+<summary><a>Show example.jl</a></summary>
+
+
+```{literalinclude} ../../julia/example.jl
+:language: julia
+```
+
+
+</details>
+
+
+<a id='API-Reference'></a>
+
+<a id='API-Reference-1'></a>
+
+## API Reference
+
+
+<a id='StanModel-interface'></a>
+
+<a id='StanModel-interface-1'></a>
+
+### StanModel interface
+
+<a id='BridgeStan.StanModel' href='#BridgeStan.StanModel'>#</a>
+**`BridgeStan.StanModel`** &mdash; *Type*.
+
+
+
+```julia
+StanModel(lib, data="", seed=204; stanc_args=[], make_args=[], warn=true)
+```
+
+A StanModel instance encapsulates a Stan model instantiated with data.
+
+Construct a Stan model from the supplied library file path and data. If lib is a path to a file ending in `.stan`, this will first compile the model.  Compilation occurs if no shared object file exists for the supplied Stan file or if a shared object file exists and the Stan file has changed since last compilation.  This is equivalent to calling [`compile_model`](julia.md#BridgeStan.compile_model) and then the constructor of `StanModel`. If `warn` is false, the warning about re-loading the same shared objects is suppressed.
+
[…]
