+++ v2.8.0/_sources/getting-started.rst.txt
@@ -1,0 +1,201 @@
+
+Getting Started
+===============
+
+Requirement: C++ toolchain
+--------------------------
+
+Stan requires a C++ tool chain consisting of
+
+* A C++14 compiler. On Windows, MSCV is *not* supported, so something like MinGW GCC is required.
+* The Gnu :command:`make` utility. On Windows, :command:`make` or :command:`mingw32-make` can be used
+
+Here are complete instructions by platform for installing both, from the CmdStan installation instructions.
+
+* `C++ tool chain installation - CmdStan User's Guide <https://mc-stan.org/docs/cmdstan-guide/installation.html#cpp-toolchain>`__
+
+Downloading BridgeStan
+----------------------
+
+.. note::
+    The :doc:`Julia <languages/julia>`, :doc:`Python <languages/python>`, and :doc:`R <languages/r>`
+    interfaces will download the source for you the first time you compile a model, and the :doc:`Rust <languages/rust>`
+    interface has an optional feature to download the source upon request.
+    This section is optional for users primarily interested in those interfaces.
+
+
+Installing BridgeStan is as simple as ensuring that the above requirements are installed and then downloading
+the source repository. All of the following ways of downloading BridgeStan will additionally download the
+`Stan <https://github.com/stan-dev/stan>`__ and `Stan Math <https://github.com/stan-dev/math>`__ libraries for you,
+and no additional dependencies are required to be installed separately for the C++ source code.
+
+
+Downloading a released archive
+______________________________
+
+Downloads of a complete copy of the source code and interfaces are available
+on `our GitHub releases page <https://github.com/roualdes/bridgestan/releases>`__.
+
+To use these, simply download the file associated with the version you wish to use,
+and unzip its contents into the folder you would like BridgeStan to be in.
+
+
+Installing the latest version with :command:`git`
+_________________________________________________
+
+If you have :command:`git` installed, you may download BridgeStan by navigating to the folder you'd like
+BridgeStan to be in and running
+
+.. code-block:: shell
+
+    git clone --recurse-submodules --shallow-submodules --depth=1 https://github.com/roualdes/bridgestan.git
+
+If you clone without the ``--recurse-submodules`` argument, you can download the required
+submodules with ``make stan-update``. The arguments ``--shallow-submodules`` and ``--depth=1`` are
+to reduce the size of the download, but are not required.
+
+
+Testing the Installation
+________________________
+
+After this, BridgeStan is installed. You can test a basic compilation by opening
+a terminal in your BridgeStan folder and running
+
+.. code-block:: shell
+
+    make test_models/multi/multi_model.so
+
+This will compile the file :file:`test_models/multi/multi.stan` into a shared library object for use with BridgeStan.
+This will require internet access the first time you run it in order
+to download the appropriate Stan compiler for your platform into
+:file:`{<bridgestan-dir>}/bin/stanc{[.exe]}`
+
+Installing an Interface
+-----------------------
+
+To see instructions for installing the BridgeStan client package in your language of
+choice, see the :doc:`Language Interfaces page <languages>`.
+
+Optional: Customizing BridgeStan
+--------------------------------
+
+BridgeStan has many compiler flags and options set by default. Many of these defaults
+are the same as those used by the CmdStan interface to Stan.
+You can override the defaults or add new flags
+on the command line when invoking :command:`make`, or make them persistent by
+creating or editing the file :file:`{<bridgestan dir>}/make/local`.
+
+For example, setting the contents of :file:`make/local` to the following
+includes compiler flags for optimization level and architecture.
+
+.. code-block:: Makefile
+
[…]
