+++ v2.9.0/_sources/languages.rst.txt
@@ -1,0 +1,27 @@
+
+Language Interfaces
+===================
+
+BridgeStan currently has clients in four languages, a public C API
+which underlies all the clients, and an example of a standalone program
+written in C.
+
+All language interfaces expose the same core functionality from the
+:doc:`C API <./languages/c-api>`.
+
+
+Additional "quality of life" features are available on a best-effort basis.
+If BridgeStan's bindings to your favorite language are missing one of these features, we welcome
+`contributions <https://github.com/roualdes/bridgestan/blob/main/CONTRIBUTING.md>`__!
+
+.. toctree::
+    :maxdepth: 2
+
+    languages/python
+    languages/julia
+    languages/r
+    languages/rust
+    languages/c-api
+
+
+
