+++ v2.9.0/_sources/internals.rst.txt
@@ -1,0 +1,18 @@
+
+How It Works
+============
+
+This section of the documentation is primarily intended for developers
+interested in contributing to BridgeStan or building their own client
+interface, but it may be interesting for users who are curious what is
+going on "behind the scenes".
+
+.. toctree::
+    :maxdepth: 2
+    :caption: Contents:
+
+    internals/ffi
+    internals/development
+    internals/testing
+    internals/documentation
+    internals/details

[…]
