+++ v1.0.2/_sources/index.rst.txt
@@ -1,0 +1,25 @@
+.. BridgeStan documentation master file, created by
+   sphinx-quickstart on Fri Sep  2 11:44:11 2022.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to BridgeStan's documentation!
+======================================
+
+BridgeStan provides efficient in-memory access through Python, Julia,
+and R to the methods of a `Stan <https://mc-stan.org>`__ model, including
+log densities, gradients, Hessians, and constraining and unconstraining
+transforms.
+
+The motivation is developing inference algorithms in
+higher-level languages for arbitrary Stan models, such as those in
+`posteriordb <https://github.com/stan-dev/posteriordb>`__.
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   getting-started
+   languages
+   internals
