### 2023-04-14 82b95bd
Adapt rust interface to changes in the C API
