### 2023-04-17 1a82893
Add version number to build docs
