+++ docs/conf.py
@@ -71,10 +71,10 @@ html_theme_options = {
     ],
     "use_edit_page_button": True,
     "switcher": {
-        "json_url": "https://wardbrian.github.io/bridgestan/latest/_static/switcher.json",
+        "json_url": "https://roualdes.github.io/bridgestan/latest/_static/switcher.json",
         "version_match": switcher_version
     },
-       "navbar_end": ["navbar-icon-links", "version-switcher"]
+       "navbar_end": ["theme-switcher", "navbar-icon-links", "version-switcher"]
 }
 
 html_context = {
