### 2023-04-18 bf91247
Utilize doc version switcher (#106)
* Add version number to build docs

* Push new docs on release

* Latex dependencies

* Build docs in 'latest' directory

* Try switcher

* Wire up release

* Add dry-run flag to release for testing

* Switch URLs

* Use inputs

* Remove extraneous switcher version

* Clean up

* Insert new versions at top of switcher

* Update existing links in repo
