### 2023-04-18 7273ee9
Manually flush cstdio in Julia test
