### 2023-04-18 c3d2c1a
Update StanModel docstring
