### 2023-04-19 d020f79
Remove rust variable name parsing
