### 2023-04-19 037827a
Remove nuts-rs specific code
