+++ rust/Cargo.toml
@@ -4,8 +4,8 @@ version = "1.0.2"
 edition = "2021"
 
 [dependencies]
-libloading = "0.7.4"
+libloading = "0.8.0"
 thiserror = "1.0.40"
 
 [build-dependencies]
-bindgen = "0.64.0"
+bindgen = "0.65.1"
