### 2023-04-19 9af4db5
Add tests for rust backend
