+++ rust/Cargo.toml
@@ -9,3 +9,6 @@ thiserror = "1.0.40"
 
 [build-dependencies]
 bindgen = "0.65.1"
+
+[dev-dependencies]
+approx = "0.5.1"
