### 2023-04-19 3fbfd0c
Improve rust documentation
