+++ rust/src/ffi.rs
@@ -1,5 +1,6 @@
 #![allow(non_camel_case_types)]
 #![allow(dead_code)]
 #![allow(clippy::all)]
+#![allow(rustdoc::broken_intra_doc_links)]
 // Include generated bindings file
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
