### 2023-04-19 5204ed8
Panic on non-utf8 parameter names and return info as cstr
