### 2023-04-19 9daf23d
Add workflow for rust tests
