### 2023-04-19 8aa757a
Cleanup rust error struct
