+++ .github/workflows/main.yaml
@@ -220,9 +220,21 @@ jobs:
           path: ./test_models/
           key: ${{ hashFiles('**/*.stan', 'src/*', 'stan/src/stan/version.hpp') }}-${{ matrix.os }}-v${{ env.CACHE_VERSION }}
 
-      - name: Run rust tests
+      - name: Install LLVM and Clang  # Required for bindgen
+        uses: KyleMayes/install-llvm-action@v1
+        with:
+          version: "15.0"
+          directory: ${{ runner.temp }}/llvm
+
+      - name: Get rust toolchain
         uses: dtolnay/rust-toolchain@stable
-        runs: cargo test --verbose --manifest-path=./rust/Cargo.toml
+
+      - name: Run rust tests
+        working-directory: ./rust
+        env:
+          LIBCLANG_PATH: ${{ runner.temp }}/llvm/lib
+          LLVM_CONFIG_PATH: ${{ runner.temp }}/llvm/bin/llvm-config
+        run: cargo test --verbose
 
   # Seperate for now, weird dynamic linking issues need resolving in GHA runner
   test_clients_windows:
