### 2023-04-20 3749748
ctypes.c_char_p is only for null-terminated strings
