### 2023-04-20 34b2f3b
Move windows rust tests to windows section
