### 2023-04-20 4806582
Test re-entrancy of print callback
