+++ python/test/test_stanmodel.py
@@ -678,10 +678,12 @@ def test_stdout_capture():
 
     # define a new, sillier, callback which lets us test thread safety
     x = 0
+
     @ctypes.CFUNCTYPE(None, ctypes.POINTER(ctypes.c_char), ctypes.c_int)
     def callback(s, n):
         nonlocal x
         x += 1
+
     m2._set_print_callback(callback, None)
 
     # call it many times from several threads
@@ -689,6 +691,7 @@ def test_stdout_capture():
         rng = m2.new_rng(1234)
         for _ in range(25):
             m2.param_constrain(np.array([theta]), include_gq=True, rng=rng)
+
     threads = [threading.Thread(target=f) for _ in range(10)]
     for t in threads:
         t.start()
