+++ src/model_rng.hpp
@@ -221,7 +221,6 @@ class bs_model {
 
   /** number of unconstrained parameters */
   int param_unc_num_ = -1;
-
 };
 
 /**
