### 2023-04-20 075994b
Add the ability to provide a callback function for printing (#108)
* Sketch: Printer callback

* Lock for callback, general cleanup

* Also add Julia capture

* Add tests

* Clean up python callback registration

* Manually flush cstdio in Julia test

* Include stddef in C header

* Clean up

* Allow an opt-out in Python

* Update StanModel docstring

* ctypes.c_char_p is only for null-terminated strings

* Test re-entrancy of print callback

* Formatting
