+++ python/test/test_stanmodel.py
@@ -645,6 +645,62 @@ def test_fr_gaussian():
         pos += 1
 
 
+def test_stdout_capture():
+    import contextlib, io
+
+    theta = 0.1
+
+    m = bs.StanModel(
+        str(STAN_FOLDER / "print" / "print_model.so"), capture_stan_prints=False
+    )
+
+    with contextlib.redirect_stdout(io.StringIO()) as f:
+        print("Hello from Python!")
+        m.log_density(np.array([theta]))
+
+    captured = f.getvalue()
+    assert captured.splitlines()[0] == "Hello from Python!"
+    assert "Stan" not in captured
+
+    m2 = bs.StanModel(str(STAN_FOLDER / "print" / "print_model.so"))
+
+    with contextlib.redirect_stdout(io.StringIO()) as f:
+        print("Hello from Python!")
+        m2.log_density(np.array([theta]))
+
+    lines = f.getvalue().splitlines()
+    assert lines[0] == "Hello from Python!"
+    assert lines[1] == "Hi from Stan!"
+    assert lines[2] == f"theta = {theta}"
+
+    # test re-entrancy
+    import ctypes, threading
+
+    # define a new, sillier, callback which lets us test thread safety
+    x = 0
+
+    @ctypes.CFUNCTYPE(None, ctypes.POINTER(ctypes.c_char), ctypes.c_int)
+    def callback(s, n):
+        nonlocal x
+        x += 1
+
+    m2._set_print_callback(callback, None)
+
+    # call it many times from several threads
+    def f():
+        rng = m2.new_rng(1234)
+        for _ in range(25):
+            m2.param_constrain(np.array([theta]), include_gq=True, rng=rng)
+
+    threads = [threading.Thread(target=f) for _ in range(10)]
+    for t in threads:
+        t.start()
+    for t in threads:
+        t.join()
+
+    assert x == 500  # 2 calls per print, 10 threads, 25 iterations
+
+
 @pytest.fixture
 def recompile_simple():
     """Recompile simple_model with autodiff hessian enable, then clean-up/restore it after test"""
