+++ src/bridgestan.cpp
@@ -2,12 +2,12 @@
 #include "model_rng.cpp"
 #include "bridgestanR.cpp"
 #include "version.hpp"
+#include "callback_stream.hpp"
+#include <sstream>
 
 int bs_major_version = BRIDGESTAN_MAJOR;
 int bs_minor_version = BRIDGESTAN_MINOR;
 int bs_patch_version = BRIDGESTAN_PATCH;
-
-#include <sstream>
 
 bs_model* bs_model_construct(const char* data, unsigned int seed,
                              char** error_msg) {
@@ -227,3 +227,28 @@ bs_rng* bs_rng_construct(unsigned int seed, char** error_msg) {
 }
 
 void bs_rng_destruct(bs_rng* rng) { delete (rng); }
+
+int bs_set_print_callback(STREAM_CALLBACK callback, char** error_msg) {
+  try {
+    if (buf != nullptr) {  // nullptr only when `outstream` is &std::cout
+      delete buf;
+      delete outstream;
+    }
+    if (callback == nullptr) {
+      outstream = &std::cout;
+      buf = nullptr;
+    } else {
+      buf = new callback_ostreambuf(callback);
+      outstream = new std::ostream(buf);
+    }
+    return 0;
+  } catch (...) {
+    if (error_msg) {
+      std::stringstream error;
+      error << "set_print_callback() failed with unknown exception"
+            << std::endl;
+      *error_msg = strdup(error.str().c_str());
+    }
+  }
+  return -1;
+}

[…]
