+++ README.md
@@ -12,7 +12,7 @@ Stan is a probabilistic programming language for coding statistical
 models.  For an introduction to what can be coded in Stan, see the
 [*Stan User's Guide*](https://mc-stan.org/docs/stan-users-guide/index.html).
 
-BridgeStan is currently shipping with Stan version 2.31.0
+BridgeStan is currently shipping with Stan version 2.32.0
 
 Documentation is available at https://roualdes.github.io/bridgestan/
 
