### 2023-04-24 d978db4
Add typedefs and variables to C API doc
