+++ .github/workflows/main.yaml
@@ -164,10 +164,10 @@ jobs:
           submodules: recursive
 
       - name: Install R
-        uses: r-lib/actions/setup-r@v2.4.0
+        uses: r-lib/actions/setup-r@v2.6.2
 
       - name: Install R dependencies
-        uses: r-lib/actions/setup-r-dependencies@v2.4.0
+        uses: r-lib/actions/setup-r-dependencies@v2.6.2
         with:
           packages: |
             any::R6
@@ -223,10 +223,10 @@ jobs:
           key: ${{ hashFiles('**/*.stan', 'src/*', 'stan/src/stan/version.hpp') }}-windows-latest-v${{ env.CACHE_VERSION }}
 
       - name: Install R
-        uses: r-lib/actions/setup-r@v2.4.0
+        uses: r-lib/actions/setup-r@v2.6.2
 
       - name: Install R dependencies
-        uses: r-lib/actions/setup-r-dependencies@v2.4.0
+        uses: r-lib/actions/setup-r-dependencies@v2.6.2
         with:
           packages: |
             any::R6
