### 2023-04-24 37ca40a
Add typedefs and variables to C API doc (#110)
* Add typedefs and variables to C API doc

* Bump setup-r for R 4.3
