### 2023-04-24 ae2a9da
Add workaround for library unload bug on windows
On windows we see deadlocks if a stan libraries are loaded
and unloaded in different threads. We work around this issue
by not unloading stan libraries.
