+++ rust/src/bs_safe.rs
@@ -5,6 +5,9 @@ use std::ffi::c_int;
 use std::ffi::c_uint;
 use std::ffi::CStr;
 use std::ffi::OsStr;
+#[cfg(windows)]
+use std::mem::forget;
+use std::mem::ManuallyDrop;
 use std::ptr::null;
 use std::ptr::NonNull;
 use std::str::Utf8Error;
@@ -13,7 +16,22 @@ use std::str::Utf8Error;
 use thiserror::Error;
 
 /// A loaded shared library for a stan model
-pub use ffi::Bridgestan as StanLibrary;
+pub struct StanLibrary(ManuallyDrop<ffi::Bridgestan>);
+
+#[cfg(windows)]
+impl Drop for StanLibrary {
+    fn drop(&mut self) {
+        let lib = unsafe { ManuallyDrop::take(&mut self.0) };
+        forget(lib.into_library());
+    }
+}
+
+#[cfg(not(windows))]
+impl Drop for StanLibrary {
+    fn drop(&mut self) {
+        unsafe { ManuallyDrop::drop(&mut self.0) };
+    }
+}
 
 /// Error type for bridgestan interface
 #[derive(Error, Debug)]
@@ -57,7 +75,10 @@ pub fn open_library<P: AsRef<OsStr>>(path: P) -> Result<StanLibrary> {
             format!("{}.{}.{}", self_major, self_minor, self_patch),
         ));
     }
-    Ok(unsafe { StanLibrary::from_library(library) }?)
+
+    let lib = unsafe { ffi::Bridgestan::from_library(library) }?;
+    let lib = ManuallyDrop::new(lib);
+    Ok(StanLibrary(lib))
 }
 
 /// A Stan model instance with data
@@ -83,7 +104,7 @@ impl<T: Borrow<StanLibrary>> Drop for Rng<T> {
     fn drop(&mut self) {
         unsafe {
             // We don't handle error messages during deconstruct
-            self.lib.borrow().bs_rng_destruct(self.rng.as_ptr());
+            self.lib.borrow().0.bs_rng_destruct(self.rng.as_ptr());
         }
     }
 }
@@ -91,7 +112,11 @@ impl<T: Borrow<StanLibrary>> Drop for Rng<T> {
 impl<T: Borrow<StanLibrary>> Rng<T> {
     pub fn new(lib: T, seed: u32) -> Result<Self> {
         let mut err = ErrorMsg::new(lib.borrow());
-        let rng = unsafe { lib.borrow().bs_rng_construct(seed as c_uint, err.as_ptr()) };
+        let rng = unsafe {
+            lib.borrow()
+                .0
+                .bs_rng_construct(seed as c_uint, err.as_ptr())
+        };
         if let Some(rng) = NonNull::new(rng) {
             drop(err);
             Ok(Self { rng, lib })
@@ -109,7 +134,7 @@ struct ErrorMsg<'lib> {
 impl<'lib> Drop for ErrorMsg<'lib> {
     fn drop(&mut self) {
         if !self.msg.is_null() {
-            unsafe { self.lib.bs_free_error_msg(self.msg) };
+            unsafe { self.lib.0.bs_free_error_msg(self.msg) };
         }
     }
 }
@@ -153,6 +178,7 @@ impl<T: Borrow<StanLibrary>> Model<T> {
             .unwrap_or(null());
         let model = unsafe {
             lib.borrow()
+                .0
                 .bs_model_construct(data_ptr, seed, err.as_ptr())
         };
         // Make sure data lives until here
@@ -187,13 +213,13 @@ impl<T: Borrow<StanLibrary>> Model<T> {
 
     /// Return the name of the model or error if UTF decode fails
[…]
