+++ rust/src/ffi.rs
@@ -4,3 +4,9 @@
 #![allow(rustdoc::broken_intra_doc_links)]
 // Include generated bindings file
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
+
+impl Bridgestan {
+    pub(crate) fn into_library(self) -> libloading::Library {
+        self.__library
+    }
+}
