+++ rust/tests/create_models.rs
@@ -2,6 +2,8 @@ use std::{
     f64::consts::PI,
     ffi::CString,
     path::{Path, PathBuf},
+    sync::mpsc::{sync_channel, Receiver, SyncSender},
+    thread::spawn,
 };
 
 use approx::assert_ulps_eq;
@@ -36,7 +38,7 @@ fn get_model<S: AsRef<str>>(name: S) -> (StanLibrary, Option<CString>) {
 }
 
 #[test]
-fn create_all() {
+fn create_all_serial() {
     let base = model_dir();
     for path in base.read_dir().unwrap() {
         let path = path.unwrap().path();
@@ -55,6 +57,263 @@ fn create_all() {
         };
         assert!(model.name().unwrap().contains(name));
     }
+}
+
+#[test]
+fn create_all_late_drop_fwd() {
+    let base = model_dir();
+    let names: Vec<String> = base
+        .read_dir()
+        .unwrap()
+        .map(|path| {
+            let path = path.unwrap().path();
+            path.file_name().unwrap().to_str().unwrap().to_string()
+        })
+        .collect();
+
+    let handles: Vec<_> = names
+        .into_iter()
+        .filter(|name| (name != "logistic") & (name != "regression") & (name != "syntax_error"))
+        .map(|name| {
+            let (lib, data) = get_model(&name);
+            let Ok(model) = Model::new(&lib, data.as_ref(), 42) else {
+                // Only those two models should fail to create.
+                assert!((name == "ode") | (name == "throw_data"));
+                return lib
+            };
+            assert!(model.name().unwrap().contains(&name));
+            drop(model);
+            lib
+        })
+        .collect();
+    handles.into_iter().for_each(|lib| drop(lib))
+}
+
+#[test]
+fn create_all_thread_serial() {
+    let base = model_dir();
+    let names: Vec<String> = base
+        .read_dir()
+        .unwrap()
+        .map(|path| {
+            let path = path.unwrap().path();
+            path.file_name().unwrap().to_str().unwrap().to_string()
+        })
+        .collect();
+
+    names.into_iter().for_each(|name| {
+        spawn(move || {
+            if (&name == "logistic") | (&name == "regression") | (&name == "syntax_error") {
+                return;
+            }
+
+            let (lib, data) = get_model(&name);
+            // Create the model with a reference
+            let Ok(model) = Model::new(&lib, data.as_ref(), 42) else {
+                    // Only those two models should fail to create.
+                    assert!((name == "ode") | (name == "throw_data"));
+                    return;
+                };
+            assert!(model.name().unwrap().contains(&name));
+        })
+        .join()
+        .unwrap()
+    })
+    //handles
+    //.into_iter()
+    //.for_each(|handle| handle.join().unwrap())
+}
+
+#[test]
+fn create_all_parallel() {
+    let base = model_dir();
+    let names: Vec<String> = base
+        .read_dir()
+        .unwrap()
+        .map(|path| {
+            let path = path.unwrap().path();
+            path.file_name().unwrap().to_str().unwrap().to_string()
+        })
+        .collect();
+
+    let handles: Vec<_> = names
+        .into_iter()
+        .map(|name| {
+            spawn(move || {
+                if (&name == "logistic") | (&name == "regression") | (&name == "syntax_error") {
+                    return;
+                }
+
+                let (lib, data) = get_model(&name);
+                // Create the model with a reference
[…]
