### 2023-04-24 ab1cdab
Implement missing functions in rust interface
