### 2023-04-25 8c8bec3
Run clippy in CI and add example
