### 2023-04-25 df129f2
Some docstring improvements
