### 2023-04-25 87e110b
Make rng optional
