### 2023-04-25 80610d6
Add cargo fmt check to CI
