### 2023-04-25 717519c
Adapt to bools in bridgestan.h
