+++ rust/src/ffi.rs
@@ -1,4 +1,5 @@
 #![allow(non_camel_case_types)]
+#![allow(non_upper_case_globals)]
 #![allow(dead_code)]
 #![allow(clippy::all)]
 #![allow(rustdoc::broken_intra_doc_links)]
