+++ rust/Cargo.toml
@@ -2,6 +2,11 @@
 name = "bridgestan"
 version = "1.0.2"
 edition = "2021"
+rust-version = "1.69"
+description = "Rust interface for BridgeStan"
+repository = "https://github.com/roualdes/bridgestan"
+license = "BSD-3-clause"
+homepage = "https://roualdes.github.io/bridgestan/latest/"
 
 [dependencies]
 libloading = "0.8.0"
