### 2023-04-25 4b2250d
Check identity of library for model and rng
