+++ rust/src/bs_safe.rs
@@ -1,22 +1,29 @@
 use crate::ffi;
 use std::borrow::Borrow;
+use std::collections::hash_map::DefaultHasher;
 use std::ffi::c_char;
 use std::ffi::c_int;
 use std::ffi::c_uint;
 use std::ffi::CStr;
 use std::ffi::OsStr;
+use std::hash::Hash;
+use std::hash::Hasher;
 #[cfg(windows)]
 use std::mem::forget;
 use std::mem::ManuallyDrop;
 use std::ptr::NonNull;
 use std::ptr::{null, null_mut};
 use std::str::Utf8Error;
+use std::time::Instant;
 
 // This is more or less equivalent to manually defining Display and From<other error types>
 use thiserror::Error;
 
 /// A loaded shared library for a stan model
-pub struct StanLibrary(ManuallyDrop<ffi::BridgeStan>);
+pub struct StanLibrary {
+    lib: ManuallyDrop<ffi::BridgeStan>,
+    id: u64,
+}
 
 // To work around a bug on windows where unloading a library
 // can lead to deadlocks.
@@ -33,7 +40,7 @@ impl Drop for StanLibrary {
 #[cfg(not(windows))]
 impl Drop for StanLibrary {
     fn drop(&mut self) {
-        unsafe { ManuallyDrop::drop(&mut self.0) };
+        unsafe { ManuallyDrop::drop(&mut self.lib) };
     }
 }
 
@@ -51,7 +58,7 @@ impl StanLibrary {
     /// need to be thread safe.
     pub unsafe fn set_print_callback(&mut self, callback: StanPrintCallback) -> Result<()> {
         let mut err = ErrorMsg::new(self);
-        let rc = unsafe { self.0.bs_set_print_callback(Some(callback), err.as_ptr()) };
+        let rc = unsafe { self.lib.bs_set_print_callback(Some(callback), err.as_ptr()) };
 
         if rc == 0 {
             Ok(())
@@ -90,7 +97,7 @@ type Result<T> = std::result::Result<T, BridgeStanError>;
 /// The library should have been compiled with bridgestan,
 /// with the same version as the rust library.
 pub fn open_library<P: AsRef<OsStr>>(path: P) -> Result<StanLibrary> {
-    let library = unsafe { libloading::Library::new(path) }.map_err(LoadingError)?;
+    let library = unsafe { libloading::Library::new(&path) }.map_err(LoadingError)?;
     let major: libloading::Symbol<*const c_int> =
         unsafe { library.get(b"bs_major_version") }.map_err(LoadingError)?;
     let major = unsafe { **major };
@@ -113,7 +120,11 @@ pub fn open_library<P: AsRef<OsStr>>(path: P) -> Result<StanLibrary> {
 
     let lib = unsafe { ffi::BridgeStan::from_library(library) }.map_err(LoadingError)?;
     let lib = ManuallyDrop::new(lib);
-    Ok(StanLibrary(lib))
+    let mut hasher = DefaultHasher::new();
+    Instant::now().hash(&mut hasher);
+    path.as_ref().hash(&mut hasher);
+    let id = hasher.finish();
+    Ok(StanLibrary { lib, id })
 }
 
 /// A Stan model instance with data
@@ -138,7 +149,7 @@ unsafe impl<T: Send + Borrow<StanLibrary>> Send for Rng<T> {}
 impl<T: Borrow<StanLibrary>> Drop for Rng<T> {
     fn drop(&mut self) {
         unsafe {
-            self.lib.borrow().0.bs_rng_destruct(self.rng.as_ptr());
+            self.lib.borrow().lib.bs_rng_destruct(self.rng.as_ptr());
         }
     }
 }
[…]
