+++ .github/workflows/main.yaml
@@ -8,7 +8,7 @@ on:
   workflow_dispatch: {}
 
 env:
-  CACHE_VERSION: 3
+  CACHE_VERSION: 4
 
 jobs:
   build_test_models:
