### 2023-04-27 9e47b87
Add test for windows bug
