+++ rust/tests/create_models.rs
@@ -232,7 +232,9 @@ fn load_order_all_serial() {
     drop(lib5);
 }
 
+// Test for bug https://github.com/roualdes/bridgestan/issues/111
 #[test]
+#[ignore]
 fn load_order_min_parallel() {
     let names = ["bernoulli", "gaussian", "jacobian"];
     let (senders, handles): (Vec<_>, Vec<_>) = names
@@ -248,7 +250,7 @@ fn load_order_min_parallel() {
                 ok_sender.send(()).unwrap();
 
                 unload_receiver.recv().unwrap();
-                drop(lib);
+                unsafe { lib.unload_library() };
                 ok_sender.send(()).unwrap();
 
                 exit_receiver.recv().unwrap();
@@ -277,12 +279,12 @@ fn load_order_min_parallel() {
     load(&senders[0]);
     load(&senders[1]);
     unload(&senders[0]);
+    exit(&senders[0]);
+    // Potential deadlock
     load(&senders[2]);
     unload(&senders[1]);
-    unload(&senders[2]);
-
-    exit(&senders[0]);
     exit(&senders[1]);
+    unload(&senders[2]);
     exit(&senders[2]);
 
     handles.into_iter().for_each(|h| h.join().unwrap());
