+++ .github/workflows/release.yaml
@@ -97,7 +97,7 @@ jobs:
 
       - name: Upload PyPI wheels
         if: ${{ !inputs.dry_run }}
-        uses: pypa/gh-action-pypi-publish@v1.8.4
+        uses: pypa/gh-action-pypi-publish@v1.8.5
         with:
           password: ${{ secrets.PYPI_TOKEN }}
           packages_dir: python/dist/
