### 2023-05-03 74c5376
Update to Stan 2.32.1 (#115)
* Bump stan from `7760933` to `42e873d`

Bumps [stan](https://github.com/stan-dev/stan) from `7760933` to `42e873d`.
- [Release notes](https://github.com/stan-dev/stan/releases)
- [Commits](https://github.com/stan-dev/stan/compare/77609338a253678f747d86ab676f71f3e6fd4df1...42e873d5879143545910310ea9f10463d18f86f7)

---
updated-dependencies:
- dependency-name: stan
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>

* Bump versions

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bo
[…]
