### 2023-05-03 2981237
Tweaks following 2.0 release action (#116)
