### 2023-05-05 af2b73d
Also use os.add_dll_directory
