+++ .github/workflows/main.yaml
@@ -233,6 +233,23 @@ jobs:
       - name: Set up Julia
         uses: julia-actions/setup-julia@v1
 
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.10"
+
+      - name: Install Python package
+        run: |
+          cd python/
+          pip install pytest
+          pip install .
+
+      - name: Run Python tests
+        run: |
+          $env:BRIDGESTAN = $(pwd)
+          cd python/
+          pytest -v
+          
       - name: Run Julia tests
         run: |
           $env:BRIDGESTAN = $(pwd)
