+++ .github/workflows/main.yaml
@@ -105,10 +105,11 @@ jobs:
 
       - name: Run tests
         run: |
-          export BRIDGESTAN=$(pwd)
           cd python/
           pytest -v
           pytest -v --run-type=ad_hessian
+        env:
+          BRIDGESTAN: ${{ github.workspace }}
 
   julia:
     needs: [build]
@@ -157,7 +158,7 @@ jobs:
     runs-on: ${{matrix.os}}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
       fail-fast: false
     steps:
       - name: Check out github
@@ -188,6 +189,12 @@ jobs:
         with:
           path: ./test_models/
           key: ${{ hashFiles('**/*.stan', 'src/*', 'stan/src/stan/version.hpp', 'Makefile') }}-${{ matrix.os }}-v${{ env.CACHE_VERSION }}
+
+      # needed for R tests until they have compilation utilities and can set this themselves.
+      - name: Set up TBB
+        if: matrix.os == 'windows-latest'
+        run: |
+          Add-Content $env:GITHUB_PATH "$(pwd)/stan/lib/stan_math/lib/tbb"
 
       - name: Run tests
         run: |
@@ -221,17 +228,6 @@ jobs:
           path: ./test_models/
           key: ${{ hashFiles('**/*.stan', 'src/*', 'stan/src/stan/version.hpp', 'Makefile') }}-windows-latest-v${{ env.CACHE_VERSION }}
 
-      - name: Install R
-        uses: r-lib/actions/setup-r@v2.6.3
-
-      - name: Install R dependencies
-        uses: r-lib/actions/setup-r-dependencies@v2.6.3
-        with:
-          packages: |
-            any::R6
-            any::testthat
-            any::devtools
-
       - name: Set up Julia
         uses: julia-actions/setup-julia@v1
 
@@ -252,14 +248,3 @@ jobs:
           cd python/
           pytest -v
 
-      # needed for R tests until they have compilation utilities and can set this themselves.
-      - name: Set up TBB
-        run: |
-          Add-Content $env:GITHUB_PATH "$(pwd)/stan/lib/stan_math/lib/tbb"
-
-      - name: Run R tests
-        run: |
-          cd R/tests/testthat
-          gcc -fpic -shared -o test_collisions.dll test_collisions.c
-          cd ../..
-          Rscript -e 'devtools::test(reporter = c(\"summary\", \"fail\"))'
