+++ .github/workflows/main.yaml
@@ -69,7 +69,7 @@ jobs:
     runs-on: ${{matrix.os}}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: [3.9, "3.10"]
       fail-fast: false
     steps:
@@ -133,6 +133,7 @@ jobs:
         uses: julia-actions/setup-julia@v1
         with:
           version: ${{ matrix.julia-version }}
+
       - name: Restore Stan
         uses: actions/cache@v3
         id: stan-cache
@@ -197,54 +198,17 @@ jobs:
           Add-Content $env:GITHUB_PATH "$(pwd)/stan/lib/stan_math/lib/tbb"
 
       - name: Run tests
+        if: matrix.os != 'windows-latest'
         run: |
           cd R/tests/testthat
           gcc -fpic -shared -o test_collisions.so test_collisions.c
           cd ../..
           Rscript -e "devtools::test(reporter = c(\"summary\", \"fail\"))"
 
-  # Seperate for now, weird dynamic linking issues need resolving in GHA runner
-  test_clients_windows:
-    needs: [build]
-    runs-on: windows-latest
-    steps:
-      - name: Check out github
-        uses: actions/checkout@v3
-        with:
-          submodules: recursive
-
-      - name: Restore Stan
-        uses: actions/cache@v3
-        id: stan-cache
-        with:
-          path: ./stan/
-          key: ${{ runner.os }}-stan-${{ hashFiles('stan/src/stan/version.hpp') }}-v${{ env.CACHE_VERSION }}
-
-
-      - name: Restore built models
-        uses: actions/cache@v3
-        id: test-models
-        with:
-          path: ./test_models/
-          key: ${{ hashFiles('**/*.stan', 'src/*', 'stan/src/stan/version.hpp', 'Makefile') }}-windows-latest-v${{ env.CACHE_VERSION }}
-
-      - name: Set up Julia
-        uses: julia-actions/setup-julia@v1
-
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.10"
-
-      - name: Install Python package
+      - name: Run tests (windows)
+        if: matrix.os == 'windows-latest'
         run: |
-          cd python/
-          pip install pytest
-          pip install .
-
-      - name: Run Python tests
-        run: |
-          $env:BRIDGESTAN = $(pwd)
-          cd python/
-          pytest -v
-
+          cd R/tests/testthat
+          gcc -fpic -shared -o test_collisions.dll test_collisions.c
+          cd ../..
+          Rscript -e 'devtools::test(reporter = c(\"summary\", \"fail\"))'
