### 2023-05-06 3552625
Add Makefile to hash in github actions
