### 2023-05-08 6e5a83e
Look for more options for mingw dlls
