+++ python/bridgestan/__init__.py
@@ -29,7 +29,7 @@ if _plt.system() == "Windows":
 
         try:
             out = subprocess.run(
-                ["where.exe", "libwinpthread-1.dll"], check=True, capture_output=True
+                ["where.exe", "libwinpthread-1.dll", "libgcc_s_seh-1.dll", "libstdc++-6.dll"], check=True, capture_output=True
             )
             mingw_dir = os.path.dirname(out.stdout.decode().splitlines()[0])
             os.add_dll_directory(mingw_dir)
