### 2023-05-08 e70d303
Windows: add TBB (and MinGW DLLs) to load path (#118)
* Windows: Add TBB to PATH on client load

* Also use os.add_dll_directory

* Look for more options for mingw dlls
