+++ .github/workflows/main.yaml
@@ -211,9 +211,6 @@ jobs:
           path: ./stan/
           key: ${{ runner.os }}-stan-${{ hashFiles('stan/src/stan/version.hpp') }}-v${{ env.CACHE_VERSION }}
 
-      - name: Setup TBB
-        run: |
-          Add-Content $env:GITHUB_PATH "$(pwd)/stan/lib/stan_math/lib/tbb"
 
       - name: Restore built models
         uses: actions/cache@v3
@@ -236,10 +233,32 @@ jobs:
       - name: Set up Julia
         uses: julia-actions/setup-julia@v1
 
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.10"
+
+      - name: Install Python package
+        run: |
+          cd python/
+          pip install pytest
+          pip install .
+
+      - name: Run Python tests
+        run: |
+          $env:BRIDGESTAN = $(pwd)
+          cd python/
+          pytest -v
+          
       - name: Run Julia tests
         run: |
           $env:BRIDGESTAN = $(pwd)
           julia --project=./julia -t 2 -e "using Pkg; Pkg.test()"
+
+      # needed for R tests until they have compilation utilities and can set this themselves.
+      - name: Set up TBB
+        run: |
+          Add-Content $env:GITHUB_PATH "$(pwd)/stan/lib/stan_math/lib/tbb"
 
       - name: Run R tests
         run: |
