### 2023-05-08 add29bf
Properly skip non-Windows test
