### 2023-05-08 32bd072
Split Windows CI (#119)
* Try to split off Julia tests

* Try to split out R

* Try to split out Python

* Debug

* Revert debug

* Properly skip non-Windows test
