+++ .github/workflows/main.yaml
@@ -11,7 +11,7 @@ env:
   CACHE_VERSION: 0
 
 jobs:
-  build_test_models:
+  build:
     runs-on: ${{matrix.os}}
     strategy:
       matrix:
@@ -64,12 +64,12 @@ jobs:
           mingw32-make.exe STAN_THREADS=true O=0 test_models -j2
         shell: pwsh
 
-  test_python_client:
-    needs: [build_test_models]
+  python:
+    needs: [build]
     runs-on: ${{matrix.os}}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: [3.9, "3.10"]
       fail-fast: false
     steps:
@@ -105,17 +105,24 @@ jobs:
 
       - name: Run tests
         run: |
-          export BRIDGESTAN=$(pwd)
-          cd python/
-          pytest -v
-          pytest -v --run-type=ad_hessian
+          pytest -v python/
+        env:
+          BRIDGESTAN: ${{ github.workspace }}
 
-  test_julia_client:
-    needs: [build_test_models]
+      - name: Run tests (Unix-specific)
+        if: matrix.os != 'windows-latest'
+        run: |
+          pytest -v python/ --run-type=ad_hessian
+        env:
+          BRIDGESTAN: ${{ github.workspace }}
+
+
+  julia:
+    needs: [build]
     runs-on: ${{matrix.os}}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
         julia-version: ["1"]
         include:
           - julia-version: "1.6"
@@ -131,6 +138,7 @@ jobs:
         uses: julia-actions/setup-julia@v1
         with:
           version: ${{ matrix.julia-version }}
+
       - name: Restore Stan
         uses: actions/cache@v3
         id: stan-cache
@@ -147,15 +155,16 @@ jobs:
 
       - name: Run tests
         run: |
-          export BRIDGESTAN=$(pwd)
           julia --project=./julia -t 2 -e "using Pkg; Pkg.test()"
+        env:
+          BRIDGESTAN: ${{ github.workspace }}
 
-  test_R_client:
-    needs: [build_test_models]
+  Rlang:
+    needs: [build]
     runs-on: ${{matrix.os}}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
       fail-fast: false
     steps:
       - name: Check out github
@@ -187,80 +196,22 @@ jobs:
           path: ./test_models/
           key: ${{ hashFiles('**/*.stan', 'src/*', 'stan/src/stan/version.hpp', 'Makefile') }}-${{ matrix.os }}-v${{ env.CACHE_VERSION }}
 
+      # needed for R tests until they have compilation utilities and can set this themselves.
+      - name: Set up TBB
+        if: matrix.os == 'windows-latest'
+        run: |
+          Add-Content $env:GITHUB_PATH "$(pwd)/stan/lib/stan_math/lib/tbb"
+
       - name: Run tests
+        if: matrix.os != 'windows-latest'
         run: |
           cd R/tests/testthat
           gcc -fpic -shared -o test_collisions.so test_collisions.c
           cd ../..
           Rscript -e "devtools::test(reporter = c(\"summary\", \"fail\"))"
 
-  # Seperate for now, weird dynamic linking issues need resolving in GHA runner
-  test_clients_windows:
-    needs: [build_test_models]
-    runs-on: windows-latest
-    steps:
-      - name: Check out github
-        uses: actions/checkout@v3
-        with:
-          submodules: recursive
-
[…]
