### 2023-05-11 #5236 @salleuska
@Nikoleta-v3 and @roualdes  I am sorry for the late reply. I had to travel in the past weeks and have unexpectedly limited access to the internet. I still have to check out some things about the software functionalities/docs, but I have some comments about the paper and a couple of issues.

**Paper**

I think that the writing of the paper could be improved and that there some things that are worth clarifying. I think that the package provides a great tool for Stan users to take advantage of internal Stan methods, allowing users to develop their own algorithms. However, it seems to me that some parts of the paper are more focused on marketing the product rather than presenting the methods the bridgestan package allows access.

**Major Comments**

1. I would suggest revising the text in "Statement of need" and clarify which methods of Stan models are usually not accessible via the standard Stan interfaces that are instead made available by bridgestan. I also think it would be useful to mention what are potential user cases or if there are already relevant developed algorithms that take advantage of the package.

2. From the text it is not clear to me what is the state of the art, for example:

- line 43. "Using a memory-compatible C interface makes this possible even if the host language (e.g., R) was compiled with a different compiler, something no prior interface which exposed Stan’s log density calculations could allow."
-> Is there previous work that allows to access to Stan's models methods? If so could you cite that? Or give an idea of the shortcomings of these interfaces?

- line 44. "Existing tools with similar functionalities include JAX (Bradbury et al., 2018) and Turing.jl via the JuliaAD ecosystem " -> What do you mean with similar functionalities? From the sentence it is unclear whether these tools offer access to methods in Stan or instead provide functions for derivatives and calculation of log-densities?

**Minor**

- line 27: "we find that rstan ranks at number 70, rstantools 179, and rstanarm 502." it is not obvious to a reader what these packages are,

- line 39. "BridgeStan offers efficient, in-memory, from memory allocated within the host language, computations of the log joint density function of a Stan model, itself implemented using highly templated C++ from the Stan math library" -> this sentence is hard to follow; I think it would helpful to elaborate a bit, and possible merge these concepts related to memory in the paragraph at line 53.

- line 47: "The Stan community by and large uses CPU hardware" - this is unclear; maybe "by" is a "buy"?

- There is an abuse of the term  "efficient" that I find distracting and I would encourage the authors to remove it when unnecessary.
