### 2023-05-11 roualdes/bridgestan#121 @salleuska
Error in R/example.R
When I run the R example I got this error

`Error in initialize(...) : unused argument (0) `

From the function help I think the `0` is an extra argument here

https://github.com/roualdes/bridgestan/blob/32bd072236a0b59e52dfeb69c0a0b7342d6d7b42/R/example.R#L7
