### 2023-05-11 30efe61
Run all examples in CI
