### 2023-05-11 #5236 @salleuska
**Code**

For now, I found one error in the R example.

https://github.com/roualdes/bridgestan/issues/121#issue-1705891136

I will check julia and python early next week.
