### 2023-05-11 6026f40
Fix R example and run examples in CI (#122)
* Remove extra argument from example.R

* Run all examples in CI

* Update github action

* Pin MacOS Julia version to 1.8
