+++ .github/workflows/main.yaml
@@ -124,14 +124,11 @@ jobs:
     runs-on: ${{matrix.os}}
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
         julia-version: ["1"]
         include:
           - julia-version: "1.6"
             os: ubuntu-latest
-          # bug in Julia 1.9: https://github.com/JuliaLang/julia/issues/49772
-          - julia-version: "1.8"
-            os: macos-latest
       fail-fast: false
     steps:
       - name: Check out github
