### 2023-05-19 d8473a9
clear up similar functionality with JAX
