### 2023-05-19 e2cc9ba
address previous work that allows access to Stan model methods
