### 2023-05-19 5d63b03
clarify primary use case of BridgeStan
