### 2023-05-19 2932c77
state more clearly why we mention rstan, rstanarm, rstantools, pystan, and cmdstanpy
