### 2023-05-19 9c687e4
clean up sentence about what BridgeStan offers
