### 2023-05-19 06c5513
remove all but two uses of the word efficient
