+++ paper/paper.bib
@@ -1,10 +1,10 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% http://bibdesk.sourceforge.net/
 
-%% Created for Edward Roualdes at 2023-01-10 10:03:35 -0800 
+%% Created for Edward Roualdes at 2023-01-10 10:03:35 -0800
 
 
-%% Saved with string encoding Unicode (UTF-8) 
+%% Saved with string encoding Unicode (UTF-8)
 
 
 
@@ -105,3 +105,19 @@
 	title = {Comments on {"Representations of knowledge in complex systems" by U. Grenander and MI Miller}},
 	volume = {56},
 	year = {1994}}
+
+@misc{httpstan,
+  title = {httpstan (4.4.0)},
+  author = {Riddell, Allen and Hartikainen, Ari and Carter, Matthew},
+  year = {2021},
+  month = mar,
+  howpublished = {PyPI}
+}
+
+@Misc{rstan,
+	title = {{RStan}: the {R} interface to {Stan}},
+	author = {{Stan Development Team}},
+	note = {R package version 2.21.8},
+	year = {2023},
+	url = {https://mc-stan.org/},
+}
