+++ rust/src/bs_safe.rs
@@ -64,7 +64,7 @@ impl StanLibrary {
     ///
     /// There seem to be issues around unloading libraries in threaded
     /// code that are not fully understood:
-    /// https://github.com/roualdes/bridgestan/issues/111
+    /// <https://github.com/roualdes/bridgestan/issues/111>
     pub unsafe fn unload_library(mut self) {
         let lib = unsafe { ManuallyDrop::take(&mut self.lib) };
         drop(lib.into_library());
