+++ rust/tests/loading.rs
@@ -1,41 +1,12 @@
+mod common;
+use common::{get_model, model_dir};
+
 use std::{
-    f64::consts::PI,
-    ffi::CString,
-    path::{Path, PathBuf},
     sync::mpsc::{sync_channel, Receiver, SyncSender},
     thread::spawn,
 };
 
-use approx::{assert_abs_diff_eq, assert_ulps_eq};
-use bridgestan::{open_library, BridgeStanError, Model, StanLibrary};
-
-fn model_dir() -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .unwrap()
-        .join("test_models")
-}
-
-/// Load stan library and corresponding data if available
-fn get_model<S: AsRef<str>>(name: S) -> (StanLibrary, Option<CString>) {
-    let name = name.as_ref();
-    let mut base = model_dir();
-    base.push(name);
-    let lib = base.join(format!("{}_model.so", name));
-    if !lib.exists() {
-        panic!("Could not find compiled model {}", name);
-    }
-    let lib = open_library(lib).unwrap();
-
-    let data_path = base.join(name).with_extension("data.json");
-
-    if data_path.exists() {
-        let contents = std::fs::read(data_path).unwrap();
-        (lib, Some(CString::new(contents).unwrap()))
-    } else {
-        (lib, None)
-    }
-}
+use bridgestan::Model;
 
 #[test]
 fn create_all_serial() {
@@ -231,7 +202,6 @@ fn load_order_all_serial() {
 
 // Test for bug https://github.com/roualdes/bridgestan/issues/111
 #[test]
-#[ignore]
 fn load_order_min_parallel() {
     let names = ["bernoulli", "gaussian", "jacobian"];
     let (senders, handles): (Vec<_>, Vec<_>) = names
@@ -313,111 +283,4 @@ fn load_parallel() {
     handles
         .into_iter()
         .for_each(|handle| handle.join().unwrap())
-}
-
-#[test]
-fn throw_data() {
-    let (lib, data) = get_model("throw_data");
-    let Err(err) = Model::new(&lib, data, 42) else {
-        panic!("throw_data model should not successfully be created.");
-    };
-
-    let BridgeStanError::ConstructFailed(msg) = err else {
-        panic!("Creating throw_data model return an unexpected error");
-    };
-    assert!(msg.contains("find this text: datafails"));
-}
-
-#[test]
-#[should_panic(expected = "number of parameters")]
-fn bad_arglength() {
-    let (lib, data) = get_model("stdnormal");
-    let model = Model::new(&lib, data, 42).unwrap();
-    let theta = vec![];
-    let mut grad = vec![];
-    let _ = model.log_density_gradient(&theta[..], true, true, &mut grad[..]);
-}
-
-#[test]
-fn logp_gradient() {
-    let (lib, data) = get_model("stdnormal");
-    let model = Model::new(&lib, data, 42).unwrap();
-    let theta = vec![1f64];
-    let mut grad = vec![0f64];
-    let logp = model
-        .log_density_gradient(&theta[..], false, true, &mut grad[..])
-        .unwrap();
-    assert_ulps_eq!(logp, (2. * PI).sqrt().recip().ln() - 0.5);
-    assert_ulps_eq!(grad[0], -1f64);
-}
-
-#[test]
-fn logp_hessian() {
-    let (lib, data) = get_model("stdnormal");
-    let model = Model::new(&lib, data, 42).unwrap();
-    let theta = vec![1f64];
-    let mut grad = vec![0f64];
[…]
