+++ rust/tests/common.rs
@@ -1,4 +1,3 @@
-
 use std::{
     ffi::CString,
     path::{Path, PathBuf},
