### 2023-05-23 9973a26
Add Rust wrapper (#105)
* Add minimal Rust example

* Update rust example

* Fix memory leak

* Reorganize

* Error handling tweak

* Use slices

* Threading

* Reorg and add NUTS example using nuts-rs

* Clean up

* Switch names

* Lint with clippy

* Version bump

* Update dependencies

* Update readme

* Also expose unc

* Switch to libloading style

* Update readme

* Bump version

* Rebase rust interface to use rng and error message

* Move rust bindings

* Fix clippy suggestions

* Remove Cargo.lock from repo

* Use Borrow to store reference to stan library

* Change Rng to use Borrow

*
[…]
