+++ rust/src/bridgestan.h
@@ -1,0 +1,1 @@
+../../src/bridgestan.h
