+++ rust/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "bridgestan"
+version = "2.0.0"
+edition = "2021"
+rust-version = "1.69"
+description = "Rust interface for BridgeStan"
+repository = "https://github.com/roualdes/bridgestan"
+license = "BSD-3-clause"
+homepage = "https://roualdes.github.io/bridgestan/latest/"
+
+[dependencies]
+libloading = "0.8.0"
+thiserror = "1.0.40"
+
+[build-dependencies]
+bindgen = "0.65.1"
+
+[dev-dependencies]
+approx = "0.5.1"
