### 2023-05-23 9dffa41
fix a misspelling and clean up the phrase standardize algorithm development
