+++ paper/paper.md
@@ -33,18 +33,17 @@ bibliography: paper.bib
 
 # Summary
 
-Stan provides a probabilistic programming language in which users can
-code Bayesian models [@Carpenter:2017; @Stan:2022]. A Stan program is
-transpiled to to a C++ class which links to the Stan math library to
-implement smooth, unconstrained posterior log densities, gradients,
-and Hessians as well as constraining/unconstraining transforms.
-Implementation is provided through automatic differentiation in the
-Stan math library [@Carpenter:2015]. `BridgeStan` provides in-memory
-access to the methods of Stan models through Python, Julia, and
-R. This allows algorithm development in these languages with the
-numerical efficiency and expressiveness of Stan models.  Furthermore,
-these features are exposed through a language-agnostic C API, allowing
-foreign function interfaces in other languages with minimal additional
+Stan provides a probabilistic programming language in which users can code
+Bayesian models [@Carpenter:2017; @Stan:2022]. A Stan program is transpiled to
+to a C++ class which links to the Stan math library to implement smooth,
+unconstrained posterior log densities, gradients, and Hessians as well as
+constraining/unconstraining transforms.  Implementation is provided through
+automatic differentiation in the Stan math library
+[@Carpenter:2015]. `BridgeStan` provides in-memory access to the methods of Stan
+models through Python, Julia, and R. This allows algorithm development in these
+languages with the numerical efficiency and expressiveness of Stan models.
+Furthermore, these features are exposed through a language-agnostic C API,
+allowing foreign function interfaces in other languages with minimal additional
 development.
 
 
@@ -58,59 +57,54 @@ state-of-the-art, gradient-based algorithms: full Bayesian inference
 with Hamiltonian Monte Carlo, Laplace approximation based on L-BFGS
 optimization, and autodiff variational inference (ADVI).
 
-In the statistical software environment R, Stan is heavily relied upon
-for development of applied statistics packages. Using Google's
-[PageRank](https://en.wikipedia.org/wiki/PageRank) algorithm on the
-dependency graph [@de-Vries:2014] amongst the 19,159 R packages listed
-on the Comprehensive R Archive Network (CRAN) as of 2022-12-31, three
-R packages that exist solely to provide access to Stan rank quite
-well: `rstan` ranks at number 70, `rstantools` 179, and
-`rstanarm` 502. Further, two Python interfaces to Stan, `pystan` and
-`cmdstanpy`, both rank in the top 600 packages by downloads on the
-Python Package Index (PyPI).
+In the statistical software environment R, Stan is heavily relied upon for
+development of applied statistics packages. Using Google's
+[PageRank](https://en.wikipedia.org/wiki/PageRank) algorithm on the dependency
+graph [@de-Vries:2014] amongst the 19,159 R packages listed on the Comprehensive
+R Archive Network (CRAN) as of 2022-12-31, three R packages that exist solely to
+provide access to Stan rank quite well: `rstan` ranks at number 70, `rstantools`
+179, and `rstanarm` 502. Further, two Python interfaces to Stan, `pystan` and
+`cmdstanpy`, both rank in the top 600 packages by downloads on the Python
+Package Index (PyPI).
 
-C++ can be cumbersome for algorithm prototyping, so developers have
-been requesting ways to access Stan models for algorithm development
-in Python, R, and Julia. `BridgeStan` answers this call, making it
-easy for algorithm developers to leverage existing Stan models in
-their evaluation, e.g., the dozens of diverse models with reference
-posteriors in [`posteriordb`](https://github.com/stan-dev/posteriordb)
-[@Magnusson:2022].  By providing access to the method of Stan model,
-`BridgeStan` will help standardize algorithm development by making
-more consistent the models being tested and the implementations of
-those models and their underlying mathematical representations.
[…]
