### 2023-05-23 #5236 @roualdes
Thanks, @salleuska, for your suggestions.  I think your comments have really helped.
 Here's a quick record of the changes we've made, based on your feedback so far.

**Major Comments**
1. Explain user cases, see [commit 5d63b03](https://github.com/roualdes/bridgestan/commit/5d63b03c77cac27ecf1e1be4272268cb7412c13d)
2. Add a new paragraph about similar offerings from other software, see [commit e2cc9ba](https://github.com/roualdes/bridgestan/commit/e2cc9bade449ef2f861d8ac4f09a9ba78d6cbacb) and [commit d8473a9](https://github.com/roualdes/bridgestan/commit/d8473a90a560e6e782f4952285769f666e10888e)

**Minor Comments**
* Specify more clearly what those R packages are all about, see [commit 2932c77](https://github.com/roualdes/bridgestan/commit/2932c77b010b2383b4c99b090da186d1841c3d7a)
* clean up sentence about what BridgeStan offers, see [commit 9c687e4](https://github.com/roualdes/bridgestan/commit/9c687e4d2a42297ae2051781f9356a56198282d8)
* replace phrase by and large, see [commit f2589c7](https://github.com/roualdes/bridgestan/commit/f2589c7ab64f7693b270b2f99c04596861d2d773)
* remove most uses of the word efficient, see [commit 06c5513](https://github.com/roualdes/bridgestan/commit/06c5513eaa9bbcc215205508f6c8b9f7ebbdf4c6)
