### 2023-05-30 13685c4
C++ implementation
