+++ python/bridgestan/__init__.py
@@ -17,7 +17,9 @@ if _plt.system() == "Windows":
         from .compile import get_bridgestan_path
 
         try:
-            out = subprocess.run(["where.exe", "tbb.dll"], check=True, capture_output=True)
+            out = subprocess.run(
+                ["where.exe", "tbb.dll"], check=True, capture_output=True
+            )
             tbb_path = os.path.dirname(out.stdout.decode().splitlines()[0])
             os.add_dll_directory(tbb_path)
         except:
@@ -29,7 +31,14 @@ if _plt.system() == "Windows":
 
         try:
             out = subprocess.run(
-                ["where.exe", "libwinpthread-1.dll", "libgcc_s_seh-1.dll", "libstdc++-6.dll"], check=True, capture_output=True
+                [
+                    "where.exe",
+                    "libwinpthread-1.dll",
+                    "libgcc_s_seh-1.dll",
+                    "libstdc++-6.dll",
+                ],
+                check=True,
+                capture_output=True,
             )
             mingw_dir = os.path.dirname(out.stdout.decode().splitlines()[0])
             os.add_dll_directory(mingw_dir)
