### 2023-06-08 #5236 @Nikoleta-v3
Hey @roualdes 👋🏻  @saumil-sh has agreed to be the second reviewer for your submission. Thank you saumil-sh🙏

All of our communications will happen here from now on. As a reviewer, the first step is to create a checklist for your review by entering:

s the top of a new comment in this thread.

These checklists contain the JOSS requirements ✅ As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines.

Reviewers are encouraged to submit issues and pull requests on the software repository (please mention #5236 so that a link is created to this thread and I can keep an eye on what is happening) or posting their review on this thread.

We aim for reviews to be completed within about 2-4 weeks, and feel free to ping me (@Nikoleta-v3) if you have any questions/concerns. 😄 🙋🏻
