### 2023-06-13 #5236 @salleuska
> Many thanks @roualdes! I did play around with Julia and Python and haven't had any issues to report. Also the documentation seems in pretty good shape to me. I'll take a look at the changes and keep you posted!

I have read the new version of the paper, it is clearer to me! There is just one sentence that I am not following:

line 35 "By providing access to the method of Stan model, BridgeStan aides algorithm development by making more consistent the models being tested and the implementations of those models and their underlying mathematical representations"
-> Plus, what does it mean here that the models (etc..) would be more consistent?

Other than this minor thing, for me, everything is good to go!
