### 2023-06-13 6dd7d3a
Python: Warn if user has 5+ copies downloaded in .bridgestan (#128)
