### 2023-06-13 #5236 @roualdes
Great.  Thanks, @salleuska.

Here's a longer explanation of what I intend to mean by more consistent.  I'd appreciate it if you could offer some suggestions on how to make my attempt at a pithy description of the following more clear.

In my experience, statisticians commonly write numerical code implementing popular statistical models to compare their new algorithm against other algorithms.  The trouble is numerical code can be tricky to write.  There are tricks to avoid under/over-flow, to maintain floating point accuracy, to implement sum-to-one constraints, etc., and the Stan Math library, upon which BridgeStan depends, seeks out implementations to remedy as many of these numerical issues that the Stan Development Team can think of and/or find.  Hence, by using BridgeStan, and thus the Stan Math library, algorithm developers can think more about their algorithm and less about numerical issues in their implementation.

Certainly, just using BridgeStan doesn't guarantee that a user doesn't implement poorly a hand-written log-sum-exp, instead of using Stan's built in and more numerically stable version [log-sum-exp](https://mc-stan.org/docs/stan-users-guide/log-sum-of-exponentials.html). Many (all?) of Stan's distributions' functions (log-densities, cumulative density functions, 1 - CDFs, etc.)  implicitly use such numerical tricks, so using BridgeStan is at least some steps forward in more consistent numerical implementations of statistical models.
