+++ .github/workflows/main.yaml
@@ -275,5 +275,5 @@ jobs:
         run: |
           cargo clippy
           cargo fmt --check
-          cargo run --example=make_model
+          cargo run --example=example
           cargo test --verbose
