### 2023-06-15 roualdes/bridgestan#131 @saumil-sh
Error in `R/example.R`
Hello,
I am [reviewing](https://github.com/openjournals/joss-reviews/issues/5236) this package for submission at JOSS. When I tried to run the `example.R` file from the `bridgestan/R` folder, I get the following error:
```r
Error in .C("bs_model_construct_R", as.character(data), as.integer(seed),  :
  "bs_model_construct_R" not available for .C() for package "bernoulli_model"
```

My guess is that it originates here https://github.com/roualdes/bridgestan/blob/7267e83ad6a8f41ff2e65d49dfcc130c0137cf9e/R/example.R#L7,
is something missing in this call
https://github.com/roualdes/bridgestan/blob/7267e83ad6a8f41ff2e65d49dfcc130c0137cf9e/R/R/bridgestan.R#L42?
