### 2023-06-15 roualdes/bridgestan#131 @WardBrian
Hi @saumil-sh -

Can you share a bit more about how you downloaded bridgestan and what commands you ran before receiving this error? We had an issue in the R example (https://github.com/roualdes/bridgestan/issues/121) which was fixed recently, but if you downloaded one of the .tar.gz from github I do not believe it would have been updated yet.
