+++ R/R/bridgestan.R
@@ -23,7 +23,7 @@ StanModel <- R6::R6Class("StanModel",
       private$seed <- seed
       private$lib <- tools::file_path_as_absolute(lib)
       private$lib_name <- tools::file_path_sans_ext(basename(lib))
-      if (is.loaded("construct_R", PACKAGE = private$lib_name)) {
+      if (is.loaded("bs_model_construct_R", PACKAGE = private$lib_name)) {
         warning(
           paste0("Loading a shared object '", lib, "' which is already loaded.\n",
                   "If the file has changed since the last time it was loaded, this load may not update the library!"
