### 2023-06-15 roualdes/bridgestan#131 @saumil-sh
I had a look at [#121](https://github.com/roualdes/bridgestan/issues/121) before posting this and I belive it is a different issue.

I followed the [documentation](https://roualdes.github.io/bridgestan/latest/getting-started.html#installing-the-latest-version-with-git) and used `git clone --recurse-submodules https://github.com/roualdes/bridgestan.git`.
Here is the relevant lines of the output of my `git log`
```sh
commit 6dd7d3a8a80b9027145fb20d230aa86e5fcf99df (HEAD -> main, origin/main, origin/HEAD)
Author: Brian Ward <bward@flatironinstitute.org>
Date:   Tue Jun 13 13:55:15 2023 -0400

    Python: Warn if user has 5+ copies downloaded in .bridgestan (#128)
```

I am on Mac M1 running Ventura 13.4 and tested the installation according to the [documentation](https://roualdes.github.io/bridgestan/latest/getting-started.html#testing-the-installation) running `make test_models/multi/multi_model.so` from the `root` directory of the repo.

Next, I created a `mamba`/`conda` environment using `mamba create -n stan -c conda-forge` and installed `r-base` and `radian`. Then I followed the relevant [documentation](https://roualdes.github.io/bridgestan/latest/languages/r.html) to install the `R` `bridgestan` package. Here is the relevant version information:
```sh
$ mamba --version
mamba 1.4.2
conda 23.3.1

$ mamba list | grep r-base
r-base                    4.3.0                h4b3f977_1    conda-forge

$ mamba list | grep radian
radian                    0.6.6              pyhd8ed1ab_0    conda-forge

$ radian
R version 4.3.0 (2023-04-21) -- "Already Tomorrow"
Platform: aarch64-apple-darwin20.0.0 (64-bit)

r$> install.packages(getwd(), repos=NULL, type="source")
.
.
.
* DONE (bridgestan)

r$> installed.packages()
           Package      LibPath                                          Version Priority
base       "base"       "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
bridgestan "bridgestan" "/Users/shah/mambaforge/envs/stan/lib/R/library" "2.0.0" NA
compiler   "compiler"   "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
datasets   "datasets"   "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
graphics   "graphics"   "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
grDevices  "grDevices"  "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
grid       "grid"       "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
methods    "methods"    "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
parallel   "parallel"   "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
R6         "R6"         "/Users/shah/mambaforge/envs/stan/lib/R/library" "2.5.1" NA
splines    "splines"    "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
stats      "stats"      "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
stats4     "stats4"     "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
tcltk      "tcltk"      "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
tools      "tools"      "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
utils      "utils"      "/Users/shah/mambaforge/envs/stan/lib/R/library" "4.3.0" "base"
```

I am aware of 2 ways of running an `R` script, one using `Rscript example.R` and another `source("example.R")` from `radian` both threw the same error mentioned in the first post in this thread.

Please let me know if I missed any information.
