### 2023-06-15 roualdes/bridgestan#131 @WardBrian
Did you build the `bernoulli_model` library before running the example? If you were able to build the `multi_model.so` file your installation is most likely working, but you may need to build the specific model the example is using first

https://github.com/roualdes/bridgestan/blob/7267e83ad6a8f41ff2e65d49dfcc130c0137cf9e/R/example.R#L1-L3
