+++ docs/languages/rust.rst
@@ -1,0 +1,44 @@
+Rust Interface
+==============
+
+`See the BridgeStan Crate documentation on docs.rs <https://docs.rs/bridgestan>`__
+
+----
+
+Installation
+------------
+
+The BridgeStan Rust client is availble on crates.io and via ``cargo``:
+
+.. code-block:: shell
+
+    cargo add bridgestan
+
+To build and use BridgeStan models, a copy of the BridgeStan C++ source code
+is required. Please follow the :doc:`Getting Started guide <../getting-started>`
+or use the Rust client in tandem with an interface such as :doc:`Python <./python>`
+which automates this process.
+
+Example Program
+---------------
+
+An example program is provided alongside the Rust crate in ``examples/example.rs``:
+
+.. raw:: html
+
+   <details>
+   <summary><a>Show example.rs</a></summary>
+
+
+.. literalinclude:: ../../rust/examples/example.rs
+   :language: Rust
+
+.. raw:: html
+
+   </details>
+
+
+API Reference
+-------------
+
+See docs.rs for the full API reference:
