### 2023-06-15 roualdes/bridgestan#131 @saumil-sh
I forgot to mention this. I did run `make test_models/bernoulli/bernoulli_model.so` and if I run it again it says:
```sh
$make test_models/bernoulli/bernoulli_model.so
make: 'test_models/bernoulli/bernoulli_model.so' is up to date.
```
Should I "remake" it using the `make -B` flag?
