### 2023-06-15 04020d1
Clean up some docs, un-mark Rng as send/sync
