### 2023-06-15 roualdes/bridgestan#131 @WardBrian
That all seems correct, then.

Could you try running the following command:
`nm test_models/bernoulli/bernoulli_model.so | grep " bs_"`

This should print out the names of the functions available in `bernoulli_model`. It should be the case that `bs_model_construct_R` is on this list, but it is good to check
