### 2023-06-15 roualdes/bridgestan#131 @saumil-sh
I ran the following in the `root` directory of the repository:
```sh
$nm test_models/bernoulli/bernoulli_model.so | grep _bs_
00000000000161ec T _bs_free_error_msg
00000000000161e4 T _bs_free_error_msg_R
0000000000016a7c T _bs_log_density
0000000000016a20 T _bs_log_density_R
0000000000016c5c T _bs_log_density_gradient
0000000000016c00 T _bs_log_density_gradient_R
0000000000016e38 T _bs_log_density_hessian
0000000000016de0 T _bs_log_density_hessian_R
000000000005c1b0 D _bs_major_version
000000000005c8c0 S _bs_minor_version
0000000000015f60 T _bs_model_construct
0000000000015f18 T _bs_model_construct_R
000000000001623c T _bs_model_destruct
0000000000016218 T _bs_model_destruct_R
0000000000016280 T _bs_model_info
0000000000016270 T _bs_model_info_R
0000000000016268 T _bs_name
0000000000016258 T _bs_name_R
0000000000016410 T _bs_param_constrain
00000000000163b0 T _bs_param_constrain_R
00000000000162cc T _bs_param_names
0000000000016288 T _bs_param_names_R
0000000000016360 T _bs_param_num
000000000001631c T _bs_param_num_R
0000000000016314 T _bs_param_unc_names
0000000000016304 T _bs_param_unc_names_R
00000000000163a8 T _bs_param_unc_num
0000000000016398 T _bs_param_unc_num_R
00000000000166d0 T _bs_param_unconstrain
000000000001668c T _bs_param_unconstrain_R
000000000001689c T _bs_param_unconstrain_json
0000000000016854 T _bs_param_unconstrain_json_R
000000000005c8c4 S _bs_patch_version
0000000000017000 T _bs_rng_construct
0000000000016fbc T _bs_rng_construct_R
000000000001720c T _bs_rng_destruct
00000000000171fc T _bs_rng_destruct_R
0000000000017218 T _bs_set_print_callback
00000000000161f0 T _bs_version_R
```
