### 2023-06-15 roualdes/bridgestan#131 @saumil-sh
Maybe relevant: https://stackoverflow.com/questions/45064207/error-with-my-c-files-in-my-package
