### 2023-06-15 roualdes/bridgestan#131 @WardBrian
Can you share your C++ compiler version and how you installed it (xcode, brew, mamba, etc)? The output of `nm` suggests it is mangling the C function names with a leading underscore. I've known of some macos compilers to do this, but it hasn't previously come up for us in this project.

You may be able to delete the model library and then use `CXXFLAGS=-fno-leading-underscore make test_models/bernoulli/bernoulli_model.so` to fix this (to be verified either with `nm` or with the example script), but it would be great to understand why this is occuring/why we haven't encountered it with other Macs we've used
