### 2023-06-15 roualdes/bridgestan#131 @saumil-sh
I will share my C++ compiler and relevant setup and I try your suggestions when I get to it tomorrow.

Python and Julia examples ran as expected.
python:
```sh
$python example.py
This model's name is bernoulli_model.
It has 1 parameters.
log_density and gradient of Bernoulli model: (-27.455880681398384, array([-7.67408425]))
```
julia:
```sh
$julia -e 'using Pkg; Pkg.activate(".."); include("example.jl")'
  Activating project at `/Volumes/shah/projects/reviews/bridgestan/bridgestan`
This model's name is bernoulli_model.
It has 1 parameters.
log_density and gradient of Bernoulli model: (-6.456684984700505, [-2.575183751495763])
```
