### 2023-06-16 roualdes/bridgestan#131 @saumil-sh
Hey,
Here is relevant information about my setup:
When I try `xcode-select --install` It directs me to I updated xcode using `System Settings > Software Update`, and that is how I updated it just before cloning the repo.
```sh
$xcode-select --version
xcode-select version 2397.
```
`clang` and `make` were installed as the dependency in the `mamba` environment I created.
```sh
$clang --version
clang version 16.0.5
Target: arm64-apple-darwin22.5.0
Thread model: posix
InstalledDir: /Users/shah/mambaforge/envs/stan/bin

$clang++ --version
clang version 16.0.5
Target: arm64-apple-darwin22.5.0
Thread model: posix
InstalledDir: /Users/shah/mambaforge/envs/stan/bin

$make --version
GNU Make 4.3
Built for arm64-apple-darwin20.0.0
```

The `xcode` command line tools provide other compilers.
```sh
g++ --version
Apple clang version 14.0.3 (clang-1403.0.22.14.1)
Target: arm64-apple-darwin22.5.0
Thread model: posix
InstalledDir: /Library/Developer/CommandLineTools/usr/bin
```

I will rebuild with no leading underscore flag and let you know how it goes.
