### 2023-06-16 roualdes/bridgestan#131 @saumil-sh
I ran `make clean` and the built the `multi` model again. There are 43 warnings generated, I am putting them here in case it may be useful.
```sh
$make test_models/multi/multi_model.so
curl -L https://github.com/stan-dev/stanc3/releases/download/v2.32.1/mac-stanc -o ./bin/stanc --retry 5 --retry-delay 10
                                 Dload  Upload   Total   Spent    Left  Speed
100 10.8M  100 10.8M    0     0   9.8M      0  0:00:01  0:00:01 --:--:-- 28.7M
chmod +x ./bin/stanc

--- Compiling Stan bridge C++ code ---
arm64-apple-darwin20.0.0-clang++ -ftree-vectorize -fPIC -fPIE -fstack-protector-strong -O2 -pipe -stdlib=libc++ -fvisibility-inlines-hidden -fmessage-length=0 -isystem /Users/shah/mambaforge/envs/stan/include -std=c++1y -Wno-unknown-warning-option -Wno-tautological-compare -Wno-sign-compare -D_REENTRANT -Wno-ignored-attributes      -I ./stan/lib/stan_math/lib/tbb_2020.3/include    -O3 -I ./stan/src -I ./stan/lib/rapidjson_1.1.0/ -I ./stan/lib/stan_math/ -I ./stan/lib/stan_math/lib/eigen_3.4.0 -I ./stan/lib/stan_math/lib/boost_1.78.0 -I ./stan/lib/stan_math/lib/sundials_6.1.1/include -I ./stan/lib/stan_math/lib/sundials_6.1.1/src/sundials -fPIC -D_FORTIFY_SOURCE=2 -isystem /Users/shah/mambaforge/envs/stan/include    -DBOOST_DISABLE_ASSERTS          -c -o src/bridgestan.o       -Wl,-L,"/Volumes/shah/projects/reviews/bridgestan/bridgestan/stan/lib/stan_math/lib/tbb" -Wl,-rpath,"/Volumes/shah/projects/reviews/bridgestan/bridgestan/stan/lib/stan_math/lib/tbb"     src/bridgestan.cpp
clang-16: warning: -Wl,-L,/Volumes/shah/projects/reviews/bridgestan/bridgestan/stan/lib/stan_math/lib/tbb: 'linker' input unused [-Wunused-command-line-argument]
clang-16: warning: -Wl,-rpath,/Volumes/shah/projects/reviews/bridgestan/bridgestan/stan/lib/stan_math/lib/tbb: 'linker' input unused [-Wunused-command-line-argument]
In file included from src/bridgestan.cpp:1:
In file included from src/bridgestan.h:5:
In file included from src/model_rng.hpp:4:
In file included from ./stan/src/stan/model/model_base.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/rev/core.hpp:10:
In file included from ./stan/lib/stan_math/stan/math/rev/core/chainable_object.hpp:6:
In file included from ./stan/lib/stan_math/stan/math/rev/core/typedefs.hpp:7:
In file included from ./stan/lib/stan_math/stan/math/rev/core/Eigen_NumTraits.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/prim/core.hpp:4:
In file included from ./stan/lib/stan_math/stan/math/prim/core/init_threadpool_tbb.hpp:6:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast.hpp:31:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/bad_lexical_cast.hpp:28:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/throw_exception.hpp:24:
./stan/lib/stan_math/lib/boost_1.78.0/boost/assert/source_location.hpp:75:14: warning: 'sprintf' is deprecated: This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead. [-Wdeprecated-declarations]
        std::sprintf( buffer, ":%ld", static_cast<long>( line() ) );
             ^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/stdio.h:188:1: note: 'sprintf' has been explicitly marked deprecated here
__deprecated_msg("This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead.")
^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/sys/cdefs.h:215:48: note: expanded from macro '__deprecated_msg'
        #define __deprecated_msg(_msg) __attribute__((__deprecated__(_msg)))
                                                      ^
In file included from src/bridgestan.cpp:1:
In file included from src/bridgestan.h:5:
In file included from src/model_rng.hpp:4:
In file included from ./stan/src/stan/model/model_base.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/rev/core.hpp:10:
In file included from ./stan/lib/stan_math/stan/math/rev/core/chainable_object.hpp:6:
In file included from ./stan/lib/stan_math/stan/math/rev/core/typedefs.hpp:7:
In file included from ./stan/lib/stan_math/stan/math/rev/core/Eigen_NumTraits.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/prim/core.hpp:4:
In file included from ./stan/lib/stan_math/stan/math/prim/core/init_threadpool_tbb.hpp:6:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast.hpp:31:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/bad_lexical_cast.hpp:28:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/throw_exception.hpp:24:
./stan/lib/stan_math/lib/boost_1.78.0/boost/assert/source_location.hpp:80:18: warning: 'sprintf' is deprecated: This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead. [-Wdeprecated-declarations]
            std::sprintf( buffer, ":%ld", static_cast<long>( column() ) );
                 ^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/stdio.h:188:1: note: 'sprintf' has been explicitly marked deprecated here
__deprecated_msg("This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead.")
^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/sys/cdefs.h:215:48: note: expanded from macro '__deprecated_msg'
        #define __deprecated_msg(_msg) __attribute__((__deprecated__(_msg)))
                                                      ^
In file included from src/bridgestan.cpp:1:
In file included from src/bridgestan.h:5:
In file included from src/model_rng.hpp:4:
In file included from ./stan/src/stan/model/model_base.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/rev/core.hpp:10:
In file included from ./stan/lib/stan_math/stan/math/rev/core/chainable_object.hpp:6:
In file included from ./stan/lib/stan_math/stan/math/rev/core/typedefs.hpp:7:
In file included from ./stan/lib/stan_math/stan/math/rev/core/Eigen_NumTraits.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/prim/core.hpp:4:
In file included from ./stan/lib/stan_math/stan/math/prim/core/init_threadpool_tbb.hpp:6:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast.hpp:32:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/try_lexical_convert.hpp:44:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/detail/converter_lexical.hpp:54:
./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/detail/converter_lexical_streams.hpp:285:21: warning: 'sprintf' is deprecated: This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead. [-Wdeprecated-declarations]
                    sprintf(begin,
                    ^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/stdio.h:188:1: note: 'sprintf' has been explicitly marked deprecated here
__deprecated_msg("This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead.")
^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/sys/cdefs.h:215:48: note: expanded from macro '__deprecated_msg'
        #define __deprecated_msg(_msg) __attribute__((__deprecated__(_msg)))
                                                      ^
In file included from src/bridgestan.cpp:1:
In file included from src/bridgestan.h:5:
In file included from src/model_rng.hpp:4:
In file included from ./stan/src/stan/model/model_base.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/rev/core.hpp:10:
In file included from ./stan/lib/stan_math/stan/math/rev/core/chainable_object.hpp:6:
In file included from ./stan/lib/stan_math/stan/math/rev/core/typedefs.hpp:7:
In file included from ./stan/lib/stan_math/stan/math/rev/core/Eigen_NumTraits.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/prim/core.hpp:4:
In file included from ./stan/lib/stan_math/stan/math/prim/core/init_threadpool_tbb.hpp:6:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast.hpp:32:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/try_lexical_convert.hpp:44:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/detail/converter_lexical.hpp:54:
./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/detail/converter_lexical_streams.hpp:297:21: warning: 'sprintf' is deprecated: This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead. [-Wdeprecated-declarations]
                    sprintf(begin,
                    ^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/stdio.h:188:1: note: 'sprintf' has been explicitly marked deprecated here
__deprecated_msg("This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead.")
^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/sys/cdefs.h:215:48: note: expanded from macro '__deprecated_msg'
        #define __deprecated_msg(_msg) __attribute__((__deprecated__(_msg)))
                                                      ^
In file included from src/bridgestan.cpp:1:
In file included from src/bridgestan.h:5:
In file included from src/model_rng.hpp:4:
In file included from ./stan/src/stan/model/model_base.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/rev/core.hpp:10:
In file included from ./stan/lib/stan_math/stan/math/rev/core/chainable_object.hpp:6:
In file included from ./stan/lib/stan_math/stan/math/rev/core/typedefs.hpp:7:
In file included from ./stan/lib/stan_math/stan/math/rev/core/Eigen_NumTraits.hpp:5:
In file included from ./stan/lib/stan_math/stan/math/prim/core.hpp:4:
In file included from ./stan/lib/stan_math/stan/math/prim/core/init_threadpool_tbb.hpp:6:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast.hpp:32:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/try_lexical_convert.hpp:44:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/detail/converter_lexical.hpp:54:
./stan/lib/stan_math/lib/boost_1.78.0/boost/lexical_cast/detail/converter_lexical_streams.hpp:310:21: warning: 'sprintf' is deprecated: This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead. [-Wdeprecated-declarations]
                    sprintf(begin,
                    ^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/stdio.h:188:1: note: 'sprintf' has been explicitly marked deprecated here
__deprecated_msg("This function is provided for compatibility reasons only.  Due to security concerns inherent in the design of sprintf(3), it is highly recommended that you use snprintf(3) instead.")
^
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/include/sys/cdefs.h:215:48: note: expanded from macro '__deprecated_msg'
        #define __deprecated_msg(_msg) __attribute__((__deprecated__(_msg)))
                                                      ^
In file included from src/bridgestan.cpp:2:
In file included from src/model_rng.cpp:4:
In file included from ./stan/src/stan/io/json/json_data.hpp:4:
In file included from ./stan/src/stan/io/json/json_data_handler.hpp:18:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string.hpp:23:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/split.hpp:16:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/iter_find.hpp:27:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/find_iterator.hpp:24:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/detail/find_iterator.hpp:18:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/function.hpp:30:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/function/detail/prologue.hpp:17:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/function/function_base.hpp:21:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/type_index.hpp:29:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/type_index/stl_type_index.hpp:47:
./stan/lib/stan_math/lib/boost_1.78.0/boost/container_hash/hash.hpp:132:33: warning: 'unary_function<bool, unsigned long>' is deprecated [-Wdeprecated-declarations]
        struct hash_base : std::unary_function<T, std::size_t> {};
                                ^
./stan/lib/stan_math/lib/boost_1.78.0/boost/container_hash/hash.hpp:609:5: note: in instantiation of template class 'boost::hash_detail::hash_base<bool>' requested here
    BOOST_HASH_SPECIALIZE(bool)
    ^
./stan/lib/stan_math/lib/boost_1.78.0/boost/container_hash/hash.hpp:581:19: note: expanded from macro 'BOOST_HASH_SPECIALIZE'
         : public boost::hash_detail::hash_base<type> \
                  ^
/Users/shah/mambaforge/envs/stan/bin/../include/c++/v1/__functional/unary_function.h:23:29: note: 'unary_function<bool, unsigned long>' has been explicitly marked deprecated here
struct _LIBCPP_TEMPLATE_VIS _LIBCPP_DEPRECATED_IN_CXX11 unary_function
                            ^
/Users/shah/mambaforge/envs/stan/bin/../include/c++/v1/__config:862:41: note: expanded from macro '_LIBCPP_DEPRECATED_IN_CXX11'
#    define _LIBCPP_DEPRECATED_IN_CXX11 _LIBCPP_DEPRECATED
                                        ^
/Users/shah/mambaforge/envs/stan/bin/../include/c++/v1/__config:847:49: note: expanded from macro '_LIBCPP_DEPRECATED'
#      define _LIBCPP_DEPRECATED __attribute__((deprecated))
                                                ^
In file included from src/bridgestan.cpp:2:
In file included from src/model_rng.cpp:4:
In file included from ./stan/src/stan/io/json/json_data.hpp:4:
In file included from ./stan/src/stan/io/json/json_data_handler.hpp:18:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string.hpp:23:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/split.hpp:16:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/iter_find.hpp:27:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/find_iterator.hpp:24:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/detail/find_iterator.hpp:18:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/function.hpp:30:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/function/detail/prologue.hpp:17:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/function/function_base.hpp:21:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/type_index.hpp:29:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/type_index/stl_type_index.hpp:47:
./stan/lib/stan_math/lib/boost_1.78.0/boost/container_hash/hash.hpp:132:33: warning: 'unary_function<char, unsigned long>' is deprecated [-Wdeprecated-declarations]
        struct hash_base : std::unary_function<T, std::size_t> {};
                                ^
./stan/lib/stan_math/lib/boost_1.78.0/boost/container_hash/hash.hpp:610:5: note: in instantiation of template class 'boost::hash_detail::hash_base<char>' requested here
    BOOST_HASH_SPECIALIZE(char)
    ^
./stan/lib/stan_math/lib/boost_1.78.0/boost/container_hash/hash.hpp:581:19: note: expanded from macro 'BOOST_HASH_SPECIALIZE'
         : public boost::hash_detail::hash_base<type> \
                  ^
/Users/shah/mambaforge/envs/stan/bin/../include/c++/v1/__functional/unary_function.h:23:29: note: 'unary_function<char, unsigned long>' has been explicitly marked deprecated here
struct _LIBCPP_TEMPLATE_VIS _LIBCPP_DEPRECATED_IN_CXX11 unary_function
                            ^
/Users/shah/mambaforge/envs/stan/bin/../include/c++/v1/__config:862:41: note: expanded from macro '_LIBCPP_DEPRECATED_IN_CXX11'
#    define _LIBCPP_DEPRECATED_IN_CXX11 _LIBCPP_DEPRECATED
                                        ^
/Users/shah/mambaforge/envs/stan/bin/../include/c++/v1/__config:847:49: note: expanded from macro '_LIBCPP_DEPRECATED'
#      define _LIBCPP_DEPRECATED __attribute__((deprecated))
                                                ^
In file included from src/bridgestan.cpp:2:
In file included from src/model_rng.cpp:4:
In file included from ./stan/src/stan/io/json/json_data.hpp:4:
In file included from ./stan/src/stan/io/json/json_data_handler.hpp:18:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string.hpp:23:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/split.hpp:16:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/iter_find.hpp:27:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/find_iterator.hpp:24:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/detail/find_iterator.hpp:18:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/function.hpp:30:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/function/detail/prologue.hpp:17:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/function/function_base.hpp:21:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/type_index.hpp:29:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/type_index/stl_type_index.hpp:47:
./stan/lib/stan_math/lib/boost_1.78.0/boost/container_hash/hash.hpp:132:33: warning: 'unary_function<signed char, unsigned long>' is deprecated [-Wdeprecated-declarations]
        struct hash_base : std::unary_function<T, std::size_t> {};
                                ^
./stan/lib/stan_math/lib/boost_1.78.0/boost/container_hash/hash.hpp:611:5: note: in instantiation of template class 'boost::hash_detail::hash_base<signed char>' requested here
    BOOST_HASH_SPECIALIZE(signed char)
    ^
./stan/lib/stan_math/lib/boost_1.78.0/boost/container_hash/hash.hpp:581:19: note: expanded from macro 'BOOST_HASH_SPECIALIZE'
         : public boost::hash_detail::hash_base<type> \
                  ^
/Users/shah/mambaforge/envs/stan/bin/../include/c++/v1/__functional/unary_function.h:23:29: note: 'unary_function<signed char, unsigned long>' has been explicitly marked deprecated here
struct _LIBCPP_TEMPLATE_VIS _LIBCPP_DEPRECATED_IN_CXX11 unary_function
                            ^
/Users/shah/mambaforge/envs/stan/bin/../include/c++/v1/__config:862:41: note: expanded from macro '_LIBCPP_DEPRECATED_IN_CXX11'
#    define _LIBCPP_DEPRECATED_IN_CXX11 _LIBCPP_DEPRECATED
                                        ^
/Users/shah/mambaforge/envs/stan/bin/../include/c++/v1/__config:847:49: note: expanded from macro '_LIBCPP_DEPRECATED'
#      define _LIBCPP_DEPRECATED __attribute__((deprecated))
                                                ^
In file included from src/bridgestan.cpp:2:
In file included from src/model_rng.cpp:4:
In file included from ./stan/src/stan/io/json/json_data.hpp:4:
In file included from ./stan/src/stan/io/json/json_data_handler.hpp:18:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string.hpp:23:
In file included from ./stan/lib/stan_math/lib/boost_1.78.0/boost/algorithm/string/split.hpp:16:
[…]
