### 2023-06-16 roualdes/bridgestan#131 @saumil-sh
Aha! I think `clang` doesn't know to compile without the leading underscores.
```sh
$CXXFLAGS=-fno-leading-underscore make test_models/bernoulli/bernoulli_model.so

--- Translating Stan model to C++ code ---
./bin/stanc  --o=test_models/bernoulli/bernoulli.hpp test_models/bernoulli/bernoulli.stan

--- Compiling C++ code ---
arm64-apple-darwin20.0.0-clang++ -fno-leading-underscore -std=c++1y -Wno-unknown-warning-option -Wno-tautological-compare -Wno-sign-compare -D_REENTRANT -Wno-ignored-attributes      -I ./stan/lib/stan_math/lib/tbb_2020.3/include    -O3 -I ./stan/src -I ./stan/lib/rapidjson_1.1.0/ -I ./stan/lib/stan_math/ -I ./stan/lib/stan_math/lib/eigen_3.4.0 -I ./stan/lib/stan_math/lib/boost_1.78.0 -I ./stan/lib/stan_math/lib/sundials_6.1.1/include -I ./stan/lib/stan_math/lib/sundials_6.1.1/src/sundials -fPIC -D_FORTIFY_SOURCE=2 -isystem /Users/shah/mambaforge/envs/stan/include    -DBOOST_DISABLE_ASSERTS          -c -x c++ -o test_models/bernoulli/bernoulli.o test_models/bernoulli/bernoulli.hpp
clang-16: error: unknown argument: '-fno-leading-underscore'
make: *** [Makefile:62: test_models/bernoulli/bernoulli_model.so] Error 1
```
