### 2023-06-16 roualdes/bridgestan#131 @saumil-sh
Maybe relevant:
https://stackoverflow.com/questions/19101815/how-can-c-code-call-an-external-function-without-a-leading-underscore
