### 2023-06-16 roualdes/bridgestan#131 @WardBrian
> Python and Julia examples ran as expected.

Because of this, I believe the issue is with R, not with the compiler, since all three languages will be using the same dylib. The Python and Julia clients both request the function names without the leading underscore, and their standard libraries seem to know to add it. But R isn’t, at least not the build from conda (I know several Mac/R users who this problem has not surfaced for).

Just to confirm I’d like @roualdes to check (using `nm`) if the symbols on his Mac have the leading underscore. If so we can be fairly confident that the issue is not with compilation but in the R runtime
