### 2023-06-16 roualdes/bridgestan#131 @roualdes
```
nm bernoulli_model.so | grep _bs_
0000000000021f20 T _bs_free_error_msg
0000000000021f10 T _bs_free_error_msg_R
00000000000235c0 T _bs_log_density
0000000000023570 T _bs_log_density_R
0000000000023c40 T _bs_log_density_gradient
0000000000023bf0 T _bs_log_density_gradient_R
00000000000242d0 T _bs_log_density_hessian
0000000000024270 T _bs_log_density_hessian_R
0000000000074128 D _bs_major_version
0000000000074c78 S _bs_minor_version
0000000000021820 T _bs_model_construct
00000000000217e0 T _bs_model_construct_R
0000000000021f80 T _bs_model_destruct
0000000000021f50 T _bs_model_destruct_R
0000000000021fe0 T _bs_model_info
0000000000021fd0 T _bs_model_info_R
0000000000021fc0 T _bs_name
0000000000021fb0 T _bs_name_R
0000000000022180 T _bs_param_constrain
0000000000022130 T _bs_param_constrain_R
0000000000022040 T _bs_param_names
0000000000021ff0 T _bs_param_names_R
00000000000220e0 T _bs_param_num
0000000000022090 T _bs_param_num_R
0000000000022080 T _bs_param_unc_names
0000000000022070 T _bs_param_unc_names_R
0000000000022120 T _bs_param_unc_num
0000000000022110 T _bs_param_unc_num_R
00000000000228d0 T _bs_param_unconstrain
0000000000022890 T _bs_param_unconstrain_R
0000000000022f40 T _bs_param_unconstrain_json
0000000000022f00 T _bs_param_unconstrain_json_R
0000000000074c7c S _bs_patch_version
0000000000024950 T _bs_rng_construct
0000000000024910 T _bs_rng_construct_R
0000000000025030 T _bs_rng_destruct
0000000000025010 T _bs_rng_destruct_R
0000000000025050 T _bs_set_print_callback
0000000000021f30 T _bs_version_R
```
