### 2023-06-16 roualdes/bridgestan#131 @roualdes
I tried R 4.2 and 4.3 on Platform: x86_64-apple-darwin20 (64-bit), Running under: macOS Monterey 12.6.6 with both the R console and radian and the R example works as is, for me.

As far as I understand this, we've narrowed the issue down to either a aarch64-apple-darwin20.0.0 issue and/or a mamba/conda issue.  I can't do anything about not having an M1 and I'm hesitant to install mamba/conda, because I don't know how to sandbox the install.

@saumil-sh would you mind trying to install R outside of a mamba/conda environment, and then trying the R example again?
