+++ rust/src/bs_safe.rs
@@ -152,6 +152,12 @@ pub struct Rng<T: Borrow<StanLibrary>> {
     lib: T,
 }
 
+// Use sites require exclusive reference which guarantees
+// that the rng is not used in multiple threads concurrently.
+unsafe impl<T: Sync + Borrow<StanLibrary>> Sync for Rng<T> {}
+unsafe impl<T: Send + Borrow<StanLibrary>> Send for Rng<T> {}
+
+
 impl<T: Borrow<StanLibrary>> Drop for Rng<T> {
     fn drop(&mut self) {
         unsafe {
