### 2023-06-16 5875b3a
R: set NAOK
