+++ R/R/bridgestan.R
@@ -303,6 +303,7 @@ StanModel <- R6::R6Class("StanModel",
         return_code = as.integer(0),
         err_msg = as.character(""),
         err_ptr = raw(8),
+        NAOK = TRUE,
         PACKAGE = private$lib_name
       )
       if (vars$return_code) {
