### 2023-06-16 2e6fbc6
Improve rust API docs
