+++ rust/src/bs_safe.rs
@@ -80,18 +80,25 @@ pub struct LoadingError(#[from] libloading::Error);
 #[derive(Error, Debug)]
 #[non_exhaustive]
 pub enum BridgeStanError {
+    /// The prodided library could not be loaded.
     #[error(transparent)]
     InvalidLibrary(#[from] LoadingError),
+    /// The version of the Stan library does not match the version of the rust crate.
     #[error("Bad Stan library version: Got {0} but expected {1}")]
     BadLibraryVersion(String, String),
+    /// The Stan library could not be loaded because it was compiled without threading support.
     #[error("The Stan library was compiled without threading support. Config was {0}")]
     StanThreads(String),
+    /// Stan returned a string that couldn't be decoded using UTF8.
     #[error("Failed to decode string to UTF8")]
     InvalidString(#[from] Utf8Error),
+    /// The model could not be instanciated, possibly because if incorrect data.
     #[error("Failed to construct model: {0}")]
     ConstructFailed(String),
+    /// Stan returned an error while computing the density.
     #[error("Failed during evaluation: {0}")]
     EvaluationFailed(String),
+    /// Setting a print-callback failed.
     #[error("Failed to set a print-callback: {0}")]
     SetCallbackFailed(String),
 }
@@ -209,7 +216,7 @@ impl<'lib> ErrorMsg<'lib> {
 
     /// Return the error message as a String.
     ///
-    /// Panics if there was no error message.
+    /// *Panics* if there was no error message.
     fn message(&self) -> String {
         NonNull::new(self.msg)
             .map(|msg| {
@@ -270,8 +277,13 @@ impl<T: Borrow<StanLibrary>> Model<T> {
     }
 
     /// Create a new `Rng` random number generator from the library underlying this model.
+    ///
     /// This can be used in `param_constrain` when values from the `generated quantities`
     /// block are desired.
+    ///
+    /// This instance can only be used with models from the same
+    /// stan library. Invalid usage will otherwise result in a
+    /// panic.
     pub fn new_rng(&self, seed: u32) -> Result<Rng<&StanLibrary>> {
         Rng::new(self.ref_library(), seed)
     }
@@ -293,14 +305,14 @@ impl<T: Borrow<StanLibrary>> Model<T> {
     ///
     /// The parameters are returned in the order they are declared.
     /// Multivariate parameters are return in column-major (more
-    /// generally last-index major) order.  Parameters are separated with
-    /// periods (`.`).  For example, `a[3]` is written `a.3` and `b[2,
-    /// 3]` as `b.2.3`.  The numbering follows Stan and is indexed from 1.
+    /// generally last-index major) order.  Parameter indices are separated
+    /// with periods (`.`).  For example, `a[3]` is written `a.3` and `b[2, 3]`
+    /// as `b.2.3`.  The numbering follows Stan and is indexed from 1.
     ///
-    /// # Arguments
-    ///
-    /// `include_tp`: Include transformed parameters
-    /// `include_gp`: Include generated quantities
+    /// If `include_tp` is set the names will also include the transformed
+    /// parameters of the Stan model after the parameters. If `include_gq` is
+    /// set, we also include the names of the generated quantities at
+    /// the very end.
     pub fn param_names(&self, include_tp: bool, include_gq: bool) -> &str {
         let cstr = unsafe {
             CStr::from_ptr(self.ffi_lib().bs_param_names(
@@ -319,7 +331,7 @@ impl<T: Borrow<StanLibrary>> Model<T> {
     ///
     /// The parameters are returned in the order they are declared.
     /// Multivariate parameters are return in column-major (more
-    /// generally last-index major) order.  Parameters are separated with
[…]
