### 2023-06-17 roualdes/bridgestan#131 @saumil-sh
Hi,
I installed an `R` instance outside mamba using `brew install r`
```sh
$R --version
R version 4.3.1 (2023-06-16) -- "Beagle Scouts"
Copyright (C) 2023 The R Foundation for Statistical Computing
Platform: aarch64-apple-darwin22.4.0 (64-bit)
```
Made sure Bernoulli example is compiled, `R6`, and `bridgestan` are installed.
et voilà
```sh
r$> source("example.R")
[1] "This model's name is bernoulli_model."
[1] "This model has 1 parameters."
[1] "log_density and gradient of Bernoulli model: -8.8324194103925, -4.27513683401048"
```
So, the `conda` distribution of the `R` is the culprit here.
