### 2023-06-17 roualdes/bridgestan#131 @roualdes
This is good to know.  Thanks @saumil-sh for discovering and helping us find the root of this issue.

What do you all think is the appropriate action item from this?  Add a note in the install doc for R?  Do nothing until this issue surfaces again?  Other?
