### 2023-06-17 roualdes/bridgestan#131 @WardBrian
I can file an issue against https://github.com/conda-forge/r-base-feedstock, which is where the Conda recipe is maintained
