### 2023-06-17 roualdes/bridgestan#131 @saumil-sh
A note in the documentation is certainly a good idea until this is resolved.
