### 2023-06-19 #5236 @saumil-sh
## Review report `BridgeStan`: Efficient in-memory access to the methods of a _Stan_ model
_Stan_ is a probabilistic programming language for statistical modeling along with the _Stan math_ library, a `C++` program with automatic differentiation capabilities. The _Stan math_ library provides a template class with a range of functions for _Stan_ models as methods. Markov chain Monte Carlo (MCMC) algorithms utilize these methods of _Stan_ model classes. MCMC are computational statistics algorithms that offer means of sampling from analytically intractable probability distributions. The development in MCMC methods, like Hamiltonian Monte Carlo (HMC), provides a performance gain enabling sophisticated statistical and Bayesian inference. The _BridgeStan_ library grants access to the methods of the _Stan_ model, a `C++` class, from `Julia`, `Python`, and `R`; thus, facilitating the development and implementation of MCMC algorithms in high-level languages.

The interfaces of _Stan_, e.g. `CmdStan`, provide means of transpiling, compiling, and accessing results. The functionality of _BridgeStan_ differs from these interfaces as it exposes the _Stan_ model class with `extern "C"` allowing interface with other programming languages. There is some overlap with `rstan`; however, `rstan` is not under active development. In principle, the reach of _BridgeStan_ can be widened to any language of choice, e.g. the repository features a contributed [`rust` interface](https://github.com/roualdes/bridgestan/tree/main/rust). The documentation and the accompanying report are adequately detailed to convey the fundamentals and workings. The report features an illustrative implementation of Metropolis-adjusted Langevin, an MCMC algorithm. Overall, I suggest revising the report and documentation for clarity. Please consider the following suggestions.

### Major comments
1. The [Stan Model Server](https://github.com/bob-carpenter/stan-model-server/), and [ReddingStan](https://github.com/dmuck/redding-stan) paved the way for _BridgeStan_, please consider citing or acknowledging them in the report.

2. Lines 61-67 are ambiguous, at least to me, whether _BridgeStan_ is an alternative to `JAX` and `Turing.jl`/`JuliAD`. If I understand correctly, automatic differentiation is implemented in the _Stan math_ library; therefore, `Jax`, `JuliaAD`, and _Stan math_ are comparable; however, this comparison is not in the scope of this submission. A relevant performance benchmark of _BridgeStan_ could be how much overhead (time, memory, or other resources) _BridgeStan_ adds to the functionality of _Stan_. Such benchmark supports claims like:
    > BridgeStan is a zero-cost abstraction built upon Stan's math library. (Line 74)

    Additionally, a benchmark could favor the choice of, for example, _Stan_ + _BridgeStan_ + `Julia` over `Turing.jl`. However, adding such a benchmark is time-consuming, and I would leave this decision to the authors.

3. Please consider substituting hyperlinks and Wikipedia references in the report with more reliable or valuable references wherever applicable. For example, please consider changing PageRank to an appropriate publication. Please consider hyperlinking the _Stan_ Discourse threads discussing the need for access to _Stan_ model methods rather than hyperlinking the _Stan_ Discourse URL.

4. An admirable scope of _BridgeStan_ is that it facilitates the prototyping algorithms. However, a new user wanting to prototype an algorithm may need help as the documentation assumes familiarity with dependencies. Please consider making the documentation accessible by being more descriptive and explicit. Please consider relevant minor suggestions.

### Minor comments
1. at first glance, it is ambiguous whether _Stan_ is required and whether documentation assumes a working installation. _Stan_ and _Stan math_ are shipped as `git submodules` when cloning the _BridgeStan_ `git` repository. Please consider mentioning this at the beginning of the section [Getting Started](https://roualdes.github.io/bridgestan/latest/getting-started.html#getting-started).

2. When running `git clone` with the `--recurse-submodules` flag, the entire `git` repositories of _Stan_ and _Stan math_ are cloned. I understand that _Stan_ and _Stan math_ are required; however, their `git` histories are not. The _BridgeStan_ `git` repository is `1.52MB` whereas _Stan_ `283MB` and _Stan math_ `642MB`! Please avoid this using `--depth 1` or `--shallow-submodules`.

3. What about pre-existing _Stan_ and the user choosing to obtain _BridgeStan_ from a package repository (of the language) of their choice, e.g. _PyPi_. Is it possible to connect pre-existing _Stan_ and _BridgeStan_? Does the user just need to declare paths for the `.stan`/`.so` and `.data.json` files? What is the `set_bridgestan_path()` in this case, and how to find it?
    > As of version 1.0.2, the Julia and Python interfaces will download the source for you, so this section is optional.

    Therefore, in any case, the _BridgeStan_ source is downloaded and is the `bridgestan_path`. Please consider making documentation explicit for this case.

4. What decides whether the _BridgeStan_ language interface can compile _Stan_ model? Currently, the `Python` and `Julia` interfaces can, and `R` and `Rust` cannot compile a _Stan model_. Please consider highlighting this in the documentation.

5. The section on [`R`-compatible function](https://roualdes.github.io/bridgestan/latest/languages/c-api.html#r-compatible-functions) mentions,
    > These are small shims which call the above functions.

    However, on its own, the section could be more helpful. Please consider (hyper-)linking all the `function_R()` to the documentation of `function()`.

6. Please add a warning against using `conda`/`mamba` `R`-recipe until the [issue #131](https://github.com/roualdes/bridgestan/issues/131) is resolved.
