### 2023-06-26 roualdes/bridgestan#136 @roualdes
Add to Getting Started doc necessary dependencies on Stan and Stan Math
From our review for the Journal of Open Source Software, https://github.com/openjournals/joss-reviews/issues/5236, it was recommended that we address more clearly our software dependencies on Stan and Stan Math, including mentioning these as included git submodules.
