### 2023-06-26 roualdes/bridgestan#137 @roualdes
Add to Getting Started doc recommendation to use --shallow-submodules 
From our review for the Journal of Open Source Software, https://github.com/openjournals/joss-reviews/issues/5236, it was recommended that we suggest to users to clone the BridgeStan repository with `--shallow-submodules` since we don't necessarily need the git histories associated with Stan and Stan Math and this will save much disk space.

On my machine cloning BridgeStan using `--shallow-submodules` reports

```
$ du -hs bridgestan
698M	bridgestan
```

and without `--shallow-submodules`

```
$ du -hs bridgestan
1.2G	bridgestan
```
