### 2023-06-26 roualdes/bridgestan#136 @bob-carpenter
Seems obvious to us, but definitely worth mentioning.  We can just add a section on dependencies and include NumPy.  Is there anything else?
