### 2023-06-26 roualdes/bridgestan#137 @bob-carpenter
We can just mention the current sizes doing it both ways.  I had no idea you could do that.
