### 2023-06-26 roualdes/bridgestan#137 @WardBrian
We create the release using `tar --exclude-vcs`, which completely removes the `.git` folders (including from the submodules). By not saving any history it gets quite small.
