### 2023-06-26 roualdes/bridgestan#137 @WardBrian
The equivalent for a git clone would be something like

```
git clone --recurse-submodules --shallow-submodules --depth=1 https://github.com/roualdes/bridgestan.git
```
