### 2023-06-26 roualdes/bridgestan#138 @roualdes
Add to Getting Started doc how to specify pre-existing Stan directory
From our review for the Journal of Open Source Software, openjournals/joss-reviews#5236, it was recommended that we add to our documentation details on how a user could link to a pre-existing Stan directory instead of cloning Stan and Stan Math as submodules from within the BridgeStan repository.
