### 2023-06-26 roualdes/bridgestan#139 @roualdes
Add to doc a section highlighting the functionality or lack there of across the interfaces
From our review for the Journal of Open Source Software, openjournals/joss-reviews#5236, it was recommended that we add details discussing the similarities and differences across the interfaces.
