### 2023-06-26 roualdes/bridgestan#138 @WardBrian
I can tackle all three of these getting started page suggestions
