### 2023-06-26 3b9dde9
Add links to documentation of _R functions in C API
