### 2023-06-26 978d5a0
Add --shallow-submodules, closes #137
