### 2023-06-26 1ce7869
Mention Stan and Math as submodules, closes #136
