+++ docs/getting-started.rst
@@ -18,7 +18,9 @@ Downloading BridgeStan
 ----------------------
 
 Installing BridgeStan is as simple as ensuring that the above requirements are installed and then downloading
-the source repository.
+the source repository. All of the following ways of downloading BridgeStan will additionally download the
+[Stan](https://github.com/stan-dev/stan) and [Stan Math](https://github.com/stan-dev/math) libraries as submodules,
+no additional dependencies are required to be installed separately for the C++ source code.
 
 As of version 1.0.2, the :doc:`Julia <languages/julia>` and
 :doc:`Python <languages/python>` interfaces will download
