### 2023-06-26 f2112c3
Describe using external download of Stan repo, closes #138
