### 2023-06-26 a2adccd
Add warning to R doc about conda
