### 2023-06-26 roualdes/bridgestan#139 @WardBrian
The best place to do this is most likely on the "languages" landing page, no? https://roualdes.github.io/bridgestan/latest/languages.html

Something like
> All language interfaces expose the same core functionality from the C API [link]. Additional "quality of life" features such as the ability to download BridgeStan's source code automatically or compile models from inside the language (rather than manually calling `make`) are available on a best-effort basis. If you are missing these features in your favorite language, we would welcome a contribution!
