### 2023-06-26 roualdes/bridgestan#139 @bob-carpenter
I like it, too and it's exactly what I think we should write in our doc.

For the article, I'd try to stay a bit more formal in tone than "we would welcome a contribution!"   Given the space limitation, I suggest just dropping that last sentence, or at most, replacing it with some hint we might add more in the future.
