### 2023-06-26 47382a1
Update Getting Started documentation (#141)
* Add --shallow-submodules, closes #137

* Mention Stan and Math as submodules, closes #136

* Describe using external download of Stan repo, closes #138

* Fix links

* Tweak wording

* Wording tweak
