### 2023-06-27 96e049e
Delete extra sentence in R docs
