+++ R/R/bridgestan.R
@@ -10,7 +10,7 @@ StanModel <- R6::R6Class("StanModel",
     #' @description
     #' Create a Stan Model instance.
     #' @param lib A path to a compiled BridgeStan Shared Object file.
-    #' @param data Either a JSON string literal,a path to a data file in JSON format ending in ".json", or the empty string.
+    #' @param data Either a JSON string literal, a path to a data file in JSON format ending in ".json", or the empty string.
     #' @param seed Seed for the RNG used in constructing the model.
     #' @return A new StanModel.
     initialize = function(lib, data, seed) {
@@ -227,7 +227,6 @@ StanModel <- R6::R6Class("StanModel",
     },
     #' @description
     #' Return the log density of the specified unconstrained parameters.
-    #' See also `StanModel$param_unconstrain()`, the inverse of this function.
     #' @param theta_unc The vector of unconstrained parameters.
     #' @param propto If `TRUE`, drop terms which do not depend on the parameters.
     #' @param jacobian If `TRUE`, include change of variables terms for constrained parameters.
@@ -249,7 +248,6 @@ StanModel <- R6::R6Class("StanModel",
     },
     #' @description
     #' Return the log density and gradient of the specified unconstrained parameters.
-    #' See also `StanModel$param_unconstrain()`, the inverse of this function.
     #' @param theta_unc The vector of unconstrained parameters.
     #' @param propto If `TRUE`, drop terms which do not depend on the parameters.
     #' @param jacobian If `TRUE`, include change of variables terms for constrained parameters.
@@ -272,7 +270,6 @@ StanModel <- R6::R6Class("StanModel",
     },
     #' @description
     #' Return the log density, gradient, and Hessian of the specified unconstrained parameters.
-    #' See also `StanModel$param_unconstrain()`, the inverse of this function.
     #' @param theta_unc The vector of unconstrained parameters.
     #' @param propto If `TRUE`, drop terms which do not depend on the parameters.
     #' @param jacobian If `TRUE`, include change of variables terms for constrained parameters.
