### 2023-06-27 #5236 @roualdes
Thanks @saumil-sh for your helpful feedback. Here's a list of the changes we made based on your feedback so far.

# Major comments
1. and 3. Both refer to commits [4f5b904](https://github.com/roualdes/bridgestan/commit/4f5b904a8bbc5131fb41f09c946641c06995a614) and [e0878dc](https://github.com/roualdes/bridgestan/commit/e0878dc0aad89e7a9e7b6ad93d52dba8241143a4) on the branch joss.
2. A benchmark isn't a great fit here, for a few reasons.

    a. Comparing BridgeStan to Stan doesn't make a great amount of sense, since there is no other way we can extract gradients from Stan in a meaningful way to compare to BridgeStan.  BridgeStan is indeed the only tool that does this.  Plus we know that at most one memory allocation occurs in BridgeStan's extraction of a gradient, so there's not much to measure.

    b. There already exists speed comparisons of Stan to Turing.jl, which is a much more meaningful test than BridgeStan to Turing.jl.

    c. We are fairly word limited, by JOSS standards on word count.

4. see Minor comments.

# Minor comments
All minor comments are addressed by pull requests [#141](https://github.com/roualdes/bridgestan/pull/141) and [#143](https://github.com/roualdes/bridgestan/pull/143).

Please let us know if you have any further comments or suggestions.
