### 2023-06-28 13e1a5e
Update README for Rust example
