### 2023-06-28 94e5655
clarify how exactly bridgestan can aide algorithm developers
