### 2023-06-28 #5236 @roualdes
@salleuska, thanks.  That's great.  I didn't directly quote you, but I modified that paragraph with heavy inspiration from your suggestions.  Here's the paragraph in total,

> C++ can be cumbersome for algorithm prototyping.  As such, developers have been
requesting ways to access Stan models for algorithm development in Python, R,
and Julia. `BridgeStan` answers this call, making it easy for algorithm
developers to incorporate existing Stan models in their evaluation, e.g., the
dozens of diverse models with reference posteriors in
[`posteriordb`](https://github.com/stan-dev/posteriordb) [@Magnusson:2022].
`BridgeStan` further aides algorithm development by leveraging the functions in
the Stan Math library, which are written to appropriately handle numerical
issues such as under/over-flow and sum-to-one constraints.  Algorithm developers
using `BridgeStan` can thus focus more on their algorithms and less on their
implementations.  These benefits also ensure more precise and stable
implementations of statistical algorithms allowing for more fair comparisons.

I also updated the branch `joss`.
