### 2023-06-28 de94024
Add section about STAN_THREADS to rust docs
