### 2023-06-29 832b32e
add new_rng wrapper around StanRNG constructor
