### 2023-06-29 bff0383
add sentence to Julia interface docs about requiring STAN_THREADS=true
