### 2023-06-29 0b16c97
add entry to Julia doc
