+++ julia/docs/src/julia.md
@@ -90,6 +90,7 @@ param_constrain!
 param_unconstrain!
 param_unconstrain_json!
 StanRNG
+new_rng
 ```
 
 ### Compilation utilities
