### 2023-06-29 edc1e6f
Add how to run the example
