### 2023-06-29 679f5c9
move Julia testing info to CONTRIBUTING
