### 2023-06-29 a184246
Rust: documentation and release pipelines (#130)
* Start to clean up Rust docs

* Update CI

* Update release action (needs testing)

* Sketch docs

* Add non-working link

* Clean up some docs, un-mark Rng as send/sync

* Re-mark rng as send+sync

* cargo fmt

* three languages -> four

* Improve rust API docs

* Capitalization fixes

* Consistency edits

* Update ffi.rst

* Typo fixes

* Update README for Rust example

* Add section about STAN_THREADS to rust docs

* Add how to run the example

---------

Co-authored-by: Adrian Seyboldt <adrian.seyboldt@gmail.com>
