+++ rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bridgestan"
-version = "2.0.0"
+version = "2.1.0"
 edition = "2021"
 rust-version = "1.69"
 description = "Rust interface for BridgeStan"
