+++ rust/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 rust-version = "1.69"
 description = "Rust interface for BridgeStan"
 repository = "https://github.com/roualdes/bridgestan"
-license = "BSD-3-clause"
+license = "BSD-3-Clause"
 homepage = "https://roualdes.github.io/bridgestan/latest/"
 
 [dependencies]
