+++ .github/workflows/release.yaml
@@ -87,7 +87,7 @@ jobs:
         with:
           name: release-artifacts
           path: |
-            python/dist/
+            python/dist/*.whl
             bridgestan-*.tar.gz
 
       - name: Create release
