### 2023-06-30 79e583c
Move Windows path logic later in runtime
