### 2023-06-30 ee80fb0
Fix: store whether path has already been set properly
