### 2023-06-30 c4b9c3a
Move 'global' keyword
