### 2023-06-30 2cf71ee
In fact, just always store abspath
