### 2023-06-30 41ffee8
Fix: Windows TBB path logic fails if BridgeStan is not yet downloaded (#149)
* Test import without

* Fix name

* Move Windows path logic later in runtime

* Fix: store whether path has already been set properly

* Move 'global' keyword

* Remove from compile step

* Rename to windows_dll_path_setup

* Try to fix example on Windows

* Use abspath

* In fact, just always store abspath
