### 2023-07-05 #5236 @saumil-sh
Apologies for not getting back earlier.

I agree that comparing _BridgeStan_ and _Stan_ doesn't make sense. I proposed an overhead benchmark as I intend not to leave any loose end for a reader. In the case of at most one memory allocation (this is fantastic!), I agree there is hardly anything to measure. I am convinced that conveying this to a reader is more relevant than referring to automatic differentiation libraries. Adding one or two sentences in this direction may be worth trading the word count.

Otherwise, I am satisfied with all other changes.
