### 2023-07-05 b316633
add specifics of log_density_gradient memory allocations
