### 2023-07-05 #5236 @roualdes
@saumil-sh, thanks. I updated the last paragraph before Section Example to include more information about the memory allocations necessary for the BridgeStan function log_density_gradient, as you suggest.  Here's the updated paragraph in total (which is also updated on our BridgeStan branch joss)

> `BridgeStan` enables memory allocated in the host language (Julia, Python, or
R), to be reused within Stan; though any language with a C foreign function
interface could be similarly interfaced to access Stan methods. For instance,
the `BridgeStan` function `log_density_gradient` has an optional output argument, the
array into which the gradient will be stored.  By avoiding unnecessary copies,
`BridgeStan` is a zero-cost abstraction built upon Stan's math library.  If no
output argument is passed to `log_density_gradient`, then at most one memory
allocation, occuring in the host language, takes place.
