### 2023-07-06 f967e4e
test: Add a test that uses sundials
