### 2023-07-06 aa8a9d0
fix: Use SUNDIALS_TARGET in Makefile
