+++ python/test/test_stanmodel.py
@@ -43,7 +43,10 @@ def test_constructor():
         b4 = bs.StanModel(throw_data_so)
 
     load_sundials = str(STAN_FOLDER / "ode_sundials" / "ode_sundials_model.so")
-    bs.StanModel(load_sundials)
+    ode_sundials_data = (
+        STAN_FOLDER / "ode_sundials" / "ode_sundials.data.json"
+    ).read_text()
+    bs.StanModel(load_sundials, ode_sundials_data)
 
 
 def test_name():
