### 2023-07-06 7bf32a4
Fix Sundials `make` config (#152)
* test: Add a test that uses sundials

* ci: Disable fail-fast flag in CI

* fix: Use SUNDIALS_TARGET in Makefile

* Delete unnecessary old model

---------

Co-authored-by: Brian Ward <bward@flatironinstitute.org>
