+++ src/bridgestan.h
@@ -40,7 +40,7 @@ extern int bs_patch_version;
  * This PRNG is used for RNG functions in the `transformed data`
  * block of the model, and then discarded.
  * @param[out] error_msg a pointer to a string that will be allocated if there
- * is an error. This must later be freed by calling `bs_free_error_msg`.
+ * is an error. This must later be freed by calling bs_free_error_msg().
  * @return pointer to constructed model or `nullptr` if construction
  * fails
  */
@@ -161,10 +161,10 @@ int bs_param_unc_num(const bs_model* m);
  * @param[in] theta_unc sequence of unconstrained parameters
  * @param[out] theta sequence of constrained parameters
  * @param[in] rng pointer to pseudorandom number generator, should be created
- * by `bs_rng_construct`. This is only required when `include_gq` is `true`,
+ * by bs_rng_construct(). This is only required when `include_gq` is `true`,
  * otherwise it can be null.
  * @param[out] error_msg a pointer to a string that will be allocated if there
- * is an error. This must later be freed by calling `bs_free_error_msg`.
+ * is an error. This must later be freed by calling bs_free_error_msg().
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
@@ -183,7 +183,7 @@ int bs_param_constrain(const bs_model* m, bool include_tp, bool include_gq,
  * @param[in] theta sequence of constrained parameters
  * @param[out] theta_unc sequence of unconstrained parameters
  * @param[out] error_msg a pointer to a string that will be allocated if there
- * is an error. This must later be freed by calling `bs_free_error_msg`.
+ * is an error. This must later be freed by calling bs_free_error_msg().
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
@@ -200,7 +200,7 @@ int bs_param_unconstrain(const bs_model* m, const double* theta,
  * @param[in] json JSON-encoded constrained parameters
  * @param[out] theta_unc sequence of unconstrained parameters
  * @param[out] error_msg a pointer to a string that will be allocated if there
- * is an error. This must later be freed by calling `bs_free_error_msg`.
+ * is an error. This must later be freed by calling bs_free_error_msg().
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
@@ -220,7 +220,7 @@ int bs_param_unconstrain_json(const bs_model* m, const char* json,
  * @param[in] theta_unc unconstrained parameters
  * @param[out] lp log density to be set
  * @param[out] error_msg a pointer to a string that will be allocated if there
- * is an error. This must later be freed by calling `bs_free_error_msg`.
+ * is an error. This must later be freed by calling bs_free_error_msg().
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
@@ -244,7 +244,7 @@ int bs_log_density(const bs_model* m, bool propto, bool jacobian,
  * @param[out] val log density to be set
  * @param[out] grad gradient to set
  * @param[out] error_msg a pointer to a string that will be allocated if there
- * is an error. This must later be freed by calling `bs_free_error_msg`.
+ * is an error. This must later be freed by calling bs_free_error_msg().
  * @return code 0 if successful and code -1 if there is an exception
  * in the underlying Stan code
  */
@@ -272,7 +272,7 @@ int bs_log_density_gradient(const bs_model* m, bool propto, bool jacobian,
  * @param[out] grad gradient to set
  * @param[out] hessian hessian to set
  * @param[out] error_msg a pointer to a string that will be allocated if there
- * is an error. This must later be freed by calling `bs_free_error_msg`.
[…]
