### 2023-07-10 32c71bc
Docs: Formatting tweaks and fixes (#155)
* Change color of code in docs

* Docs: Improve usage of Sphinx directives, cross-linking

* Tweak color, fix typo
