### 2023-07-10 50caabb
Rename pointer for consistency
