+++ python/example.py
@@ -2,6 +2,10 @@ import numpy as np
 
 import bridgestan as bs
 
+# These paths are what they are because this example lives in a subfolder
+# of the BridgeStan repository. If you're running this on your own, you
+# will most likely want to delete the next line (to have BridgeStan
+# download its sources for you) and change the paths on the following two
 bs.set_bridgestan_path("..")
 
 stan = "../test_models/bernoulli/bernoulli.stan"
