+++ julia/docs/src/julia.md
@@ -20,7 +20,7 @@ BridgeStan is registered on JuliaRegistries each release.
 ] add BridgeStan
 ```
 
-The first time you need it, the BridgeStan source code will be downloaded
+The first time you compile a model, the BridgeStan source code will be downloaded
 as an [Artifact](https://pkgdocs.julialang.org/v1/artifacts/). If you
 prefer to use a source distribution of BridgeStan, consult the following section.
 
