### 2023-07-17 #5236 @Nikoleta-v3
Hey @roualdes thank you for your patience! I have not further comments. Just two minor typos in the manuscript:

- Line 8: ~to~

- Line 35: I believe it should be `aids`

At this point could you please:

- [ ] (if necessary) Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release 1.0.0
 Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
