+++ paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 Stan provides a probabilistic programming language in which users can code
 Bayesian models [@Carpenter:2017; @Stan:2022]. A Stan program is transpiled to
-to a C++ class which links to the Stan math library to implement smooth,
+to a C++ class that links to the Stan math library to implement smooth,
 unconstrained posterior log densities, gradients, and Hessians as well as
 constraining/unconstraining transforms.  Implementation is provided through
 automatic differentiation in the Stan math library
@@ -49,13 +49,13 @@ with minimal additional development.
 
 # Statement of need
 
-Stan was developed for applied statisticians working on real world
+Stan was developed for applied statisticians working on real-world
 problems and has been used by hundreds of thousands of researchers and
-practitioners across the social, biological, and physical sciences,
+practitioners across the social, biological, and physical sciences, as well as
 engineering, education, sports, and finance. Stan provides several
 state-of-the-art, gradient-based algorithms: full Bayesian inference
 with Hamiltonian Monte Carlo, Laplace approximation based on L-BFGS
-optimization, and autodiff variational inference (ADVI).
+optimization, and automatic differentiation variational inference (ADVI).
 
 In the statistical software environment R, Stan is heavily relied upon for
 development of applied statistics packages. Using Google's
@@ -78,31 +78,31 @@ the Stan Math library, which are written to appropriately handle numerical
 issues such as under/over-flow and sum-to-one constraints.  Algorithm developers
 using `BridgeStan` can thus focus more on their algorithms and less on their
 implementations.  These benefits also ensure more precise and stable
-implementations of statistical algorithms allowing for more fair comparisons.
+implementations of statistical algorithms, allowing for more fair comparisons.
 
 `BridgeStan` is an interface, written in C-compatible C++, between a Stan
-program and any higher level language which exposes a C foreign function
+program and any higher level language that exposes a C foreign function
 interface. Julia, Python, and R each have C foreign function interfaces. Using
 memory allocated within such higher level languages, `BridgeStan` provides
-computations of the log joint density function, and its gradient, of a Stan
+computations of the log joint density function and corresponding gradient of a Stan
 model, which is itself implemented using highly templated C++ from the Stan math
 library. Using a memory-compatible C interface makes this possible even if the
 host language (e.g., R) was compiled with a different compiler, something no
-prior interface which exposed Stan's log density calculations could allow.
+prior interface that exposed Stan's log density calculations could allow.
 
 Other software in the Stan ecosystem provides some overlapping features with
 `BridgeStan`.  For instance, `rstan` [@rstan] provides functions `log_prob` and
 `grad_log_prob`, which provide access to the log joint density and its gradient.
 Similarly, `httpstan` [@httpstan] offers `log_prob` and `log_prob_grad`.  Such
 cases of similar functionality are unfortunately limited.  As of 2023-05-19,
-`rstan` via CRAN is still on Stan version 2.21.0 (released 2019-10-18) and the
+`rstan` via CRAN is still on Stan version 2.21.0 (released 2019-10-18), and the
 development version of `rstan`, which is not hosted on CRAN, is on Stan version
 2.26.1 (released 2021-02-17), while the latest version of Stan is on 2.32.2
 (released 2023-05-15).  Further, `rstan` is limited to the host language R.  On
-the other hand, `httpstan` is a Python package which offers a REST API,
+the other hand, `httpstan` is a Python package that offers a REST API,
 primarily targeting the Stan algorithms, which allows some limited access to the
 methods of a Stan model.  The REST API may be used by languages other than
-Python, but by design cannot take advantage of direct memory access of the host
[…]
