### 2023-07-17 37174f6
Acknowledge Seth's funding
