+++ python/bridgestan/__version.py
@@ -1,2 +1,2 @@
-__version__ = "2.1.1"
+__version__ = "2.1.2"
 __version_info__ = tuple(map(int, __version__.split(".")))
