### 2023-07-20 #5236 @roualdes
Tagged v2.1.2: https://github.com/roualdes/bridgestan/releases/tag/v2.1.2
Zenodo up to date: https://zenodo.org/record/8169248
![image](https://github.com/openjournals/joss-reviews/assets/3252729/88a04a86-8dd1-4d51-b561-28e6bbf74e9b)
